% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioeconomyFunctions.R
\name{addBiofuelsSector}
\alias{addBiofuelsSector}
\title{Adds a new biofuels sector to an existing model. This includes 3 new industries and 1 new commodity.}
\usage{
addBiofuelsSector(
  modelName,
  newIndustryInfo,
  newCommodityInfo,
  simCommodityInfo,
  percentage,
  inputPurchases,
  valueAdded,
  newEnvData
)
}
\arguments{
\item{newIndustryInfo}{dataframe with 3 different technologies on the rows and the following variables (Code,Name, Primary_product, Price, Percentage_Prod) on the columns.}

\item{newCommodityInfo}{dataframe with the information of the new commodity. Column variables (Code, Name, Primary_producers).}

\item{simCommodityInfo}{list with the information of the similar commodity. Elements: (Code, Name, PrimaryProducer_Code, PrimaryProducer_Name, Price).}

\item{percentage}{numeric (0,1] that refers to the \% of the original commodity (fuels) that the new commodity (biofuels) will replace.}

\item{inputPurchases}{(nComm+1)x 3 matrix with the amount spent in each of the nComm existing commodities to produce 1 GGE of biofuel in each of the new industries.}

\item{valueAdded}{2x3 matrix with the first 2 components of value added (Compensation of employees and Taxes on production and imports, less subsidies) required 
to produce 1 GGE of biofuel in each of the new industries.}

\item{newEnvData}{(# environmental flows x 3) matrix with the data for all the environmental flows per dollar of output for the new industries.}

\item{modelname}{Name of the model from a config file.}
}
\value{
A list with USEEIO model components and attributes modified and ready for calculation.
}
\description{
This function adds 3 new industries (that correspond to 3 different technologies) to produce 1 new commodity (biofuels).
This new commodity (biofuels) will be a perfect substitute of fuels (commodity) produced by Petroleum Refineries.
This function prepares (read and organize data from the BEA tables), modify the Make and Use table to add the new biofuels sector, 
build the model (create the required matrices) and update B and W matrices after its construction to return a model including the new sector
ready for calculation.See details of the methodology on "LCA New tech Methodology (v01 CGA 2020-07-23).pdf".
}
