% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormattingFunctions.R
\name{addSlashandNameItem}
\alias{addSlashandNameItem}
\title{Appends name after slash and formats all in lower case}
\usage{
addSlashandNameItem(v, loc)
}
\arguments{
\item{v, }{character vector of names}

\item{loc, }{string for location}
}
\value{
v, the original char v with /loc added
}
\description{
Appends name after slash and formats all in lower case
}
