% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{adjustBEACPItoIOIndustry2012Schema}
\alias{adjustBEACPItoIOIndustry2012Schema}
\title{Adjust CPI from GDP industries to IO industries (2012 schema) at Detail, Summary, and Sector IO levels.}
\usage{
adjustBEACPItoIOIndustry2012Schema()
}
\value{
A list contains IO-based CPI at Detail, Summary, and Sector IO levels.
}
\description{
Adjust CPI from GDP industries to IO industries (2012 schema) at Detail, Summary, and Sector IO levels.
}
