% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{adjustBEAGrossOutouttoIOIndustry2012Schema}
\alias{adjustBEAGrossOutouttoIOIndustry2012Schema}
\title{Adjust multi-year USEEIO gross output by model-specified currency year.}
\usage{
adjustBEAGrossOutouttoIOIndustry2012Schema()
}
\value{
A dataframe contains adjusted multi-year USEEIO gross output.
Adjust gross output from GDP industries to IO industries (2012 schema) at Detail, Summary, and Sector IO levels.

A list contains IO-based gross output at Detail, Summary, and Sector IO levels.
}
\description{
Adjust multi-year USEEIO gross output by model-specified currency year.
}
