% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{adjustBEAGrossOutputtoIOIndustry2012Schema}
\alias{adjustBEAGrossOutputtoIOIndustry2012Schema}
\title{Adjust gross output from GDP industries to IO industries (2012 schema) at Detail, Summary, and Sector IO levels.}
\usage{
adjustBEAGrossOutputtoIOIndustry2012Schema()
}
\value{
A list contains IO-based gross output at Detail, Summary, and Sector IO levels.
}
\description{
Adjust gross output from GDP industries to IO industries (2012 schema) at Detail, Summary, and Sector IO levels.
}
