% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{adjustBEAValueAddedtoIOIndustry2012Schema}
\alias{adjustBEAValueAddedtoIOIndustry2012Schema}
\title{Adjust Value Added ($) from GDP industries to IO industries (2012 schema) at Summary and Sector IO levels.}
\usage{
adjustBEAValueAddedtoIOIndustry2012Schema()
}
\value{
A list contains IO-based Value Added at Summary and Sector IO levels.
}
\description{
Adjust Value Added ($) from GDP industries to IO industries (2012 schema) at Summary and Sector IO levels.
}
