% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIOFunctions.R
\name{adjustMarginswithWIOSectors}
\alias{adjustMarginswithWIOSectors}
\title{Adjust Margins to the correct dimensions based on WIO sectors}
\usage{
adjustMarginswithWIOSectors(model, WIO)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{WIO}{A list with WIO specifications and data}
}
\value{
A model object with a modified Margins object based on WIO sectors
}
\description{
Adjust Margins to the correct dimensions based on WIO sectors
}
