% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustPrice.R
\name{adjustMultiplierPriceType}
\alias{adjustMultiplierPriceType}
\title{Adjust multipliers from producer to purchaser price.}
\usage{
adjustMultiplierPriceType(matrix, currency_year, model)
}
\arguments{
\item{matrix}{A matrix representing the multiplier that needs price type adjustment.}

\item{currency_year}{An integer representing the currency year.}

\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A matrix representing the multiplier that is adjusted to purchaser price.
}
\description{
Adjust multipliers from producer to purchaser price.
}
