% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustPrice.R
\name{adjustMultiplierPriceYear}
\alias{adjustMultiplierPriceYear}
\title{Adjust multipliers from IO year to currency year price.}
\usage{
adjustMultiplierPriceYear(matrix_name, currency_year, model)
}
\arguments{
\item{matrix_name}{Name of matrix representing the multiplier that needs price year adjustment.}

\item{currency_year}{An integer representing the currency year.}

\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A matrix representing the multiplier that is adjusted to currency year price.
}
\description{
Adjust multipliers from IO year to currency year price.
}
