% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUIOFunctions.R
\name{adjustMultiYearObjectsForMUIO}
\alias{adjustMultiYearObjectsForMUIO}
\title{Adjust MultiYearObjects in MUIO models such that they have the same dimensions as the other model objects}
\usage{
adjustMultiYearObjectsForMUIO(
  model,
  physComIndeces,
  physIndIndeces,
  econComIndeces,
  econIndIndeces
)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{physComIndeces}{Indeces that specify the location of the physical MUIO commodities in the commodity lists}

\item{physIndIndeces}{Indeces that specify the location of the physical MUIO industries in the industry lists}

\item{econComIndeces}{Indeces that specify the location of the economic MUIO commodities (i.e., the "Other" sectors, X in code) in the commodity lists}

\item{econIndIndeces}{Indeces that specify the location of the economic MUIO industries (i.e., the "Other" sectors, X in code) in the industry lists}
}
\value{
A model with the multiyear objects adjusted for the MUIO industries/commodities.
}
\description{
Adjust MultiYearObjects in MUIO models such that they have the same dimensions as the other model objects
}
