% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIOFunctions.R
\name{adjustMultiYearObjectsForWIO}
\alias{adjustMultiYearObjectsForWIO}
\title{Adjust MultiYearObjects in WIO models such that they have the same dimensions as the other model objects}
\usage{
adjustMultiYearObjectsForWIO(model, WIO)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{WIO}{A dataframe with WIO specifications}
}
\value{
spec A dataframe with the specified sectors added to the spec$NAICSSetorCW object
}
\description{
Adjust MultiYearObjects in WIO models such that they have the same dimensions as the other model objects
}
