% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustPrice.R
\name{adjustResultMatrixPrice}
\alias{adjustResultMatrixPrice}
\title{Adjust price year and type (producer's or purchaser's) of a model result matrix.
Model result matrices are M, M_d, N, N_d
Year adjustments from 2007-2018 supported}
\usage{
adjustResultMatrixPrice(
  matrix_name,
  currency_year,
  purchaser_price = TRUE,
  model
)
}
\arguments{
\item{matrix_name}{Name of the result matrix that needs price adjustment, e.g. "N"}

\item{currency_year}{An integer representing the currency year, e.g. 2018.}

\item{purchaser_price}{A logical value indicating whether to adjust producer's price to purchaser's price.}

\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A model result matrix after price adjustment
}
\description{
Adjust price year and type (producer's or purchaser's) of a model result matrix.
Model result matrices are M, M_d, N, N_d
Year adjustments from 2007-2018 supported
}
