% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{adjustUSEEIOGrossOutputbyCPIYear}
\alias{adjustUSEEIOGrossOutputbyCPIYear}
\title{Adjust multi-year USEEIO gross output by model-specified currency year.}
\usage{
adjustUSEEIOGrossOutputbyCPIYear()
}
\value{
A dataframe contains adjusted multi-year USEEIO gross output.
}
\description{
Adjust multi-year USEEIO gross output by model-specified currency year.
}
