% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{aggregateMasterCrosswalk}
\alias{aggregateMasterCrosswalk}
\title{Disaggregate the MasterCrosswalk to include the new sectors for disaggregation}
\usage{
aggregateMasterCrosswalk(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
crosswalk with aggregated sectors removed
}
\description{
Disaggregate the MasterCrosswalk to include the new sectors for disaggregation
}
