% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{aggregateMatrix}
\alias{aggregateMatrix}
\title{Aggregate matrix by rows then by columns}
\usage{
aggregateMatrix(matrix, from_level, to_level, model)
}
\description{
Aggregate matrix by rows then by columns
}
