% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{aggregateModel}
\alias{aggregateModel}
\title{Aggregate a model based on specified source file}
\usage{
aggregateModel(model)
}
\arguments{
\item{model}{Model file loaded with IO tables}
}
\value{
An aggregated model.
}
\description{
Aggregate a model based on specified source file
}
