% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{aggregateResultMatrix}
\alias{aggregateResultMatrix}
\title{Aggregate result matrix by rows and columns}
\usage{
aggregateResultMatrix(matrix, from_level, to_level)
}
\arguments{
\item{matrix}{A matrix with sectors as rows and columns}

\item{from_level}{The level of BEA code this matrix starts at}

\item{to_level}{The level of BEA code this matrix will be aggregated to}
}
\value{
aggregated matrix with sectors as rows and columns
}
\description{
Aggregate result matrix by rows and columns
}
