% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{aggregateResultMatrixbyRow}
\alias{aggregateResultMatrixbyRow}
\title{Aggregate result matrix by rows}
\usage{
aggregateResultMatrixbyRow(matrix, to_level, crosswalk)
}
\arguments{
\item{matrix}{A matrix with sectors as rows}

\item{to_level}{The level of BEA code this matrix will be aggregated to}

\item{crosswalk}{Sector crosswalk between levels of detail}
}
\value{
An aggregated matrix with sectors as rows
}
\description{
Aggregate result matrix by rows
}
