% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{aggregateResultTable}
\alias{aggregateResultTable}
\title{Aggregate result matrix by rows}
\usage{
aggregateResultTable(matrix, from_level, to_level, crosswalk)
}
\arguments{
\item{matrix}{A matrix with sectors as rows}

\item{from_level}{The level of BEA code this matrix starts at}

\item{to_level}{The level of BEA code this matrix will be aggregated to}

\item{crosswalk}{Sector crosswalk between levels of detail}
}
\value{
aggregated matrix with sectors as rows
}
\description{
Aggregate result matrix by rows
}
