% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{aggregateSatelliteTable}
\alias{aggregateSatelliteTable}
\title{Aggreagte (FlowAmount in) satellite tables from one BEA level to another}
\usage{
aggregateSatelliteTable(sattable, from_level, to_level)
}
\arguments{
\item{sattable}{A satellite table to be aggregated based on the level (Detail, Summary, or Sector) of BEA code.}

\item{from_level}{The level of BEA code in the satellite table.}

\item{to_level}{The level of BEA code this satellite table will be aggregated to.}
}
\value{
A more aggregated satellite table.
}
\description{
Aggreagte (FlowAmount in) satellite tables from one BEA level to another
}
