% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{aggregateSector}
\alias{aggregateSector}
\title{Aggregate a sector in a table}
\usage{
aggregateSector(model, mainSector, sectorToRemove, tableType, domestic = FALSE)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{mainSector}{Sector to aggregate to (string)}

\item{sectorToRemove}{Sector to be aggregated into mainSector, then removed from table (string)}

\item{tableType}{String to designate either Make or Use table}

\item{domestic}{Boolean to indicate whether to aggregate the UseTransactions or DomesticUseTransactions table}
}
\value{
aggregated table
}
\description{
Aggregate a sector in a table
}
