% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{aggregateVA}
\alias{aggregateVA}
\title{Aggregate the MakeTable based on specified source file}
\usage{
aggregateVA(model, aggregationSpecs)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{aggregationSpecs}{Specifications for aggregation}
}
\value{
An aggregated MakeTable.
}
\description{
Aggregate the MakeTable based on specified source file
}
