% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{aggSatelliteTable}
\alias{aggSatelliteTable}
\title{Aggregate satellite tables from static file based on specs}
\usage{
aggSatelliteTable(model, sattable, sat)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{sattable}{A standardized satellite table with resource and emission names from original sources.}

\item{sat}{The abbreviation for the satellite table.}
}
\value{
A standardized satellite table with aggregated sectors added.
}
\description{
Aggregate satellite tables from static file based on specs
}
