% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{applyRAS}
\alias{applyRAS}
\title{Integrate pre-adjustment of t_r, t_c and t (tolerance level) with RAS function.}
\usage{
applyRAS(m0, t_r, t_c, relative_diff, absolute_diff, max_itr)
}
\arguments{
\item{m0}{A matrix object.}

\item{t_r}{A vector setting the target row sums of the matrix.}

\item{t_c}{A vector setting the target column sums of the matrix.}

\item{relative_diff}{A numeric value setting the relative difference of the two numerical vectors.}

\item{absolute_diff}{A numeric value setting the mean absolute difference of the two numerical vectors.}

\item{max_itr}{A numeric value setting the maximum number of iterations to try for convergence.
Default is 1,000,000.}
}
\value{
A RAS balanced matrix.
}
\description{
Integrate pre-adjustment of t_r, t_c and t (tolerance level) with RAS function.
}
