% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{AssembleMake}
\alias{AssembleMake}
\title{Assemble Disaggregated Make table from the various disaggregated components.}
\usage{
AssembleMake(
  originalMake,
  originalRowIndex,
  originalColIndex,
  disaggCols,
  disaggRows,
  disaggIntersection
)
}
\arguments{
\item{originalMake}{Dataframe. The original Make table before disaggregation}

\item{originalRowIndex}{Integer. The row index, in the original Make table, of the sector to be disaggregated}

\item{originalColIndex}{Integer. The column index, in the original Make table, of the sector to be disaggregated}

\item{disaggCols}{Dataframe. Previously disaggregated columns of the Make table.}

\item{disaggRows}{Dataframe. Previously disaggregated rows of the Make table.}

\item{disaggIntersection}{Dataframe. Previously disaggregated intersection of the Make table.}
}
\value{
The Make table as a dataframe with the disaggregated rows, columns, and intersection included
}
\description{
Assemble Disaggregated Make table from the various disaggregated components.
}
