% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIOFunctions.R
\name{assembleWIOModel}
\alias{assembleWIOModel}
\title{Build a WIO-style model by including data read from wiospecs folder}
\usage{
assembleWIOModel(model)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}
}
\value{
A model with the UseTransactions matrix modified with WIO specs.
}
\description{
Build a WIO-style model by including data read from wiospecs folder
}
