% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationFunctions.R
\name{barplotFlowbyGroup}
\alias{barplotFlowbyGroup}
\title{Bar plot of flow by group}
\usage{
barplotFlowbyGroup(table, modelspecs)
}
\arguments{
\item{table}{Supports a model table}

\item{modelspecs}{The specs for a given model}
}
\description{
Bar plot of flow by group
}
