% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationFunctions.R
\name{barplotIndicatorScoresbySector}
\alias{barplotIndicatorScoresbySector}
\title{Bar plot of indicator scores calculated from totals by sector and displayed by BEA Sector Level}
\usage{
barplotIndicatorScoresbySector(
  model_list,
  totals_by_sector_name,
  indicator_code
)
}
\arguments{
\item{model_list}{List of EEIO models with IOdata, satellite tables, and indicators loaded}

\item{totals_by_sector_name}{The name of one of the totals by sector tables available in model$SatelliteTables$totals_by_sector}

\item{indicator_code}{The code of the indicator of interest from the model$Indicators}
}
\description{
Bar plot of indicator scores calculated from totals by sector and displayed by BEA Sector Level
}
