% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIOTables.R
\name{BEAMakeAndUseTables}
\alias{BEAMakeAndUseTables}
\title{Externalize the BEA make and use tables}
\usage{
BEAMakeAndUseTables(iolevel, makeoruse, year, redef)
}
\arguments{
\item{iolevel}{"Detail", "Sector", or "Summary" level}

\item{makeoruse}{Select either "Make" or "Use"}

\item{year}{Year of interest between 2010--2018}

\item{redef}{Select either "BeforeRedef" or "AfterRedef"}
}
\value{
A dataframe of BEA make or use tables
}
\description{
Externalize the BEA make and use tables
}
