% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{BLS_QCEW_2013}
\alias{BLS_QCEW_2013}
\title{BLS QCEW state ESTAB and EMP table for 2013}
\format{
A data frame with 106238 obs. and 6 variables
\describe{
 \item{GeoFips}{integer}
 \item{GeoName}{text code}
 \item{NAICS}{character}
 \item{Year}{integer}
 \item{ESTAB}{number of establishment}
 \item{EMP}{number of employee}
}
}
\source{
\url{https://www.bls.gov/cew/downloadable-data-files.htm}
}
\usage{
BLS_QCEW_2013
}
\description{
BLS QCEW state ESTAB and EMP table for 2013
}
\keyword{datasets}
