% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{buildDisaggFullUse}
\alias{buildDisaggFullUse}
\title{Build a Full Use table using the Use transactions, Use value added, and final demand model objects}
\usage{
buildDisaggFullUse(model, disagg)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}
}
\value{
dataframe representing a use table that includes the Use transactions, Use value added, and final demand sectors
}
\description{
Build a Full Use table using the Use transactions, Use value added, and final demand model objects
}
