% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{buildEconomicMatrices}
\alias{buildEconomicMatrices}
\title{Construct the economic matrices of an IO model based on loaded IO tables.}
\usage{
buildEconomicMatrices(model)
}
\arguments{
\item{model}{An EEIO model object with model specs, IO tables}
}
\value{
A list with EEIO economic matrices.
}
\description{
Construct the economic matrices of an IO model based on loaded IO tables.
}
