% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{buildIOModel}
\alias{buildIOModel}
\title{Build an IO model with economic components only.}
\usage{
buildIOModel(modelname, configpaths = NULL)
}
\arguments{
\item{modelname}{Name of the model from a config file.}

\item{configpaths}{str vector, paths (including file name) of model configuration file
and optional agg/disagg configuration file(s). If NULL, built-in config files are used.}
}
\value{
A list of IO model with only economic components
}
\description{
Build an IO model with economic components only.
}
