% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Margins.R
\name{buildMarginAllocationMatrix}
\alias{buildMarginAllocationMatrix}
\title{Create margin_allocation matrix to allocate fractions by margin sector
Currently uses sector output to provide that allocation}
\usage{
buildMarginAllocationMatrix(all_margin_sectors, model)
}
\arguments{
\item{all_margin_sectors, }{vector of sector codes}

\item{model, }{a fully built model}
}
\value{
matrix, margin types x margin sectors with values being fractions of type to each sector
}
\description{
Create margin_allocation matrix to allocate fractions by margin sector
Currently uses sector output to provide that allocation
}
