% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{buildModel}
\alias{buildModel}
\title{Build an EEIO model.}
\usage{
buildModel(modelname, configfile = NULL)
}
\arguments{
\item{modelname}{Name of the model from a config file.}

\item{configfile}{Directory of model configuration file.}
}
\value{
A list of EEIO model components and attributes
}
\description{
Build an EEIO model.
}
