% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{buildModelwithImportFactors}
\alias{buildModelwithImportFactors}
\title{Create import Use table and validate domestic+import against full use model .}
\usage{
buildModelwithImportFactors(model)
}
\arguments{
\item{model, }{An EEIO model object with model specs and crosswalk table loaded}
}
\value{
A model object with explicit import components.
}
\description{
Create import Use table and validate domestic+import against full use model .
}
