% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{buildPriceMatrices}
\alias{buildPriceMatrices}
\title{Construct the price adjustment matrices, Rho, Tau, and Phi}
\usage{
buildPriceMatrices(model)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables}
}
\value{
A list with EEIO price adjustment matrices.
}
\description{
Construct the price adjustment matrices, Rho, Tau, and Phi
}
