% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initiation_Model.R
\name{calculate}
\alias{calculate}
\title{Calculate total emissions/resources (LCI) or total impact for USEEIO model for a given demand vector and perspective.}
\usage{
calculate(model, perspective)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{perspective}{Perspective of the model, can be "DIRECT", "INTERMEDIATE", or "FINAL".}
}
\value{
A list with LCI and LCIA results of the EEIO model.
}
\description{
Calculate total emissions/resources (LCI) or total impact for USEEIO model for a given demand vector and perspective.
}
