% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{calculateBalancedDomesticTables}
\alias{calculateBalancedDomesticTables}
\title{Calculate the domestic use transactions and final demand tables after RAS balancing}
\usage{
calculateBalancedDomesticTables(model, disagg, balancedFullUse)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}

\item{balancedFullUse}{A fullUse table (including UseTransactions, UseValueAdded, and FinalDemand), created to determine whether RAS balancing is needed}
}
\value{
list containing balanced domesticFinalDemand and domesticUseTransactions dataframes.
}
\description{
Calculate the domestic use transactions and final demand tables after RAS balancing
}
