% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustPrice.R
\name{calculateBasicbyProducerPriceRatio}
\alias{calculateBasicbyProducerPriceRatio}
\title{Calculate basic to producer price ratio.}
\usage{
calculateBasicbyProducerPriceRatio(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A dataframe of basic to producer price ratio.
}
\description{
Calculate basic to producer price ratio.
}
