% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{calculateDefaultIntersection}
\alias{calculateDefaultIntersection}
\title{Creates a square dataframe matrix with values assigned based on default percentages}
\usage{
calculateDefaultIntersection(
  originalIntersection,
  defaultPercentages,
  newSectorCodes
)
}
\arguments{
\item{originalIntersection}{int value of the original intersection to be disaggregated}

\item{defaultPercentages}{vector of allocation percentages}

\item{newSectorCodes}{vector of named disaggregated sectors}
}
\value{
square dataframe matrix with new sectors as row and column names with default values
}
\description{
Creates a square dataframe matrix with values assigned based on default percentages
}
