% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R, R/CalculationFunctions.R
\name{calculateDirectPerspectiveLCI}
\alias{calculateDirectPerspectiveLCI}
\title{Multiply the B matrix and the scaling vector c.}
\usage{
calculateDirectPerspectiveLCI(B, c)

calculateDirectPerspectiveLCI(B, c)
}
\arguments{
\item{B}{Marginal impact per unit of the environmental flows.}

\item{c}{Scaling vector.}

\item{B}{Marginal impact per unit of the environmental flows.}

\item{c}{Scaling vector.}
}
\value{
Transposed m_d with total impacts in form of sector x flows.

Transposed m_d with total impacts in form of sector x flows.
}
\description{
Multiply the B matrix and the scaling vector c.

Multiply the B matrix and the scaling vector c.
}
\references{
Yang, Yi, Wesley W. Ingwersen, Troy R. Hawkins, Michael Srocka, and David E. Meyer.
2017. “USEEIO: A New and Transparent United States Environmentally-Extended Input-Output Model.”
Journal of Cleaner Production 158 (August): 308–18. https://doi.org/10.1016/j.jclepro.2017.04.150.
SI1, Equation 8.

Yang, Yi, Wesley W. Ingwersen, Troy R. Hawkins, Michael Srocka, and David E. Meyer.
2017. “USEEIO: A New and Transparent United States Environmentally-Extended Input-Output Model.”
Journal of Cleaner Production 158 (August): 308–18. https://doi.org/10.1016/j.jclepro.2017.04.150.
SI1, Equation 8.
}
