% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{calculateFinalPerspectiveLCIA}
\alias{calculateFinalPerspectiveLCIA}
\title{The final perspective LCIA aligns impacts with sectors consumed by final users.
Multiply the N matrix and the diagonal of demand, y.}
\usage{
calculateFinalPerspectiveLCIA(N, y)
}
\arguments{
\item{N, }{a model N matrix, direct + indirect impact per unit of the environmental flows.}

\item{y, }{a model demand vector}
}
\value{
A matrix with total impacts in form of sector x impact categories.
}
\description{
The final perspective LCIA aligns impacts with sectors consumed by final users.
Multiply the N matrix and the diagonal of demand, y.
}
\references{
Yang, Yi, Wesley W. Ingwersen, Troy R. Hawkins, Michael Srocka, and David E. Meyer.
2017. “USEEIO: A New and Transparent United States Environmentally-Extended Input-Output Model.”
Journal of Cleaner Production 158 (August): 308–18. https://doi.org/10.1016/j.jclepro.2017.04.150.
SI1, Equation 10.
}
