% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{calculateGhoshB}
\alias{calculateGhoshB}
\title{Function to get supply side Ghosh matrix B. B can be calculated based on commodity-by-commodity or 
industry-by-industry direct requirement matrix: (Miller and Blair)
for industry-by-industry requirements: B = inverse(qhat) * A(industry-by-industry)*qhat,
where q is the total commodity output
for commodity-by-commodity requirements: B = inverse(xhat) * A(industry-by-industry)*xhat,
where x is the total inudstry output}
\usage{
calculateGhoshB(model, use_domestic_requirements = FALSE)
}
\arguments{
\item{model}{An EEIO model object with model specs, IO tables, satellite tables, and indicators loaded}

\item{use_domestic_requirements}{A logical value: if TRUE, use domestic demand and L_d matrix;
if FALSE, use complete demand and L matrix.}
}
\value{
B, Ghosh B matrix
}
\description{
Function to get supply side Ghosh matrix B. B can be calculated based on commodity-by-commodity or 
industry-by-industry direct requirement matrix: (Miller and Blair)
for industry-by-industry requirements: B = inverse(qhat) * A(industry-by-industry)*qhat,
where q is the total commodity output
for commodity-by-commodity requirements: B = inverse(xhat) * A(industry-by-industry)*xhat,
where x is the total inudstry output
}
