% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{calculateHouseholdEmissions}
\alias{calculateHouseholdEmissions}
\title{Calculate household emissions from B_h}
\usage{
calculateHouseholdEmissions(
  model,
  f,
  location,
  characterized = FALSE,
  show_RoW = FALSE
)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{f}{A demand vector with names as one or more model sectors and
numeric values in USD with the same dollar year as model.}

\item{location, }{str optional location code for demand vector, required for two-region models}

\item{characterized, }{bool, TRUE to characterize using C matrix, FALSE to show LCI}

\item{show_RoW, }{bool, if TRUE, include rows for commodities in RoW, e.g. `111CA/RoW` in result objects.
Only valid currently for models with ExternalImportFactors.}
}
\value{
A result vector with rows for final demand sector(s)
}
\description{
Calculate household emissions from B_h
}
