% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{calculateIndicatorScoresforTotalsBySector}
\alias{calculateIndicatorScoresforTotalsBySector}
\title{Adds an indicator score to a totals by sector table. A short cut alternative to getting totals before model result}
\usage{
calculateIndicatorScoresforTotalsBySector(
  model,
  totals_by_sector_name,
  indicator_name
)
}
\arguments{
\item{model}{A EEIO model with model specs, IO tables, satellite tables, and indicators loaded}

\item{totals_by_sector_name}{The name of one of the totals by sector tables available in model$SatelliteTables$totals_by_sector}

\item{indicator_name}{The name of the indicator of interest from the model$Indicators$factors}
}
\value{
a totals_by_sector table with fields from the Indicator table "Code" and "Amount", and calculated "IndicatorScore" added
}
\description{
Adds an indicator score to a totals by sector table. A short cut alternative to getting totals before model result
}
