% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIOTables.R
\name{calculateIndustryCommodityOutput}
\alias{calculateIndustryCommodityOutput}
\title{Calculate industry and commodity output vectors from model components.}
\usage{
calculateIndustryCommodityOutput(model)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}
}
\value{
An EEIO model with industry and commodity output added
}
\description{
Calculate industry and commodity output vectors from model components.
}
