% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{calculateLeontiefInverse}
\alias{calculateLeontiefInverse}
\title{Calculate Leontief inverse from direct requirements matrix.}
\usage{
calculateLeontiefInverse(A)
}
\arguments{
\item{A}{Direct Requirements matrix.}
}
\value{
Leontief inverse.
}
\description{
Calculate Leontief inverse from direct requirements matrix.
}
