% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{calculateMarginSectorImpacts}
\alias{calculateMarginSectorImpacts}
\title{Calculate sector margin impacts in the form of M and N Matrix}
\usage{
calculateMarginSectorImpacts(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A list with M_margin and N_margin
}
\description{
Calculate sector margin impacts in the form of M and N Matrix
}
