% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustPrice.R
\name{calculateModelIOYearbyYearPriceRatio}
\alias{calculateModelIOYearbyYearPriceRatio}
\title{Calculate model IO year by year price ratio.}
\usage{
calculateModelIOYearbyYearPriceRatio(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A dataframe of model IO year by year price ratio.
}
\description{
Calculate model IO year by year price ratio.
}
