% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalImportFactors.R
\name{calculateMwithImportFactors}
\alias{calculateMwithImportFactors}
\title{Derives an M matrix for total embodied flows from domestic and imported supply chains.}
\usage{
calculateMwithImportFactors(model)
}
\arguments{
\item{model, }{An EEIO model object with model specs and crosswalk table loaded}
}
\value{
An M matrix of flows x sector
}
\description{
Derives an M matrix for total embodied flows from domestic and imported supply chains.
}
