% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{calculateOutputRatio}
\alias{calculateOutputRatio}
\title{Generate Output Ratio table, flexible to Commodity/Industry output and model Commodity/Industry type}
\usage{
calculateOutputRatio(model, output_type = "Commodity")
}
\description{
Generate Output Ratio table, flexible to Commodity/Industry output and model Commodity/Industry type
}
