% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{calculatePercentContributiontoImpact}
\alias{calculatePercentContributiontoImpact}
\title{Dot multiplies two vectors to calculate an impact score and the percent contribution each score to the total}
\usage{
calculatePercentContributiontoImpact(x, y)
}
\arguments{
\item{x, }{numeric vector of length n}

\item{y, }{numeric vector of length n}
}
\value{
A dataframe sorted by contribution (high-to-low), also showing "x","y","impact"
}
\description{
Dot multiplies two vectors to calculate an impact score and the percent contribution each score to the total
}
