% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustPrice.R
\name{calculateProducerbyPurchaserPriceRatio}
\alias{calculateProducerbyPurchaserPriceRatio}
\title{Calculate producer to purchaser price ratio.}
\usage{
calculateProducerbyPurchaserPriceRatio(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A dataframe of producer to purchaser price ratio.
}
\description{
Calculate producer to purchaser price ratio.
}
