% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{calculateProductofLeontiefAndProductionDemand}
\alias{calculateProductofLeontiefAndProductionDemand}
\title{Calculate scaling vector with appropriate production demand vector}
\usage{
calculateProductofLeontiefAndProductionDemand(model, use_domestic)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes}

\item{use_domestic, }{a logical value indicating whether to use domestic demand vector}
}
\value{
c, a numeric vector with total $ values for each sector in model
}
\description{
Calculate scaling vector with appropriate production demand vector
}
