% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{calculateSectorContributiontoImpact}
\alias{calculateSectorContributiontoImpact}
\title{Calculate the percent contribution of sectors to an N indicator result
Uses model L matrix for total requirements and D matrix for direct indicator result}
\usage{
calculateSectorContributiontoImpact(model, sector, indicator, domestic = FALSE)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes}

\item{sector, }{str, index of a model sector for use in the M matrix, e.g. "221100/us"}

\item{indicator, }{str, index of a model indicator for use in the C matrix, e.g. "Acidification Potential"}

\item{domestic, }{boolean, sets model to use domestic flow matrix.  Default is FALSE.}
}
\value{
A dataframe sorted from highest process contribution "contribution", also showing "x","y","impact"
}
\description{
Calculate the percent contribution of sectors to an N indicator result
Uses model L matrix for total requirements and D matrix for direct indicator result
}
