% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{calculateSectorPurchasedbySectorSourcedImpact}
\alias{calculateSectorPurchasedbySectorSourcedImpact}
\title{Calculates sector x sector impacts from a given demand vector.}
\usage{
calculateSectorPurchasedbySectorSourcedImpact(y, model, indicator)
}
\arguments{
\item{y}{a model demand vector or user-specified demand vector
that must have the same dimension with the model demand vector}

\item{model}{A complete EEIO model: a list with USEEIO model components and attributes}

\item{indicator}{str, index of a model indicator, e.g. "Acidification Potential"}
}
\value{
A matrix of impacts in the form of sector purchased x sector sourced,
negative values should be interpreted as "reduced impacts".
}
\description{
Calculates sector x sector impacts from a given demand vector.
}
