% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIOFunctions.R
\name{calculateWIOOutputs}
\alias{calculateWIOOutputs}
\title{Calculate the total industry and commodity outputs for the expanded WIO tables}
\usage{
calculateWIOOutputs(model, WIO)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{WIO}{A list with WIO specifications and data}
}
\value{
A model object which contain the model$Commodity or model$Industry objects with WIO sectors
}
\description{
Calculate the total industry and commodity outputs for the expanded WIO tables
}
