% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustPrice.R
\name{calculateYearbyModelIOYearPriceRatio}
\alias{calculateYearbyModelIOYearPriceRatio}
\title{Calculate year by model IO year price ratio.}
\usage{
calculateYearbyModelIOYearPriceRatio(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A data.frame of year by model IO year price ratio.
}
\description{
Calculate year by model IO year price ratio.
}
