% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalImportFactors.R
\name{castImportFactors}
\alias{castImportFactors}
\title{Converts import factors table (of commodities) into flows x sector matrix-like format}
\usage{
castImportFactors(IFTable, model)
}
\arguments{
\item{IFTable, }{dataframe of import factors}

\item{model}{An EEIO model object with model specs, IO tables, and matrices loaded}
}
\value{
A matrix of flows x sector
}
\description{
Converts import factors table (of commodities) into flows x sector matrix-like format
}
