% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Census_ManufacturingMiningSectors_NAICSCodeName}
\alias{Census_ManufacturingMiningSectors_NAICSCodeName}
\title{Census Manufacturing and Mining Sectors NAICS code and name table}
\format{A dataframe with 8257 obs. and 2 variables:
\describe{
 \item{NAICS_Code}{numeric value}
 \item{NAICS_Name}{text code}
}}
\usage{
Census_ManufacturingMiningSectors_NAICSCodeName
}
\description{
Census Manufacturing and Mining Sectors NAICS code and name table
}
\keyword{datasets}
