% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{checkDuplicateFlowsBySector}
\alias{checkDuplicateFlowsBySector}
\title{Check duplicates across satellite tables.}
\usage{
checkDuplicateFlowsBySector(sattable_ls)
}
\arguments{
\item{sattable_ls}{A list of satellite tables}
}
\value{
Messages about whether there are duplicates across satellite tables
}
\description{
Check duplicates across satellite tables.
}
