% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIndicators.R
\name{checkIndicatorforFlows}
\alias{checkIndicatorforFlows}
\title{Checks an LCIA indicator to ensure that flows exist in the model for that indicator}
\usage{
checkIndicatorforFlows(factors, flows)
}
\arguments{
\item{factors}{A dataframe of indicator characterization factors}

\item{flows}{A dataframe of flows from model SatelliteTables}
}
\description{
Checks an LCIA indicator to ensure that flows exist in the model for that indicator
}
