% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{checkNamesandOrdering}
\alias{checkNamesandOrdering}
\title{Check order of names (n1 and n2). Stop function execution if n1 != n2.}
\usage{
checkNamesandOrdering(n1, n2, note)
}
\arguments{
\item{n1}{Name vector #1}

\item{n2}{Name vector #2}

\item{note}{Note about n1 and n2}
}
\description{
Check order of names (n1 and n2). Stop function execution if n1 != n2.
}
