% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIOFunctions.R
\name{checkWIOBalance}
\alias{checkWIOBalance}
\title{Check balance of WIO waste and recycling sectors}
\usage{
checkWIOBalance(model, sectorType = "Waste")
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{sectorType}{A string that indicates to compare the balance of waste or recycling generation and treatment sectors.}
}
\value{
A model with the UseTransactions matrix modified with WIO specs.
}
\description{
Check balance of WIO waste and recycling sectors
}
