% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{collapseTBS}
\alias{collapseTBS}
\title{Collapse a totals by sector table so that each flow sector combination exists only once}
\usage{
collapseTBS(tbs, model)
}
\arguments{
\item{tbs}{totals by sector sourced from satellite table}

\item{model}{An EEIO model object with model specs and IO table loaded}
}
\value{
aggregated totals by sector
}
\description{
Collapse a totals by sector table so that each flow sector combination exists only once
}
