% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQualityFunctions.R
\name{combineDQItoString}
\alias{combineDQItoString}
\title{Collapses 3d DQI matrices into a single string rounded to a single decimal}
\usage{
combineDQItoString(dqi_3d)
}
\arguments{
\item{dqi_3d}{A 3d matrix of dqi scores.}
}
\value{
A single matrix of collapsed dqi scores as string
}
\description{
Collapses 3d DQI matrices into a single string rounded to a single decimal
}
