% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateiorFunctions.R
\name{compare2RCommodityTotals}
\alias{compare2RCommodityTotals}
\title{Compare commodity totals between the specified 2R model objects}
\usage{
compare2RCommodityTotals(q_One, q_Two)
}
\arguments{
\item{q_One}{A vector of commodity totals derived from specific model object}

\item{q_Two}{A vector of commodity totals dervied from a different model object than q_One}
}
\value{
A list of sectors that failed the comparison between the two specified q vectors.
}
\description{
Compare commodity totals between the specified 2R model objects
}
