% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateiorFunctions.R
\name{compare2RVectorTotals}
\alias{compare2RVectorTotals}
\title{Compare  totals between the specified 2R model vectors}
\usage{
compare2RVectorTotals(v_One, v_Two)
}
\arguments{
\item{v_One}{A vector of totals derived from specific 2R model object}

\item{v_Two}{A vector of totals dervied from a different 2R model object than v_One}
}
\value{
A list of sectors that failed the comparison between the two specified q vectors.
}
\description{
Compare  totals between the specified 2R model vectors
}
