% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{compareCommodityOutputandDomesticUseplusProductionDemand}
\alias{compareCommodityOutputandDomesticUseplusProductionDemand}
\title{Compares the total commodity output against the summation of model domestic Use and production demand}
\usage{
compareCommodityOutputandDomesticUseplusProductionDemand(
  model,
  tolerance = 0.05
)
}
\arguments{
\item{model, }{EEIOmodel object completely built}
}
\value{
vector, a vector of relative different in calculation from sector output by sector
}
\description{
Compares the total commodity output against the summation of model domestic Use and production demand
}
