% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{compareDisaggValues}
\alias{compareDisaggValues}
\title{Allocate values specified by the .yml disaggregation specs to the correct places in a disaggregated row/column of the Use/Make tables.}
\usage{
compareDisaggValues(df0, df1, abs_diff = TRUE, tolerance)
}
\arguments{
\item{df0}{data frame 0}
}
\value{

}
\description{
Allocate values specified by the .yml disaggregation specs to the correct places in a disaggregated row/column of the Use/Make tables.
}
