% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{compareEandDomesticLCIResult}
\alias{compareEandDomesticLCIResult}
\title{Compares the total flows against the model result calculation with the domestic demand vector and direct perspective}
\usage{
compareEandDomesticLCIResult(model, tolerance = 0.05)
}
\arguments{
\item{model, }{EEIOmodel object completely built}
}
\description{
Compares the total flows against the model result calculation with the domestic demand vector and direct perspective
}
