% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareModels.R
\name{compareFlowTotals}
\alias{compareFlowTotals}
\title{Compare flow totals for two models}
\usage{
compareFlowTotals(modelA, modelB)
}
\arguments{
\item{modelA, }{a useeior model}

\item{modelB}{a useeior model}
}
\value{
df_confrontation, the validate::confrontation result of the flow total comparison
}
\description{
Compare flow totals for two models
}
