% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{compareMatrices}
\alias{compareMatrices}
\title{Compare two matrices, calculate percentage difference (m1-m2)/m1.
Dimensions of the two matrices must be the same.}
\usage{
compareMatrices(m1, m2, percentage_diff = FALSE)
}
\description{
Compare two matrices, calculate percentage difference (m1-m2)/m1.
Dimensions of the two matrices must be the same.
}
