% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{compareModelResult}
\alias{compareModelResult}
\title{Validate df1 against df0 based on specified conditions}
\usage{
compareModelResult(df0, df1, abs_diff = TRUE, tolerance)
}
\arguments{
\item{df0}{A data.frame to be compared against}

\item{df1}{A data.frame to be validated}

\item{abs_diff}{A boolean value indicating whether to compare the absolute difference between df1 and df0}

\item{tolerance}{A numeric value setting tolerance of the comparison}
}
\value{
A list contains confrontation details and validation results
}
\description{
Validate df1 against df0 based on specified conditions
}
