% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{compareOutputandDomesticResult}
\alias{compareOutputandDomesticResult}
\title{Compares the total sector output against the model result calculation with the domestic demand vector and direct perspective}
\usage{
compareOutputandDomesticResult(model, tolerance = 0.05)
}
\arguments{
\item{model, }{EEIOmodel object completely built}
}
\value{
vector, a vector of relative different in calculation from sector output by sector
}
\description{
Compares the total sector output against the model result calculation with the domestic demand vector and direct perspective
}
