% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{compareOutputandLeontiefXDemand}
\alias{compareOutputandLeontiefXDemand}
\title{Compares the total sector output against the model result calculation with the demand vector. and direct perspective.
Uses the model$FinalDemand and model$L
Works for the domestic model with the equivalent tables}
\usage{
compareOutputandLeontiefXDemand(model, use_domestic = FALSE, tolerance = 0.05)
}
\arguments{
\item{model, }{EEIOmodel object completely built}
}
\value{
a vector of relative different in calculation from sector output by sector
}
\description{
Compares the total sector output against the model result calculation with the demand vector. and direct perspective.
Uses the model$FinalDemand and model$L
Works for the domestic model with the equivalent tables
}
