% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{compareOutputfromMakeandUse}
\alias{compareOutputfromMakeandUse}
\title{Compare commodity or industry output calculated from Make and Use tables.}
\usage{
compareOutputfromMakeandUse(model, output_type = "Commodity")
}
\arguments{
\item{model}{A model list object with model specs and IO tables listed}

\item{output_type}{A string indicating commodity or industry output.}
}
\value{
A vector of relative difference between commodity or industry output
calculated from Supply and Use tables.
}
\description{
Compare commodity or industry output calculated from Make and Use tables.
}
