% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSatellites.R
\name{conformTbStoIOSchema}
\alias{conformTbStoIOSchema}
\title{Take a totals-by-sector df and maps flows to the model schema}
\usage{
conformTbStoIOSchema(tbs, sat_spec, model)
}
\arguments{
\item{tbs, }{totals-by-sector df}

\item{sat_spec, }{a standard specification for a single satellite table}

\item{model}{an EEIO model with IO tables loaded}
}
\value{
a totals-by-sector df with the sectors and flow amounts corresponding to the model schema
}
\description{
Take a totals-by-sector df and maps flows to the model schema
}
