% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{conformTbStoStandardSatTable}
\alias{conformTbStoStandardSatTable}
\title{Generate a standard satellite table with coefficients (kg/$) and only columns completed in the original satellite table.}
\usage{
conformTbStoStandardSatTable(sattable)
}
\arguments{
\item{sattable}{A satellite table contains FlowAmount already aggregated and transformed to coefficients.}
}
\value{
A standard satellite table with coefficients (kg/$) and only columns completed in the original satellite table.
}
\description{
Generate a standard satellite table with coefficients (kg/$) and only columns completed in the original satellite table.
}
