% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{convertUsefromPURtoBAS}
\alias{convertUsefromPURtoBAS}
\title{Convert Use table in the Supply-Use framework from purchasers' price (PUR)
to basic price (BAS)}
\usage{
convertUsefromPURtoBAS(UseSUT_PUR, specs, io_codes)
}
\arguments{
\item{UseSUT_PUR, }{a Use table (from the Supply-Use framework) in purchasers' price (PUR)}

\item{specs, }{model specifications.}

\item{io_codes, }{a list of BEA IO codes.}
}
\value{
A Use table in basic price (BAS)
}
\description{
Convert Use table in the Supply-Use framework from purchasers' price (PUR)
to basic price (BAS)
}
