% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQualityFunctions.R
\name{createBdqi}
\alias{createBdqi}
\title{Creates a 3d matrix of data quality scores (flow by sector)}
\usage{
createBdqi(model)
}
\arguments{
\item{model}{A complete EEIO model object with TbS}
}
\value{
A 3d matrix, third dimension is the individual data quality indicator
}
\description{
Creates a 3d matrix of data quality scores (flow by sector)
}
