% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioeconomyFunctions.R
\name{createBioeconomyModel}
\alias{createBioeconomyModel}
\title{Creates BioEconomy model.}
\usage{
createBioeconomyModel(
  modelname,
  newSectorCode,
  newSectorName,
  similarSectorCode,
  percentage,
  inputPurchases,
  newEnvData
)
}
\arguments{
\item{modelname}{Name of the model from a config file.}

\item{newSectorCode}{string/character that refers to the code/identifier that will be used for the new sector in the matrices.}

\item{newSectorName}{string/character that refers to the name given to the new sector.}

\item{similarSectorCode}{string/character that refers to the code/identifier for the similar existing sector.}

\item{percentage}{numeric (0,1] that refers to the \% of output of the original sector that new sector will produce.}

\item{inputPurchases}{(n+1)x 1 column vector with the amount spent in each of the n existing commodities to produce "Total Industry Output" of the new sector.}

\item{newEnvData}{(# environmental flows x 1) column vector with the data for all the environmental flows per dollar of output for the new sector.}
}
\value{
A list with USEEIO model components and attributes modified and ready for calculation.
}
\description{
This function prepares (read and organize data from the BEA tables), modify the Make and Use table to add the new sector, 
build the model (create the required matrices) and update B and W matrices after its construction to return a bioeconomy model
ready for calculation.
}
