% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{createBlankIntersection}
\alias{createBlankIntersection}
\title{Creates an empty dataframe matrix of disaggregated sectors.}
\usage{
createBlankIntersection(newSectorCodes)
}
\arguments{
\item{newSectorCodes}{vector of named disaggregated sectors}
}
\value{
square dataframe matrix with new sectors as row and column names
}
\description{
Creates an empty dataframe matrix of disaggregated sectors.
}
