% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{createCfromFactorsandBflows}
\alias{createCfromFactorsandBflows}
\title{Generate C matrix from indicator factors and a model B matrix}
\usage{
createCfromFactorsandBflows(factors, B_flows)
}
\arguments{
\item{factors, }{df in model$Indicators$factors format}

\item{B, }{the model B matrix to use for reference}
}
\value{
a C matrix in indicator x flow format
}
\description{
Generate C matrix from indicator factors and a model B matrix
}
