% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDemandVectors.R
\name{createDemandID}
\alias{createDemandID}
\title{Creates an ID for a demand vector based on user provided demandmeta data}
\usage{
createDemandID(demandmeta)
}
\arguments{
\item{demandmeta, }{a data frame of metadata with one row for a demand with
columns Type, System, Name, Year, Location and ID}
}
\value{
character vector with a name
}
\description{
Creates an ID for a demand vector based on user provided demandmeta data
}
