% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQualityFunctions.R
\name{createMdqi}
\alias{createMdqi}
\title{Creates M_dqi matrix from B_dqi}
\usage{
createMdqi(model)
}
\arguments{
\item{model}{A complete EEIO model object.}
}
\value{
A 3d matrix of dqi scores for M
}
\description{
Using the formula: ((B * B_dqi) \%*\% L) / M
}
