% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQualityFunctions.R
\name{createNdqi}
\alias{createNdqi}
\title{Creates N_dqi matrix from D_dqi using the formula
((D * D_dqi) %*% L) / N}
\usage{
createNdqi(model)
}
\arguments{
\item{model}{A complete EEIO model object.}
}
\value{
A 3d matrix of dqi scores for N
}
\description{
Creates N_dqi matrix from D_dqi using the formula
((D * D_dqi) %*% L) / N
}
