% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigDocumentation.R
\docType{data}
\name{DataSourceConfiguration}
\alias{DataSourceConfiguration}
\title{DataSourceConfiguration}
\format{
A list of data source attributes
\describe{
 \item{Title}{string, Name of source},
 \item{Author}{string, name or orgs or individuals}
 \item{DataYear}{numeric, year of data},
 \item{URL}{string, URL data retrieved from}
 \item{Primary}{boolean, TRUE if this is the primary source}
}
}
\usage{
DataSourceConfiguration
}
\description{
DataSourceConfiguration
}
\keyword{datasets}
