% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigDocumentation.R
\docType{data}
\name{DemandConfiguration}
\alias{DemandConfiguration}
\title{Demand Configuration}
\format{
A list of demand attibutes
\describe{
 \item{DemandYear}{numeric, year of demand}
 \item{DemandSubsystems}{list of names of demand subsystems}
}
}
\usage{
DemandConfiguration
}
\description{
Demand Configuration
}
\keyword{datasets}
