% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandFunctions.R
\docType{data}
\name{DemandVectorFunctionRegistry}
\alias{DemandVectorFunctionRegistry}
\title{Registry of functions that construct various demand vector in the form of as a named list with nested names
as keys and function name as values}
\format{
An object of class \code{list} of length 2.
}
\usage{
DemandVectorFunctionRegistry
}
\description{
Registry of functions that construct various demand vector in the form of as a named list with nested names
as keys and function name as values
}
\keyword{datasets}
