% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Margins.R
\name{deriveFinalConsumerMarginSectorImpacts}
\alias{deriveFinalConsumerMarginSectorImpacts}
\title{Prepare M and N matrices with sector margin impacts}
\usage{
deriveFinalConsumerMarginSectorImpacts(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A list with M_margin and N_margin
}
\description{
Prepare M and N matrices with sector margin impacts
}
