% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMarginImpacts.R
\name{deriveMarginSectorImpacts}
\alias{deriveMarginSectorImpacts}
\title{Prepare M and U matrices with sector margin impacts}
\usage{
deriveMarginSectorImpacts(model, margin_type = "industry")
}
\arguments{
\item{model}{A complete EEIO model}

\item{margin_type}{A character value: can be "industry" or "final consumer".}
}
\value{
A list with M_margin and U_margin
}
\description{
Prepare M and U matrices with sector margin impacts
}
