% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalImportFactors.R
\name{deriveMMatrix}
\alias{deriveMMatrix}
\title{Derives an aggregate M matrix from M_d and M_m based on the Consumption demand vector and
FINAL perspective. Results from this M matrix match those calculated using the Import Emission
Factors when using the Consumption demand vector and FINAL perspective.}
\usage{
deriveMMatrix(model)
}
\arguments{
\item{model, }{An EEIO model object with model specs and crosswalk table loaded}
}
\value{
An M matrix of flows x sector
}
\description{
Derives an aggregate M matrix from M_d and M_m based on the Consumption demand vector and
FINAL perspective. Results from this M matrix match those calculated using the Import Emission
Factors when using the Consumption demand vector and FINAL perspective.
}
