% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_GrossOutput_IO}
\alias{Detail_GrossOutput_IO}
\title{Detail 2007-2017 Gross Output (2012 schema)}
\format{A data frame with 405 obs. and 11 variables}
\usage{
Detail_GrossOutput_IO
}
\description{
TBD
}
\keyword{datasets}
