% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_GrossOutput_IO_17sch}
\alias{Detail_GrossOutput_IO_17sch}
\title{Detail 1997-2023 Gross Output (2017 schema)}
\format{
A dataframe with 402 obs. and 27 variables
}
\source{
\url{https://apps.bea.gov/industry/Release/ZIP/UGdpByInd.zip}
}
\usage{
Detail_GrossOutput_IO_17sch
}
\description{
Detail 1997-2023 Gross Output (2017 schema)
}
\keyword{datasets}
