% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_Import_2012_BeforeRedef}
\alias{Detail_Import_2012_BeforeRedef}
\title{Detail 2012 Import Before Redefinition (2012 schema)}
\format{
A dataframe with 405 obs. and 427 variables
}
\source{
\url{https://apps.bea.gov/industry/xls/io-annual/ImportMatrices_Before_Redefinitions_DET_2007_2012.xlsx}
}
\usage{
Detail_Import_2012_BeforeRedef
}
\description{
Detail 2012 Import Before Redefinition (2012 schema)
}
\keyword{datasets}
