% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_Import_2017_BeforeRedef_17sch}
\alias{Detail_Import_2017_BeforeRedef_17sch}
\title{Detail 2017 Import Before Redefinition (2017 schema)}
\format{
A dataframe with 402 obs. and 424 variables
}
\source{
\url{https://apps.bea.gov/industry/xls/io-annual/ImportMatrices_Before_Redefinitions_DET_2017.xlsx}
}
\usage{
Detail_Import_2017_BeforeRedef_17sch
}
\description{
Detail 2017 Import Before Redefinition (2017 schema)
}
\keyword{datasets}
