% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_Make_2012_BeforeRedef}
\alias{Detail_Make_2012_BeforeRedef}
\title{Detail 2012 Make Before Redefinition (2012 schema)}
\format{A data frame with 406 obs. and 406 variables}
\usage{
Detail_Make_2012_BeforeRedef
}
\description{
TBD
}
\keyword{datasets}
