% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_Make_2017_AfterRedef_17sch}
\alias{Detail_Make_2017_AfterRedef_17sch}
\title{Detail 2017 Make After Redefinition (2017 schema)}
\format{
A dataframe with 403 obs. and 403 variables
}
\source{
\url{https://apps.bea.gov/industry/iTables\%20Static\%20Files/AllTablesIO.zip}
}
\usage{
Detail_Make_2017_AfterRedef_17sch
}
\description{
Detail 2017 Make After Redefinition (2017 schema)
}
\keyword{datasets}
