% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_Margins_2012_BeforeRedef}
\alias{Detail_Margins_2012_BeforeRedef}
\title{Detail Margins (Before Redef) table for 2012 (2012 schema)}
\format{
A dataframe with 61844 obs. and 9 variables:
\describe{
 \item{NIPACode}{text code}
 \item{MarginsCategory}{text category name, like 'Therapeutic medical equipment'}
 \item{CommodityCode}{BEA_2012_Detail_Code}
 \item{CommodityDescription}{BEA_2012_Detail_Name}
 \item{ProducersValue}{USD2012}
 \item{Transportation}{USD2012}
 \item{Wholesale}{USD2012}
 \item{Retail}{USD2012}
 \item{PurchasersValue}{USD2012}
}
}
\source{
\url{https://apps.bea.gov/industry/xls/underlying-estimates/Margins_Before_Redefinitions_2007_2012_DET.xlsx}
}
\usage{
Detail_Margins_2012_BeforeRedef
}
\description{
Detail Margins (Before Redef) table for 2012 (2012 schema)
}
\keyword{datasets}
