% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_Margins_2012_BeforeRedef_17sch}
\alias{Detail_Margins_2012_BeforeRedef_17sch}
\title{Detail Margins (Before Redef) table for 2012 (2017 schema)}
\format{
A dataframe with 58032 obs. and 9 variables:
\describe{
 \item{NIPACode}{text code}
 \item{MarginsCategory}{text category name, like 'Therapeutic medical equipment'}
 \item{CommodityCode}{BEA_2017_Detail_Code}
 \item{CommodityDescription}{BEA_2017_Detail_Name}
 \item{ProducersValue}{USD2017}
 \item{Transportation}{USD2017}
 \item{Wholesale}{USD2017}
 \item{Retail}{USD2017}
 \item{PurchasersValue}{USD2017}
}
}
\source{
\url{https://apps.bea.gov/industry/xls/underlying-estimates/Margins_Before_Redefinitions_2017.xlsx}
}
\usage{
Detail_Margins_2012_BeforeRedef_17sch
}
\description{
Detail Margins (Before Redef) table for 2012 (2017 schema)
}
\keyword{datasets}
