% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_PCE_2012}
\alias{Detail_PCE_2012}
\title{Detail PCE Bridge table for 2012 (2012 schema)}
\format{
A dataframe with 712 obs. and 9 variables:
\describe{
 \item{NIPACode}{text code}
 \item{PCECategory}{text category name, like 'Therapeutic medical equipment'}
 \item{CommodityCode}{BEA_2012_Detail_Code}
 \item{CommodityDescription}{BEA_2012_Detail_Name}
 \item{ProducersValue}{USD2012}
 \item{Transportation}{USD2012}
 \item{Wholesale}{USD2012}
 \item{Retail}{USD2012}
 \item{PurchasersValue}{USD2012}
}
}
\source{
\url{https://apps.bea.gov/industry/xls/underlying-estimates/PCEBridge_2007_2012_DET.xlsx}
}
\usage{
Detail_PCE_2012
}
\description{
Detail PCE Bridge table for 2012 (2012 schema)
}
\keyword{datasets}
