% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_Supply_2012}
\alias{Detail_Supply_2012}
\title{Detail 2012 Supply (2012 schema)}
\format{
A dataframe with 406 obs. and 417 variables
}
\source{
\url{https://apps.bea.gov/industry/iTables\%20Static\%20Files/AllTablesSUP.zip}
}
\usage{
Detail_Supply_2012
}
\description{
Detail 2012 Supply (2012 schema)
}
\keyword{datasets}
