% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_Supply_2017}
\alias{Detail_Supply_2017}
\title{Detail 2017 Supply (2017 schema)}
\format{
A dataframe with 403 obs. and 414 variables
}
\source{
\url{https://apps.bea.gov/industry/iTables\%20Static\%20Files/AllTablesSUP.zip}
}
\usage{
Detail_Supply_2017
}
\description{
Detail 2017 Supply (2017 schema)
}
\keyword{datasets}
