% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_Use_2007_PRO_BeforeRedef}
\alias{Detail_Use_2007_PRO_BeforeRedef}
\title{Detail 2007 Use Producer's Value Before Redefinition (2012 schema)}
\format{A data frame with 428 obs. and 411 variables}
\usage{
Detail_Use_2007_PRO_BeforeRedef
}
\description{
TBD
}
\keyword{datasets}
