% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_Use_2007_PUR_AfterRedef}
\alias{Detail_Use_2007_PUR_AfterRedef}
\title{Detail 2007 Use Purchaser's Value After Redefinition (2012 schema)}
\format{A data frame with 403 obs. and 428 variables}
\usage{
Detail_Use_2007_PUR_AfterRedef
}
\description{
TBD
}
\keyword{datasets}
