% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_Use_2012_PRO_AfterRedef}
\alias{Detail_Use_2012_PRO_AfterRedef}
\title{Detail 2012 Use Producer's Value After Redefinition (2012 schema)}
\format{A dataframe with 411 obs. and 428 variables}
\usage{
Detail_Use_2012_PRO_AfterRedef
}
\description{
TBD
}
\keyword{datasets}
