% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_Use_2012_PRO_BeforeRedef}
\alias{Detail_Use_2012_PRO_BeforeRedef}
\title{Detail 2012 Use Producer's Value Before Redefinition (2012 schema)}
\format{
A dataframe with 411 obs. and 428 variables
}
\source{
\url{https://apps.bea.gov//industry/iTables\%20Static\%20Files/AllTablesIO.zip}
}
\usage{
Detail_Use_2012_PRO_BeforeRedef
}
\description{
TBD
}
\keyword{datasets}
