% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_Use_2012_PUR_BeforeRedef}
\alias{Detail_Use_2012_PUR_BeforeRedef}
\title{Detail 2012 Use Purchaser's Value Before Redefinition (2012 schema)}
\format{A data frame with 403 obs. and 428 variables}
\usage{
Detail_Use_2012_PUR_BeforeRedef
}
\description{
TBD
}
\keyword{datasets}
