% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_Use_SUT_2012}
\alias{Detail_Use_SUT_2012}
\title{Detail 2012 Use (under the Supply-Use framework, 2012 schema)}
\format{
A dataframe with 414 obs. and 426 variables
}
\source{
\url{https://apps.bea.gov/industry/iTables\%20Static\%20Files/AllTablesSUP.zip}
}
\usage{
Detail_Use_SUT_2012
}
\description{
Detail 2012 Use (under the Supply-Use framework, 2012 schema)
}
\keyword{datasets}
