% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_Use_SUT_2017_17sch}
\alias{Detail_Use_SUT_2017_17sch}
\title{Detail 2017 Use (under the Supply-Use framework, 2017 schema)}
\format{
A dataframe with 411 obs. and 423 variables
}
\source{
\url{https://apps.bea.gov/industry/iTables\%20Static\%20Files/AllTablesSUP.zip}
}
\usage{
Detail_Use_SUT_2017_17sch
}
\description{
Detail 2017 Use (under the Supply-Use framework, 2017 schema)
}
\keyword{datasets}
