% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_ValueAddedCodeName_2012}
\alias{Detail_ValueAddedCodeName_2012}
\title{Detail Value Added Code and Name (2012 schema)}
\format{
A dataframe with 3 obs. and 2 variables
}
\source{
\url{https://apps.bea.gov//industry/iTables\%20Static\%20Files/AllTablesIO.zip}
}
\usage{
Detail_ValueAddedCodeName_2012
}
\description{
TBD
}
\keyword{datasets}
