% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateCol}
\alias{disaggregateCol}
\title{Disaggregate a single column from a table.}
\usage{
disaggregateCol(
  originalColVector,
  disagg_specs,
  duplicate = FALSE,
  notUniform = FALSE
)
}
\arguments{
\item{originalColVector}{A dataframe containing the column to disaggregate}

\item{disagg_specs}{Specifications for disaggregating the current Table}

\item{duplicate}{A flag that indicates whether the disaggregated columns are to be duplicated or not (e.g. for CPI values)}

\item{notUniform}{A flag that indicates whether the disaggregated columns are to be disaggregated in uniform manner or not}
}
\value{
A dataframe with the original column disaggregated.
}
\description{
Disaggregate a single column from a table.
}
