% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateiorFunctions.R
\name{disaggregateCPI}
\alias{disaggregateCPI}
\title{Disaggregate CPI table to ensure the correct dimensions}
\usage{
disaggregateCPI(df, model)
}
\arguments{
\item{df, }{CPI table}

\item{model}{An EEIO form USEEIO model object with model specs and IO meta data loaded.}
}
\value{
An expanded CPI table with values replicated for disaggregated sectors.
}
\description{
Disaggregate CPI table to ensure the correct dimensions
}
