% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregatedRatios}
\alias{disaggregatedRatios}
\title{Calculate ratios of throughputs from the disaggregated sectors}
\usage{
disaggregatedRatios(model, disagg, output_type = "Commodity")
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}

\item{output_type}{A string value indicating whether to obtain "Commodity" or "Industry" ratios}
}
\value{
disaggRatios A dataframe which contain the disaggregated ratios for the disaggregated sectors
}
\description{
Calculate ratios of throughputs from the disaggregated sectors
}
