% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateInternationalTradeAdjustment}
\alias{disaggregateInternationalTradeAdjustment}
\title{Disaggregate model$InternationalTradeAdjustments vector in the main model object}
\usage{
disaggregateInternationalTradeAdjustment(
  model,
  disagg,
  ratios = NULL,
  adjustmentByCommodity = FALSE
)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}

\item{ratios}{Specific ratios to be used for the disaggregation of the InternationalTradeAdjusment object in place of using economic totals to derive the ratios.}

\item{adjustmentByCommodity}{Flag to denote whether to disaggregate the InternationalTradeAdjustmentbyCommodity object which is only present in industry models}
}
\value{
newInternationalTradeAdjustment A vector which contains the InternationalTradeAdjustment for the disaggregated sectors
}
\description{
Disaggregate model$InternationalTradeAdjustments vector in the main model object
}
