% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateMargins}
\alias{disaggregateMargins}
\title{Disaggregate model$Margins dataframe in the main model object}
\usage{
disaggregateMargins(model, disagg)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}
}
\value{
newMargins A dataframe which contain the margins for the disaggregated sectors
}
\description{
Disaggregate model$Margins dataframe in the main model object
}
