% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateMasterCrosswalk}
\alias{disaggregateMasterCrosswalk}
\title{Disaggregate the MasterCrosswalk to include the new sectors for disaggregation}
\usage{
disaggregateMasterCrosswalk(crosswalk, disagg)
}
\arguments{
\item{crosswalk}{MasterCrosswalk from NAICS to BEA.}
}
\value{
crosswalk with new sectors added.
}
\description{
Disaggregate the MasterCrosswalk to include the new sectors for disaggregation
}
