% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateModel}
\alias{disaggregateModel}
\title{Disaggregate a model based on specified source file}
\usage{
disaggregateModel(model, configpaths = NULL)
}
\arguments{
\item{model}{Model file loaded with IO tables}

\item{configpaths}{str vector, paths (including file name) of disagg configuration file(s).
If NULL, built-in config files are used.}
}
\value{
A disaggregated model.
}
\description{
Disaggregate a model based on specified source file
}
