% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateOutputs}
\alias{disaggregateOutputs}
\title{Disaggregate Commodity and Industry Output model objects}
\usage{
disaggregateOutputs(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
model A complete EEIO model: a list with USEEIO model components and attributes.
}
\description{
Disaggregate Commodity and Industry Output model objects
}
