% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateSectorDFs}
\alias{disaggregateSectorDFs}
\title{Disaggregate model$Commodity or model$Industry dataframes in the main model object}
\usage{
disaggregateSectorDFs(model, disagg, list_type)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}

\item{list_type}{string indicating whether to disaggregate model$Industry or model$Commodity dataframe.}
}
\value{
newSectors A dataframe which contain the disaggregated model$Commodity or model$Industry objects
}
\description{
Disaggregate model$Commodity or model$Industry dataframes in the main model object
}
