% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateSectorLists}
\alias{disaggregateSectorLists}
\title{Disaggregate sector lists in the main model object}
\usage{
disaggregateSectorLists(model, disagg)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}
}
\value{
newLists A list of two character arrays which contain the disaggregated commodity and industry codes
}
\description{
Disaggregate sector lists in the main model object
}
