% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateSectorNames}
\alias{disaggregateSectorNames}
\title{Disaggregate SectorNames model objects}
\usage{
disaggregateSectorNames(model, disagg, newNames)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}

\item{newNames}{Dataframe contaning the sector codes and names for the disaggregated sectors}
}
\value{
newSectorNames A dataframe the disaggregated sector names without the location component
}
\description{
Disaggregate SectorNames model objects
}
