% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateSetup}
\alias{disaggregateSetup}
\title{Setup the configuration specs based on the input files}
\usage{
disaggregateSetup(model, configpaths = NULL)
}
\arguments{
\item{model}{Model file loaded with IO tables}

\item{configpaths}{str vector, paths (including file name) of disagg configuration file(s).
If NULL, built-in config files are used.}
}
\value{
A model object with the correct disaggregation specs.
}
\description{
Setup the configuration specs based on the input files
}
