% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateTaxLessSubsidies}
\alias{disaggregateTaxLessSubsidies}
\title{Disaggregate TaxLessSubsidies dataframe in the main model object}
\usage{
disaggregateTaxLessSubsidies(model, disagg)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}
}
\value{
newTLS A dataframe which contain the TaxLessSubsidies including the disaggregated sectors
}
\description{
Disaggregate TaxLessSubsidies dataframe in the main model object
}
