% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateUseTable}
\alias{disaggregateUseTable}
\title{Disaggregate Use table based on specs}
\usage{
disaggregateUseTable(model, disagg, domestic = FALSE)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}

\item{domestic}{A logical value indicating whether to disaggregate domestic final demand.}
}
\value{
A standardized make table with old sectors removed and new sectors added.
}
\description{
Disaggregate Use table based on specs
}
