% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateVA}
\alias{disaggregateVA}
\title{Disaggregate Value Added based on specs}
\usage{
disaggregateVA(model, disagg)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}
}
\value{
A standardized Vale Added table with old sectors removed and new sectors with manual and default allocations added.
}
\description{
Disaggregate Value Added based on specs
}
