% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{downloadDataCommonsfile}
\alias{downloadDataCommonsfile}
\title{downloads files from the Data Commons and stores in a local temporary data directory}
\usage{
downloadDataCommonsfile(source, subdirectory)
}
\arguments{
\item{source}{The name of the source file incluidng any subfolders (e.g. "traci_2_1/TRACI_2.1_v1.parquet")}

\item{subdirectory}{The name of the package where the source file is stored on Data Commons (e.g. "lciafmt")}
}
\description{
downloads files from the Data Commons and stores in a local temporary data directory
}
