% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{downloadDataFile}
\alias{downloadDataFile}
\title{Downloads files from a given source and stores in a local data directory}
\usage{
downloadDataFile(source, subdirectory, url, query_string, localpath)
}
\arguments{
\item{source}{The name of the source file (e.g. "TRACI_2.1_v1.parquet")}

\item{subdirectory}{The name of the subdirectory where the source file is stored on
remote location including any subfolders (e.g. "lciafmt/traci_2_1"). May be empty.}

\item{url}{The url of the data storage location}

\item{query_string}{A string for URLs where query string is present after file name.
e.g. "?download=1". May be empty.}

\item{localpath}{A string value looked up from LocalStorageConfig with a path for local storage
like "flowsa/FlowBySector"}
}
\value{
NULL if successful
}
\description{
Downloads files from a given source and stores in a local data directory
}
