% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrosswalkFunctions.R
\name{downloadNAICS2to6DigitsFile}
\alias{downloadNAICS2to6DigitsFile}
\title{Download NAICS file for 2-6 digit NAICS codes if not present.}
\usage{
downloadNAICS2to6DigitsFile(year)
}
\arguments{
\item{year}{int, 2017, 2012, or 2007 accepted.}
}
\value{
FileName str
}
\description{
Download NAICS file for 2-6 digit NAICS codes if not present.
}
