% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandFunctions.R
\name{extractAndFormatDemandVector}
\alias{extractAndFormatDemandVector}
\title{Read demand vector from a csv file and format for use in model calculations}
\usage{
extractAndFormatDemandVector(file_path, demand_name, model)
}
\arguments{
\item{file_path}{str, path to csv file containing demand data}

\item{demand_name}{str, name of demand data as field header}

\item{model, }{a model}
}
\value{
a demand vector formatted for use in calculating model results
}
\description{
Read demand vector from a csv file and format for use in model calculations
}
