% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{extractValidationResult}
\alias{extractValidationResult}
\title{Extract validation passes or failures}
\usage{
extractValidationResult(validation, failure = TRUE)
}
\arguments{
\item{validation}{A data.frame contains validation details}

\item{failure}{A logical value indicating whether to report failure or not}
}
\value{
A data.frame contains validation results
}
\description{
Extract validation passes or failures
}
