% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{FAF_2014}
\alias{FAF_2014}
\title{FAF table for 2014}
\format{
A dataframe with 549137 obs. and 12 variables
\describe{
 \item{fr_orig}{integer}
 \item{dms_origst}{integer}
 \item{dms_destst}{integer}
 \item{fr_dest}{integer}
 \item{fr_inmode}{integer}
 \item{dms_mode}{integer}
 \item{fr_outmode}{integer}
 \item{sctg2}{integer}
 \item{trade_type}{integer}
 \item{tons_2014}{numeric}
 \item{value_2014}{numeric}
 \item{tmiles_2014}{numeric}
}
}
\source{
\url{https://faf.ornl.gov/fafweb/}
}
\usage{
FAF_2014
}
\description{
FAF table for 2014
}
\keyword{datasets}
