% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{FAF_2016}
\alias{FAF_2016}
\title{FAF table for 2016}
\format{
A dataframe with 550886 obs. and 12 variables
\describe{
 \item{fr_orig}{integer}
 \item{dms_origst}{integer}
 \item{dms_destst}{integer}
 \item{fr_dest}{integer}
 \item{fr_inmode}{integer}
 \item{dms_mode}{integer}
 \item{fr_outmode}{integer}
 \item{sctg2}{integer}
 \item{trade_type}{integer}
 \item{tons_2016}{numeric}
 \item{value_2016}{numeric}
 \item{tmiles_2016}{numeric}
}
}
\source{
\url{https://faf.ornl.gov/fafweb/}
}
\usage{
FAF_2016
}
\description{
FAF table for 2016
}
\keyword{datasets}
