% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigurationFunctions.R
\name{findModelConfigurationFiles}
\alias{findModelConfigurationFiles}
\title{Get model config files}
\usage{
findModelConfigurationFiles()
}
\value{
vector of model config files
}
\description{
Get model config files
}
