% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowMappingFunctions.R
\docType{data}
\name{flowmapping}
\alias{flowmapping}
\title{Import flow mapping file.}
\format{
An object of class \code{data.frame} with 3210 rows and 28 columns.
}
\usage{
flowmapping
}
\description{
Import flow mapping file.
}
\keyword{datasets}
