% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandFunctions.R
\name{formatDemandVector}
\alias{formatDemandVector}
\title{Format a named demand vector with partial sectors to have all the rows and ordering needed}
\usage{
formatDemandVector(dv, L)
}
\arguments{
\item{dv, }{a user provided demand vector. See calculateEEIOModel()}

\item{L, }{the L matrix for the given model, used as a reference}
}
\value{
A named vector with values for all names in L and ordered like L
}
\description{
Format a named demand vector with partial sectors to have all the rows and ordering needed
}
