% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormattingFunctions.R
\name{formatLCIAforIOMB}
\alias{formatLCIAforIOMB}
\title{Format a dataframe having LCIA table structure for IOMB process.
Change row names of the dataframe to "code/names/locationcode".}
\usage{
formatLCIAforIOMB(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A dataframe, having LCIA table structure, with row names from "code" to "code/names/locationcode".
}
\description{
Format a dataframe having LCIA table structure for IOMB process.
Change row names of the dataframe to "code/names/locationcode".
}
