% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormattingFunctions.R
\name{formatLocationforStateModels}
\alias{formatLocationforStateModels}
\title{Formats the location by appending "US-" before models}
\usage{
formatLocationforStateModels(r)
}
\arguments{
\item{r, }{string of region name}
}
\value{
r, "US-r"
}
\description{
Formats the location by appending "US-" before models
}
