% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormattingFunctions.R
\name{formatSectorMetaDataforIOMB}
\alias{formatSectorMetaDataforIOMB}
\title{Format the sector meta data of a model for IOMB process.}
\usage{
formatSectorMetaDataforIOMB(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A dataframe, having sector meta data table structure, with column names required by IOMB.
}
\description{
Format the sector meta data of a model for IOMB process.
}
