% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{formatValidationResult}
\alias{formatValidationResult}
\title{Format validation result}
\usage{
formatValidationResult(df, tolerance)
}
\arguments{
\item{df}{A data.frame to be validated}

\item{tolerance}{A numeric value setting tolerance of the comparison}
}
\value{
A list contains formatted validation results
}
\description{
Format validation result
}
