% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{generatDomesticUse}
\alias{generatDomesticUse}
\title{Generate domestic Use table by adjusting Use table based on Import matrix.}
\usage{
generatDomesticUse(Use, specs)
}
\arguments{
\item{Use}{An original Use table.}

\item{specs}{Model specifications.}
}
\value{
A Domestic Use table.
}
\description{
Generate domestic Use table by adjusting Use table based on Import matrix.
}
