% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateiorFunctions.R
\name{generate2RDirectRequirementsfromUseWithTrade}
\alias{generate2RDirectRequirementsfromUseWithTrade}
\title{Generate direct requirements Use table for 2 region models using domestic 
Use table with trade data generated by stateior}
\usage{
generate2RDirectRequirementsfromUseWithTrade(model, domestic)
}
\arguments{
\item{model}{An EEIO form USEEIO model object with model specs and IO meta data loaded.}

\item{domestic}{A logical parameter indicating whether to DR or Domestic DR.}
}
\value{
A 2-region direct requirements table generated using the domestic Use table with trade
}
\description{
Generate direct requirements Use table for 2 region models using domestic 
Use table with trade data generated by stateior
}
