% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{generateCommodityMixMatrix}
\alias{generateCommodityMixMatrix}
\title{Generate Commodity Mix matrix.}
\usage{
generateCommodityMixMatrix(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
Commodity Mix matrix of the model.
}
\description{
Generate Commodity Mix matrix.
}
