% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{generateCommodityOutputforYear}
\alias{generateCommodityOutputforYear}
\title{Generate Commodity output by transforming Industry output using Commodity Mix matrix.}
\usage{
generateCommodityOutputforYear(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A dataframe contains adjusted Commodity output.
}
\description{
Generate Commodity output by transforming Industry output using Commodity Mix matrix.
}
