% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{generateFlowtoDollarCoefficient}
\alias{generateFlowtoDollarCoefficient}
\title{Calculates intensity coefficient (kg/$) for a standard satellite table.}
\usage{
generateFlowtoDollarCoefficient(
  sattable,
  outputyear,
  referenceyear,
  location_acronym,
  IsRoUS = FALSE,
  model,
  output_type = "Industry"
)
}
\arguments{
\item{sattable}{A standardized satellite table with resource and emission names from original sources.}

\item{outputyear}{Year of Industry output.}

\item{referenceyear}{Year of the currency reference.}

\item{location_acronym}{Abbreviated location name of the model, e.g. "US" or "GA".}

\item{IsRoUS}{A logical parameter indicating whether to adjust Industry output for Rest of US (RoUS).}

\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{output_type}{Type of the output, e.g. "Commodity" or "Industry"}
}
\value{
A dataframe contains intensity coefficient (kg/$).
}
\description{
Calculates intensity coefficient (kg/$) for a standard satellite table.
}
