% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{generateImportCostVector}
\alias{generateImportCostVector}
\title{Generate import cost vector from Use and Import matrix.}
\usage{
generateImportCostVector(Use, specs)
}
\arguments{
\item{Use, }{dataframe of a Use table}

\item{specs, }{list of model specifications}
}
\value{
An import cost vector with names as commodity codes
}
\description{
Generate import cost vector from Use and Import matrix.
}
