% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{generateInternationalTradeAdjustmentVector}
\alias{generateInternationalTradeAdjustmentVector}
\title{Generate international trade adjustment vector from Use and Import matrix.}
\usage{
generateInternationalTradeAdjustmentVector(Use, specs)
}
\arguments{
\item{Use, }{dataframe of a Use table}

\item{specs, }{list of model specifications}
}
\value{
An international trade adjustment vector with names as commodity codes
}
\description{
Generate international trade adjustment vector from Use and Import matrix.
}
