% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIndicators.R
\name{generateLCIA}
\alias{generateLCIA}
\title{Generating LCIA output formatted for useeiopy using LCIA_indicators static file}
\usage{
generateLCIA(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A dataframe of the LCIA factors for all indicators used in the model
}
\description{
Generating LCIA output formatted for useeiopy using LCIA_indicators static file
}
