% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{generateMarketSharesfromMake}
\alias{generateMarketSharesfromMake}
\title{Generate Market Shares matrix from Make table.}
\usage{
generateMarketSharesfromMake(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
Market Shares matrix of the model.
}
\description{
Generate Market Shares matrix from Make table.
}
