% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{generateModelIdentifier}
\alias{generateModelIdentifier}
\title{Create a unique hash identifier for a model}
\usage{
generateModelIdentifier(model)
}
\arguments{
\item{model, }{any model object}
}
\value{
char string
}
\description{
Create a unique hash identifier for a model
}
