% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{generateModelSectorSchema}
\alias{generateModelSectorSchema}
\title{Create sector schema for a model}
\usage{
generateModelSectorSchema(model)
}
\arguments{
\item{model, }{any model object}
}
\value{
char string
}
\description{
Create sector schema for a model
}
