% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_IOfunctions.R
\name{generateNonScrapRatios}
\alias{generateNonScrapRatios}
\title{Generate non-scrap ratios}
\usage{
generateNonScrapRatios()
}
\value{
A datframe with rows being model industries and a column for "non_scrap_ratios" for that industry.
}
\description{
Generate non-scrap ratios
}
