% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{generateOutputBasedAllocationFactorsByGroup}
\alias{generateOutputBasedAllocationFactorsByGroup}
\title{Generate output-based allocation factors for a dataframe of BEA codes and a grouping variable.}
\usage{
generateOutputBasedAllocationFactorsByGroup(codeswithgroups)
}
\arguments{
\item{codeswithgroups}{A dataframe contains two columns: "Code" and "Group".}
}
\value{
A dataframe contains codes, groups, and allocation factors.
}
\description{
Generate output-based allocation factors for a dataframe of BEA codes and a grouping variable.
}
