% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_IOfunctions.R
\name{generatePriceAdjustedCommodityOutputforYear}
\alias{generatePriceAdjustedCommodityOutputforYear}
\title{Generate Commodity output by transforming Industry output using Commodity Mix matrix.}
\usage{
generatePriceAdjustedCommodityOutputforYear(outputyear, location_acronym,
  IsRoU, model)
}
\arguments{
\item{outputyear}{Year of Industry output.}

\item{location_acronym}{Abbreviated location name of the model, e.g. "US" or "GA".}

\item{IsRoU}{A logical parameter indicating whether to adjust Industry output for Rest of US (RoU).}

\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A dataframe contains adjusted Commodity output.
}
\description{
Generate Commodity output by transforming Industry output using Commodity Mix matrix.
}
