% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{generateTaxLessSubsidiesTable}
\alias{generateTaxLessSubsidiesTable}
\title{Generate tax less subsidies table using BEA Supply table which include
total product supply in basic price and tax less subsidies for all commodities}
\usage{
generateTaxLessSubsidiesTable(model)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}
}
\value{
A data.frame containing CommodityCode, basic price, tax less subsidies,
and producer price of total product supply
}
\description{
Generate tax less subsidies table using BEA Supply table which include
total product supply in basic price and tax less subsidies for all commodities
}
