% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSatellites.R
\name{generateTbSfromSatSpec}
\alias{generateTbSfromSatSpec}
\title{Reads a satellite table specification and generates a totals-by-sector table}
\usage{
generateTbSfromSatSpec(sat_spec, model)
}
\arguments{
\item{sat_spec, }{a standard specification for a single satellite table}

\item{model}{A model list object with model specs and IO tables listed}
}
\value{
a totals-by-sector dataframe
}
\description{
Reads a satellite table specification and generates a totals-by-sector table
}
