% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICS.R
\name{get2012NAICS7to10DigitsCodeName}
\alias{get2012NAICS7to10DigitsCodeName}
\title{Get 2012 7-10 digit NAICS codes and names (agricultural, manufacturing, and mining industries).}
\usage{
get2012NAICS7to10DigitsCodeName()
}
\value{
data frame with columns NAICS_year_Code and NAICS_year_Name.
}
\description{
Get 2012 7-10 digit NAICS codes and names (agricultural, manufacturing, and mining industries).
}
