% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICS.R
\name{get2012NAICS7to12Digits}
\alias{get2012NAICS7to12Digits}
\title{Gets 2012 NAICS codes from 7 to 12 digits in the form of a crosswalk.}
\usage{
get2012NAICS7to12Digits()
}
\value{
data frame with columns NAICS_year_Code and NAICS_year_Name.
}
\description{
Gets 2012 NAICS codes from 7 to 12 digits in the form of a crosswalk.
}
