% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{getAggregationSpecs}
\alias{getAggregationSpecs}
\title{Obtain aggregation specs from input files}
\usage{
getAggregationSpecs(model)
}
\arguments{
\item{model}{Model file loaded with IO tables}
}
\value{
A model with the specified aggregation and disaggregation specs.
}
\description{
Obtain aggregation specs from input files
}
