% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateOutput.R
\name{getAlternativeStateCommodityOutputEstimates}
\alias{getAlternativeStateCommodityOutputEstimates}
\title{Estimate state commodity output}
\usage{
getAlternativeStateCommodityOutputEstimates(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A dataframe contains state commodity output for specified state with row names being BEA sector code.
}
\description{
Estimate state commodity output
}
