% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationFunctions.R
\name{getBEASectorColorMapping}
\alias{getBEASectorColorMapping}
\title{Uses VizualizationEssentials.yml to get a mapping of the to the BEA Sector Color scheme}
\usage{
getBEASectorColorMapping(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes}
}
\value{
df with mapping with model BaseIOLevel codes to BEA Sector Codes, Names, and colors
}
\description{
Uses VizualizationEssentials.yml to get a mapping of the to the BEA Sector Color scheme
}
