% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioeconomyFunctions.R
\name{getColIndex}
\alias{getColIndex}
\title{Obtains the column index of a given industry or commodity in the Make or Use table}
\usage{
getColIndex(table, code)
}
\arguments{
\item{table}{the Make or the Use table}

\item{code}{the identifier BEA code (string)}
}
\value{
the column index of the commodity/industry indicated through the code.
}
\description{
Obtains the column index of a given industry or commodity in the Make or Use table
}
