% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{getDefaultAllocationPercentages}
\alias{getDefaultAllocationPercentages}
\title{Obtain a vector of allocation percentages from the specified source file based on disaggregations specifications.}
\usage{
getDefaultAllocationPercentages(FileDF, disagg, numNewSectors, output)
}
\arguments{
\item{FileDF}{dataframe of Make or Use disaggregation data}

\item{disagg}{Specifications for disaggregating the current Table}

\item{numNewSectors}{Int. Number of new sectors in the disaggregation}

\item{output}{String indicating whether allocation values should reference "Commodity" or "Industry" outputs by default}
}
\value{
vector of allocation percentages
}
\description{
Obtain a vector of allocation percentages from the specified source file based on disaggregations specifications.
}
