% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{getDisaggCommodityPercentages}
\alias{getDisaggCommodityPercentages}
\title{Obtain default disaggregation percentages for commodities from the disaggregation input files.}
\usage{
getDisaggCommodityPercentages(disagg)
}
\arguments{
\item{disagg}{Specifications for disaggregating the current Model}
}
\value{
A dataframe with the default disaggregation percentages for the Commodities of the current model
}
\description{
Obtain default disaggregation percentages for commodities from the disaggregation input files.
}
