% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{getDisaggIndustryPercentages}
\alias{getDisaggIndustryPercentages}
\title{Obtain default disaggregation percentages for industries from the disaggregation input files.}
\usage{
getDisaggIndustryPercentages(disagg)
}
\arguments{
\item{disagg}{Specifications for disaggregating the current Model}
}
\value{
A dataframe with the default disaggregation percentages for the Industries of the current model
}
\description{
Obtain default disaggregation percentages for industries from the disaggregation input files.
}
