% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{getDisaggregationSpecs}
\alias{getDisaggregationSpecs}
\title{Obtain aggregation and disaggregation specs from input files}
\usage{
getDisaggregationSpecs(model)
}
\arguments{
\item{model}{Model file loaded with IO tables}
}
\value{
A model with the specified aggregation and disaggregation specs.
}
\description{
Obtain aggregation and disaggregation specs from input files
}
