% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Margins.R
\name{getFinalConsumerMarginsTable}
\alias{getFinalConsumerMarginsTable}
\title{Generate Margins table using either Industry Margins (BEA Margins) or Final Consumer Margins (BEA PCE and PEQ Bridge data).}
\usage{
getFinalConsumerMarginsTable(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A dataframe containing CommodityCode, and margins for ProducersValue, Transportation, Wholesale, Retail and PurchasersValue.
}
\description{
Generate Margins table using either Industry Margins (BEA Margins) or Final Consumer Margins (BEA PCE and PEQ Bridge data).
}
