% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowsaFunctions.R
\name{getFlowbySector}
\alias{getFlowbySector}
\title{Load flowsa's FlowBySector df}
\usage{
getFlowbySector(sat_spec)
}
\arguments{
\item{sat_spec, }{a specification file}
}
\value{
A data frame for flowsa data in sector by region totals format
}
\description{
Load flowsa's FlowBySector df
}
