% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowsaFunctions.R
\name{getFlowbySectorCollapsed}
\alias{getFlowbySectorCollapsed}
\title{Calls the Python flowsa package's getFlowBySector method}
\usage{
getFlowbySectorCollapsed(method_name)
}
\arguments{
\item{method_name}{The name}
}
\value{
A data frame for flowsa data in sector by region totals format
}
\description{
Calls the Python flowsa package's getFlowBySector method
}
