% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hybridization.R
\name{getHybridizationFiles}
\alias{getHybridizationFiles}
\title{Setup the hybridization specs based on the input files}
\usage{
getHybridizationFiles(model, configpaths = NULL)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{configpaths}{str vector, paths (including file name) of disagg configuration file(s).
If NULL, built-in config files are used.}
}
\value{
A model object with the correct hybridization specs.
}
\description{
Setup the hybridization specs based on the input files
}
