% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hybridization.R
\name{getHybridizationSpecs}
\alias{getHybridizationSpecs}
\title{Obtain hybridization specs from input files}
\usage{
getHybridizationSpecs(model, configpaths = NULL)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{configpaths}{str vector, paths (including file name) of configuration file(s).
If NULL, built-in config files are used.}
}
\value{
A model with the specified hybridization specs.
}
\description{
Obtain hybridization specs from input files
}
