% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{getIndex}
\alias{getIndex}
\title{Return the index where a sector occurrs in a sectorList}
\usage{
getIndex(sectorList, sector)
}
\arguments{
\item{sectorList}{Dataframe (of strings) to match the index of the sector param}

\item{sector}{String of the sector to look the index for}
}
\value{
Index of sector in sectorList
}
\description{
Return the index where a sector occurrs in a sectorList
}
