% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_IOfunctions.R
\name{getIndustryUseofImportedCommodities}
\alias{getIndustryUseofImportedCommodities}
\title{Determine proportion of Imports in Industr use.}
\usage{
getIndustryUseofImportedCommodities()
}
\value{
A dataframe contains "imports_used_by_industry", "total_used_by_industry", and "import_ratio_of_industry_use".
}
\description{
Determine proportion of Imports in Industr use.
}
