% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCIAfmtFunctions.R
\name{getInventoryMethod}
\alias{getInventoryMethod}
\title{Calls the Python flowsa package's getFlowBySector method}
\usage{
getInventoryMethod(subset)
}
\arguments{
\item{method_name}{The name}
}
\value{
A dataframe for flowsa data in sector by region totals format
}
\description{
Calls the Python flowsa package's getFlowBySector method
}
