% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{getMarginsTable}
\alias{getMarginsTable}
\title{Adjust CPI from GDP industries to IO industries (2012 schema) at Detail, Summary, and Sector IO levels.}
\usage{
getMarginsTable(model, marginsource)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{marginsource}{A character indicating the source of Margins, either "Industry" or "FinalConsumer".}
}
\value{
A list contains IO-based CPI at Detail, Summary, and Sector IO levels.
Generate Margins table using either Industry Margins (BEA Margins) or Final Consumer Margins (BEA PCE and PEQ Bridge data).

A dataframe containing CommodityCode, and margins for ProducersValue, Transportation, Wholesale, Retail and PurchasersValue.
}
\description{
Adjust CPI from GDP industries to IO industries (2012 schema) at Detail, Summary, and Sector IO levels.
}
