% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigurationFunctions.R
\name{getModelConfiguration}
\alias{getModelConfiguration}
\title{Gets a model configuration}
\usage{
getModelConfiguration(modelname)
}
\arguments{
\item{modelname}{The name of the model.}
}
\value{
A list of model specifications.
}
\description{
Gets a model configuration
}
