% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeModel.R
\name{getModelCrosswalk}
\alias{getModelCrosswalk}
\title{Get model crosswalk based on BaseIOSchema and BaseIOLevel}
\usage{
getModelCrosswalk(model)
}
\arguments{
\item{model}{EEIO model with specs assigned}
}
\value{
crosswalk as dataframe
}
\description{
Get model crosswalk based on BaseIOSchema and BaseIOLevel
}
