% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUIOFunctions.R
\name{getMUIOSectors}
\alias{getMUIOSectors}
\title{Obtain specs that indicate which sectors need to}
\usage{
getMUIOSectors(model, configpaths = NULL)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{configpaths}{str vector, paths (including file name) of configuration file(s).
If NULL, built-in config files are used.}
}
\value{
A model with the specified MUIO specs.
}
\description{
Obtain specs that indicate which sectors need to
}
