% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrosswalkFunctions.R
\name{getNAICS2012to2017Concordances}
\alias{getNAICS2012to2017Concordances}
\title{Get 2012 NAICS to 2017 NAICS concordances at 6-digit level.}
\usage{
getNAICS2012to2017Concordances()
}
\value{
data frame with columns '2012 NAICS Code', '2012 NAICS Title',
'2017 NAICS Code', and '2017 NAICS Title'.
}
\description{
Get 2012 NAICS to 2017 NAICS concordances at 6-digit level.
}
