% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrosswalkFunctions.R
\name{getNAICS2to6DigitsCodeName}
\alias{getNAICS2to6DigitsCodeName}
\title{Get 2-6 digit NAICS codes and names for year specified.}
\usage{
getNAICS2to6DigitsCodeName(year)
}
\arguments{
\item{year}{int. 2012 or 2007 accepted.}
}
\value{
dataframe with columns NAICS_year_Code and NAICS_year_Name.
}
\description{
Get 2-6 digit NAICS codes and names for year specified.
}
