% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrosswalkFunctions.R
\name{getNAICS7to10Digits}
\alias{getNAICS7to10Digits}
\title{Get 7-10 digit NAICS codes in a crosswalk format for year specified.}
\usage{
getNAICS7to10Digits(year)
}
\arguments{
\item{year}{int. 2012 or 2007 accepted.}
}
\value{
data frame with columns NAICS_7, NAICS_8, NAICS_9, NAICS_10.
}
\description{
Get 7-10 digit NAICS codes in a crosswalk format for year specified.
}
