% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrosswalkFunctions.R
\name{getNAICS7to10DigitsCodeName}
\alias{getNAICS7to10DigitsCodeName}
\title{Get 7-10 digit NAICS codes and names (agricultural, manufacturing, and mining industries) for year specified.}
\usage{
getNAICS7to10DigitsCodeName(year)
}
\arguments{
\item{year}{int. 2012 or 2007 accepted.}
}
\value{
data frame with columns NAICS_year_Code and NAICS_year_Name.
}
\description{
Get 7-10 digit NAICS codes and names (agricultural, manufacturing, and mining industries) for year specified.
}
