% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICS.R
\name{getNAICSCodeName}
\alias{getNAICSCodeName}
\title{Get NAICS codes and names for year specified.}
\usage{
getNAICSCodeName(year)
}
\arguments{
\item{year}{int. 2012 or 2007 accepted.}
}
\value{
dataframe with columns NAICS_year_Code and NAICS_year_Name.
}
\description{
Get NAICS codes and names for year specified.
}
