% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICStoBEAAllocation.R
\name{getNAICStoBEAAllocation}
\alias{getNAICStoBEAAllocation}
\title{Determine allocation factors between NAICS and BEA sectors based on IO output.}
\usage{
getNAICStoBEAAllocation(year)
}
\arguments{
\item{year}{Year of model Industry output.}
}
\value{
A table of allocation factors between NAICS and BEA sectors.
}
\description{
Determine allocation factors between NAICS and BEA sectors based on IO output.
}
