% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{getPRObyPURRatios}
\alias{getPRObyPURRatios}
\title{Calculate PRO/PUR price ratios using BEA PCE and PEQ Bridge data.}
\usage{
getPRObyPURRatios(specs)
}
\arguments{
\item{specs}{Model specifications.}
}
\value{
A dataframe contains PRO/PUR ratios for Detail BEA sectors.
}
\description{
Calculate PRO/PUR price ratios using BEA PCE and PEQ Bridge data.
}
