% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioeconomyFunctions.R
\name{getRowIndex}
\alias{getRowIndex}
\title{Obtains the row index of a given industry or commodity in the Make or Use table}
\usage{
getRowIndex(table, code)
}
\arguments{
\item{table}{the Make or the Use table}

\item{code}{the identifier BEA code (string)}
}
\value{
the row index of the commodity/industry indicated through the code.
}
\description{
Obtains the row index of a given industry or commodity in the Make or Use table
}
