% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R, R/CalculationFunctions.R
\name{getScalingVector}
\alias{getScalingVector}
\title{Multiply the Leontief inverse L and the demand vector.}
\usage{
getScalingVector(L, demand)

getScalingVector(L, demand)
}
\arguments{
\item{L}{Leontief inverse.}

\item{demand}{Final demand vector.}

\item{L}{Leontief inverse.}

\item{demand}{Final demand vector.}
}
\value{
Scaling vector.

Scaling vector.
}
\description{
Multiply the Leontief inverse L and the demand vector.

Multiply the Leontief inverse L and the demand vector.
}
\references{
Yang, Yi, Wesley W. Ingwersen, Troy R. Hawkins, Michael Srocka, and David E. Meyer.
2017. “USEEIO: A New and Transparent United States Environmentally-Extended Input-Output Model.”
Journal of Cleaner Production 158 (August): 308–18. https://doi.org/10.1016/j.jclepro.2017.04.150.
SI1, Equation 8.

Yang, Yi, Wesley W. Ingwersen, Troy R. Hawkins, Michael Srocka, and David E. Meyer.
2017. “USEEIO: A New and Transparent United States Environmentally-Extended Input-Output Model.”
Journal of Cleaner Production 158 (August): 308–18. https://doi.org/10.1016/j.jclepro.2017.04.150.
SI1, Equation 8.
}
