% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{getSchemaCode}
\alias{getSchemaCode}
\title{Return the schema subscript for accessing useeior objects}
\usage{
getSchemaCode(specs)
}
\arguments{
\item{specs}{list of model specs must include BaseIOSchema}
}
\value{
schema, str in form "YYsch" or NULL for 2012
}
\description{
Return the schema subscript for accessing useeior objects
}
