% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{getStandardSatelliteTableFormat}
\alias{getStandardSatelliteTableFormat}
\title{Load the template of standard satellite table.}
\usage{
getStandardSatelliteTableFormat()
}
\value{
A dataframe with the columns of the standard sat table format from the IO model builder.
}
\description{
Load the template of standard satellite table.
}
