% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateOutput.R
\name{getStateAgOutput}
\alias{getStateAgOutput}
\title{Get state agriculture Commodity output for 2012 and 2017}
\usage{
getStateAgOutput()
}
\description{
Get state agriculture Commodity output for 2012 and 2017
}
