% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateOutput.R
\name{getStateCommodityOutputEstimates}
\alias{getStateCommodityOutputEstimates}
\title{Estimate state commodity output}
\usage{
getStateCommodityOutputEstimates(location_acronym)
}
\arguments{
\item{location_acronym}{Abbreviated location name of the model, e.g. "US" or "GA".}
}
\value{
A dataframe contains state commodity output for specified state with row names being BEA sector code.
}
\description{
Estimate state commodity output
}
