% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateOutput.R
\name{getStateGDP}
\alias{getStateGDP}
\title{Get industry-level GDP for all states at a specific year.}
\usage{
getStateGDP(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2018 specifying the year of interest.}
}
\value{
A dataframe contains state GDP for all states at a specific year.
}
\description{
Get industry-level GDP for all states at a specific year.
}
