% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateOutput.R
\name{getStateIndustryOutput}
\alias{getStateIndustryOutput}
\title{Estimate Summary-level state industry output based on BEA and alternative sources.}
\usage{
getStateIndustryOutput(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2018 specifying the year of interest.}
}
\value{
A dataframe contains Summary-level state industry output for all states and a specific industry.
}
\description{
Estimate Summary-level state industry output based on BEA and alternative sources.
}
