% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandFunctions.R
\name{getUSTotalConsProd}
\alias{getUSTotalConsProd}
\title{Generate a dataframe containing US total consumption and production.}
\usage{
getUSTotalConsProd(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A dataframe of US total consumption and production.
}
\description{
Generate a dataframe containing US total consumption and production.
}
