% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{getValueAddedTotalsbySector}
\alias{getValueAddedTotalsbySector}
\title{Get value added from BEA input-output use table, convert to standard totals_by_sector format.}
\usage{
getValueAddedTotalsbySector(model)
}
\arguments{
\item{model}{A EEIO model with model specs and IO tables loaded}
}
\value{
A value-added totals_by_sector table with fields of standard totals_by_sector
}
\description{
Get value added from BEA input-output use table, convert to standard totals_by_sector format.
}
