% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{getVectorOfCodes}
\alias{getVectorOfCodes}
\title{Extract desired columns from SchemaInfo, return vectors with strings of codes.}
\usage{
getVectorOfCodes(ioschema, iolevel, colName)
}
\arguments{
\item{iolevel}{Level of detail, can be "Sector", "Summary, "Detail".}

\item{colName}{A text value specifying desired column name.}
}
\value{
A vector of codes.
}
\description{
Extract desired columns from SchemaInfo, return vectors with strings of codes.
}
