% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIOFunctions.R
\name{getWIOSpecs}
\alias{getWIOSpecs}
\title{Obtain WIO specs from input files}
\usage{
getWIOSpecs(model, configpaths = NULL)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{configpaths}{str vector, paths (including file name) of configuration file(s).
If NULL, built-in config files are used.}
}
\value{
A model with the specified WIO specs.
}
\description{
Obtain WIO specs from input files
}
