% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareModels.R
\name{groupandsumTbSbyFlowLoc}
\alias{groupandsumTbSbyFlowLoc}
\title{Sum totals_by_sector by flow and location}
\usage{
groupandsumTbSbyFlowLoc(TbS)
}
\arguments{
\item{TbS}{a totals-by-sector list from a model}
}
\value{
flowtotal, a df with 1 column of flowamount with a total for that flow given in rowname
}
\description{
Sum totals_by_sector by flow and location
}
