% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{harmonizeDFsbyrowname}
\alias{harmonizeDFsbyrowname}
\title{Align two dfs by row}
\usage{
harmonizeDFsbyrowname(df1, df2)
}
\arguments{
\item{df1, }{a dataframe with row names}

\item{df2, }{another dataframe with row names matching some of df1}
}
\value{
, a list with harmonized df1 in position 1 and harmonized df2 in position 2
}
\description{
Align two dfs by row
}
