% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationFunctions.R
\name{heatmapSatelliteTableCoverage}
\alias{heatmapSatelliteTableCoverage}
\title{Bar plot of indicator scores calculated from totals by sector and displayed by BEA Sector Level to compare scores across models}
\usage{
heatmapSatelliteTableCoverage(model)
}
\arguments{
\item{model}{A complete EEIO model}
}
\description{
Bar plot of indicator scores calculated from totals by sector and displayed by BEA Sector Level to compare scores across models
}
