% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationFunctions.R
\name{heatmapSectorRanking}
\alias{heatmapSectorRanking}
\title{SMM tool like heatmap showing ranking of sectors}
\usage{
heatmapSectorRanking(
  model,
  matrix,
  indicators,
  sector_to_remove,
  y_title,
  N_sector
)
}
\arguments{
\item{model}{A complete EEIO model}

\item{matrix}{A matrix from model result}

\item{sector_to_remove}{Code of one or more BEA sectors that will be removed from the plot. Can be "".}

\item{y_title}{The title of y axis, excluding unit.}

\item{N_sector}{A numeric value indicating number of sectors to show in the ranking}
}
\description{
SMM tool like heatmap showing ranking of sectors
}
