% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hybridization.R
\name{hybridizeAMatrix}
\alias{hybridizeAMatrix}
\title{Hybridize the technology matrix based on specified source file}
\usage{
hybridizeAMatrix(model, domestic = FALSE)
}
\arguments{
\item{model}{An EEIO model object with model specs, IO tables, and satellite tables loaded}

\item{domestic}{A logical value indicating whether to hybridize domestic matrix.}
}
\value{
The A matrix for a hybridized model.
}
\description{
Hybridize the technology matrix based on specified source file
}
