% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hybridization.R
\name{hybridizeBMatrix}
\alias{hybridizeBMatrix}
\title{Hybridize the environmental matrix based on specified source file}
\usage{
hybridizeBMatrix(model)
}
\arguments{
\item{model}{An EEIO model object with model specs, IO tables, and satellite tables loaded}
}
\value{
The B matrix for a hybridized model.
}
\description{
Hybridize the environmental matrix based on specified source file
}
