% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hybridization.R
\name{hybridizeModelObjects}
\alias{hybridizeModelObjects}
\title{Update remaining model objects following hybridization of matrices for consistency}
\usage{
hybridizeModelObjects(model)
}
\arguments{
\item{model}{A completed EEIO model object}
}
\value{
model with remaining objects expanded to include hybrid process data
}
\description{
Update remaining model objects following hybridization of matrices for consistency
}
