% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIOFunctions.R
\name{includeFullUseWIO}
\alias{includeFullUseWIO}
\title{Include the WIO elements of the Use table in the correct configuration}
\usage{
includeFullUseWIO(model, WIO)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{WIO}{A list with WIO specifications and data}
}
\value{
A model with the UseTransactions matrix modified with WIO specs.
}
\description{
Include the WIO elements of the Use table in the correct configuration
}
