% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIOFunctions.R
\name{includeMakeWIO}
\alias{includeMakeWIO}
\title{Include the WIO elements of the Make table in the correct configuration}
\usage{
includeMakeWIO(model, WIO)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{WIO}{A list with WIO specifications and data}
}
\value{
A model with the MakeTransactions matrix modified with WIO specs.
}
\description{
Include the WIO elements of the Make table in the correct configuration
}
