% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioeconomyFunctions.R
\name{initialCalculations}
\alias{initialCalculations}
\title{Do initial calculations to determine the amounts of GGE produced in future bio-economy by each of the 3 new industries.}
\usage{
initialCalculations(
  modelName,
  newIndustryInfo,
  newCommodityInfo,
  simCommodityInfo,
  percentage
)
}
\arguments{
\item{newIndustryInfo}{dataframe with 3 different technologies on the rows and the following variables (Code,Name, Primary_product, Price, Percentage_Prod)}

\item{newCommodityInfo}{dataframe with the information of the new commodity. Column variables (Code, Name, Primary_producers)}

\item{simCommodityInfo}{list with the information of the similar commodity. Elements (Code, Name, PrimaryProducer_Code, PrimaryProducer_Name, Price)}

\item{percentage}{numeric (0,1] that refers to the \% of the original commodity (fuels) that the new commodity (biofuels) will replace.}

\item{modelname}{Name of the model from a config file.}
}
\value{
The model with basic information added and initial calculations, and print the amount of biofuel in GGE produced by each technology/industry in the future bio-economy.
}
\description{
Do initial calculations to determine the amounts of GGE produced in future bio-economy by each of the 3 new industries.
}
