% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQualityFunctions.R
\name{initializeArray}
\alias{initializeArray}
\title{Helper function to create a 3d array for dqi matrices}
\usage{
initializeArray(dqi, num_matrices, name_matrices)
}
\arguments{
\item{dqi}{A complete EEIO model object with TbS}

\item{num_matrices, }{int of number of matrices (e.g. 5)}

\item{name_matrices, }{name of each matrix}
}
\value{
A 3d matrix, of correct dimensions
}
\description{
Helper function to create a 3d array for dqi matrices
}
