% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeModel.R
\name{initializeModel}
\alias{initializeModel}
\title{Initialize model with specifications and fundamental crosswalk table.}
\usage{
initializeModel(modelname)
}
\arguments{
\item{modelname}{Name of the model from a config file.}
}
\value{
A list of model specifications and fundamental crosswalk table.
}
\description{
Initialize model with specifications and fundamental crosswalk table.
}
