% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIOFunctions.R
\name{initializeWIOObjects}
\alias{initializeWIOObjects}
\title{Initialize all the WIO model objects simultaneously.}
\usage{
initializeWIOObjects(model)
}
\arguments{
\item{model}{An EEIO model object with model sepcs and IO tables loaded}
}
\value{
A model object with the WIO lists initialized but empty
}
\description{
Initialize all the WIO model objects simultaneously.
}
