% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandFunctions.R
\name{isDemandVectorValid}
\alias{isDemandVectorValid}
\title{A function to validate a user provided demand vector}
\usage{
isDemandVectorValid(dv, L)
}
\arguments{
\item{dv}{a user provided demand vector}

\item{L, }{the L matrix for the given model, used as a reference}
}
\value{
A logical value indicating demand vector is valid or not.
}
\description{
A function to validate a user provided demand vector
}
