% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationFunctions.R
\name{lineplotFlowCoefficients}
\alias{lineplotFlowCoefficients}
\title{Line plot of flow coefficients}
\usage{
lineplotFlowCoefficients(flow, matrix, modelspecs)
}
\arguments{
\item{flow}{An index name for a flow in a model matrix (full flow name)}

\item{matrix}{Supports a model B, M, or U matrix}

\item{modelspecs}{The specs for a given model}
}
\description{
Line plot of flow coefficients
}
