% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_FlowMappingFunctions.R
\name{listOriginalSources}
\alias{listOriginalSources}
\title{List original sources of environmental flows.}
\usage{
listOriginalSources()
}
\description{
List original sources of environmental flows.
}
