% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIndicators.R
\name{loadandbuildIndicators}
\alias{loadandbuildIndicators}
\title{Loads data for all model indicators as listed in model specs}
\usage{
loadandbuildIndicators(model)
}
\arguments{
\item{list}{a model object with IO data loaded}
}
\value{
list a model object with Indicators added
}
\description{
Loads data for all model indicators as listed in model specs
}
