% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSatellites.R
\name{loadandbuildSatelliteTables}
\alias{loadandbuildSatelliteTables}
\title{Loads data for all satellite tables as lists in model specs}
\usage{
loadandbuildSatelliteTables(model)
}
\arguments{
\item{model}{A model list object with model specs and IO tables listed}
}
\value{
A model object with Satellite tables added
}
\description{
Loads data for all satellite tables as lists in model specs
}
