% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIOTables.R
\name{loadBEAMakeorUseTable}
\alias{loadBEAMakeorUseTable}
\title{Externalize the BEA make and use tables}
\usage{
loadBEAMakeorUseTable(iolevel, makeoruse, year, redef)
}
\arguments{
\item{iolevel}{"Detail", "Sector", or "Summary" level}

\item{makeoruse}{Select either "Make" or "Use"}

\item{year}{If iolevel = "Detail", then has to be 2012, otherwise select between 2010--2018}

\item{redef}{Select either "BeforeRedef" or "AfterRedef"}
}
\value{
A dataframe of "PRO" BEA make or use tables
}
\description{
Externalize the BEA make and use tables
}
