% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIOTables.R
\name{loadBEAtables}
\alias{loadBEAtables}
\title{Load BEA IO tables in a list based on model config.}
\usage{
loadBEAtables(specs)
}
\arguments{
\item{specs}{Model specifications.}
}
\value{
A list with BEA IO tables.
}
\description{
Load BEA IO tables in a list based on model config.
}
