% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSatellites.R
\name{loadbuildSatelliteTables}
\alias{loadbuildSatelliteTables}
\title{Loads data for all satellite tables as lists in model specs}
\usage{
loadbuildSatelliteTables(model)
}
\arguments{
\item{list}{a model object with IO data loaded}
}
\value{
list a model object with Satellite tables added
}
\description{
Loads data for all satellite tables as lists in model specs
}
