% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadGOandCPI.R
\name{loadChainPriceIndexTable}
\alias{loadChainPriceIndexTable}
\title{Load Chain Price Index table based on model specifications.}
\usage{
loadChainPriceIndexTable(specs)
}
\arguments{
\item{specs}{Specifications of the model.}
}
\value{
A data.frame of Chain Price Index.
}
\description{
Load Chain Price Index table based on model specifications.
}
