% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIOTables.R
\name{loadCommodityandIndustryOutput}
\alias{loadCommodityandIndustryOutput}
\title{Prepare commodity and industry output of an EEIO form USEEIO model.}
\usage{
loadCommodityandIndustryOutput(model)
}
\arguments{
\item{model}{A model object with model specs and fundamental IO data loaded.}
}
\value{
A list with USEEIO model economic components.
}
\description{
Prepare commodity and industry output of an EEIO form USEEIO model.
}
