% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigurationFunctions.R
\name{loadConfig}
\alias{loadConfig}
\title{Load model configuration.}
\usage{
loadConfig(modelname)
}
\arguments{
\item{modelname}{The name of the model.}
}
\value{
A list of model specifications.
}
\description{
Load model configuration.
}
