% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{loadDataCommonsfile}
\alias{loadDataCommonsfile}
\title{Load the static file originating from Data Commons either by loading from local directory
or downloading from Data Commons and 
saving to local directory}
\usage{
loadDataCommonsfile(static_file)
}
\arguments{
\item{static_file}{The name of a static file, including the subdirectories}
}
\value{
The static file originating from Data Commons
}
\description{
Load the static file originating from Data Commons either by loading from local directory
or downloading from Data Commons and 
saving to local directory
}
