% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{loadDataFile}
\alias{loadDataFile}
\title{Load the file originating from a remote location}
\usage{
loadDataFile(static_file, location)
}
\arguments{
\item{static_file}{The name of a static file, including the subdirectories}

\item{location}{The name of the data store location, which is used to look up the URL in DataStorageConfig.yml}
}
\value{
The static file loaded from the location or local cache
}
\description{
Load the file originating from a remote location
}
