% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDemandVectors.R
\name{loadDefaultDemandVectorMeta}
\alias{loadDefaultDemandVectorMeta}
\title{Loads a package stored csv of demand vector meta for vectors to be created for every model with type and system specified
This function adds additional year, location and IDs along with the type and system based on the model given}
\usage{
loadDefaultDemandVectorMeta(model)
}
\arguments{
\item{model}{An EEIO model that has been initialized}
}
\value{
a data frame of metadata with cols Type, System, Name, Year, Location and ID with rows for each default
}
\description{
Loads a package stored csv of demand vector meta for vectors to be created for every model with type and system specified
This function adds additional year, location and IDs along with the type and system based on the model given
}
