% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDemandVectors.R
\name{loadDemandVectors}
\alias{loadDemandVectors}
\title{Adds demand vectors based on model specs to model object}
\format{
A list with lists of totals by sector and formatted satellite tables
}
\usage{
loadDemandVectors(model)
}
\arguments{
\item{model}{A model list object with the specs object listed}
}
\value{
eeio model with a list of demand vectors added
}
\description{
Adds demand vectors based on model specs to model object
}
