% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadExternalImportFactors.R
\name{loadExternalImportFactors}
\alias{loadExternalImportFactors}
\title{Load and prepare import coefficients}
\usage{
loadExternalImportFactors(model)
}
\arguments{
\item{model}{An EEIO form USEEIO model object with model specs loaded}
}
\value{
M_m, matrix of import coefficients (flow x sector).
}
\description{
Load and prepare import coefficients
}
