% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIndicators.R
\name{loadFactors}
\alias{loadFactors}
\title{Load indicator factors based on spec from static or dynamic source}
\usage{
loadFactors(ind_spec)
}
\arguments{
\item{ind_spec}{Specification of an indicator}
}
\value{
A dataframe of factors with factor_fields
}
\description{
Load indicator factors based on spec from static or dynamic source
}
