% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadGDPData.R
\name{loadGDPtables}
\alias{loadGDPtables}
\title{Load GDP tables in a list based on model config.}
\usage{
loadGDPtables(specs)
}
\arguments{
\item{specs}{Specifications of the model.}
}
\value{
A list with GDP tables.
}
\description{
Load GDP tables in a list based on model config.
}
