% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIOTables.R
\name{loadImportMatrix}
\alias{loadImportMatrix}
\title{Load, format, and save import matrix as a USEEIO model object.}
\usage{
loadImportMatrix(model, io_codes)
}
\arguments{
\item{model}{A model object with model specs loaded.}

\item{io_codes}{A list of BEA IO codes.}
}
\value{
Import, df of use table imports.
}
\description{
Load, format, and save import matrix as a USEEIO model object.
}
