% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIOTables.R
\name{loadIOcodes}
\alias{loadIOcodes}
\title{Load BEA IO codes in a list based on model config.}
\usage{
loadIOcodes(specs)
}
\arguments{
\item{specs}{Model specifications.}
}
\value{
A list with BEA IO codes.
}
\description{
Load BEA IO codes in a list based on model config.
}
