% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIOTables.R
\name{loadIOData}
\alias{loadIOData}
\title{Prepare economic components of an EEIO model.}
\usage{
loadIOData(model, configpaths = NULL)
}
\arguments{
<<<<<<< hybridization
\item{model}{An EEIO model object with model specs loaded}
=======
\item{model}{An EEIO form USEEIO model object with model specs loaded}
>>>>>>> develop

\item{configpaths}{str vector, paths (including file name) of model configuration file
and optional agg/disagg configuration file(s). If NULL, built-in config files are used.}
}
\value{
<<<<<<< hybridization
A list with EEIO model economic components.
=======
A list with EEIO form USEEIO model economic components.
>>>>>>> develop
}
\description{
Prepare economic components of an EEIO model.
}
