% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIOTables.R
\name{loadIOmeta}
\alias{loadIOmeta}
\title{Prepare metadata of economic components of an EEIO form USEEIO model.}
\usage{
loadIOmeta(model)
}
\arguments{
\item{model}{A model object with model specs loaded.}
}
\value{
A list with USEEIO model economic components' metadata.
}
\description{
Prepare metadata of economic components of an EEIO form USEEIO model.
}
