% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrosswalkFunctions.R
\name{loadMasterCrosswalk}
\alias{loadMasterCrosswalk}
\title{Function to externalize the BEA to NAICS crosswalk}
\usage{
loadMasterCrosswalk()
}
\value{
A crosswalk linking 2007 and 2012 NAICS codes to 2012 Sector, Summary, and Detail BEA codes
}
\description{
Function to externalize the BEA to NAICS crosswalk
}
