% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadGOandCPI.R
\name{loadNationalGrossOutputTable}
\alias{loadNationalGrossOutputTable}
\title{Load US Gross Output table based on model specifications.}
\usage{
loadNationalGrossOutputTable(specs)
}
\arguments{
\item{specs}{Specifications of the model.}
}
\value{
A data.frame of US Gross Output.
}
\description{
Load US Gross Output table based on model specifications.
}
