% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIOTables.R
\name{loadNationalIOData}
\alias{loadNationalIOData}
\title{Prepare economic components of an EEIO form USEEIO model.}
\usage{
loadNationalIOData(model)
}
\arguments{
\item{model}{A model object with model specs loaded.}
}
\value{
A list with USEEIO model economic components.
}
\description{
Prepare economic components of an EEIO form USEEIO model.
}
