% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIOTables.R
\name{loadTwoRegionStateIOtables}
\alias{loadTwoRegionStateIOtables}
\title{Load two-region state IO tables in a list based on model config.}
\usage{
loadTwoRegionStateIOtables(specs)
}
\arguments{
\item{specs}{Model specifications.}
}
\value{
A list with state IO tables.
}
\description{
Load two-region state IO tables in a list based on model config.
}
