% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQuality.R
\name{lookupDQBoundScore}
\alias{lookupDQBoundScore}
\title{Constants to determine which bound type to use for DQ indicator}
\usage{
lookupDQBoundScore(raw_score, dqi, scoring_bounds)
}
\description{
Constants to determine which bound type to use for DQ indicator
}
