% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{mapFIPS5toLocationNames}
\alias{mapFIPS5toLocationNames}
\title{Maps a vector of 5-digit FIPS codes to location names}
\usage{
mapFIPS5toLocationNames(fipscodes, fipssystem)
}
\arguments{
\item{fipscodes}{A vector of 5 digit FIPS codes}

\item{fipssystem}{A text value specifying FIPS System, can be FIPS_2015}
}
\value{
A vector of location names where matches are found
}
\description{
Maps a vector of 5-digit FIPS codes to location names
}
