% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowMappingFunctions.R
\name{mapFlowbyCodeandCategory}
\alias{mapFlowbyCodeandCategory}
\title{Map resource and emission names in a standardized satellite table to new names by code and category.
Must set original source before getting this information.}
\usage{
mapFlowbyCodeandCategory(originalcas, originalcategory = "",
  sattablemeta)
}
\arguments{
\item{originalcategory}{Original category of flows, an optional parameter.}

\item{sattablemeta}{Meta data of the satellite table.}

\item{originalname}{Original name of flows.}
}
\description{
Map resource and emission names in a standardized satellite table to new names by code and category.
Must set original source before getting this information.
}
