% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowMappingFunctions.R
\name{mapFlowbyNameandCategory}
\alias{mapFlowbyNameandCategory}
\title{Map resource and emission names in a standardized satellite table to new names by name and category.
Must set original source before getting this information.}
\usage{
mapFlowbyNameandCategory(originalname, originalcategory = "", sattablemeta)
}
\arguments{
\item{originalname}{Original name of flows.}

\item{originalcategory}{Original category of flows, an optional parameter.}

\item{sattablemeta}{Meta data of the satellite table.}
}
\value{
A full flow information as a vector of 6 components using original name only.
}
\description{
Map resource and emission names in a standardized satellite table to new names by name and category.
Must set original source before getting this information.
}
