% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowsaFunctions.R
\name{mapFlowBySectorCollapsedtoRegionalTotals}
\alias{mapFlowBySectorCollapsedtoRegionalTotals}
\title{Maps flows by sector collapsed format to totals by sector and region format}
\usage{
mapFlowBySectorCollapsedtoRegionalTotals(fbsct)
}
\arguments{
\item{fbsct}{A df in the flowbysector collapsed format with additional columns removed}
}
\value{
A df of sector by region totals
}
\description{
Maps flows by sector collapsed format to totals by sector and region format
}
