% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{mapFlowTotalsbySectorandLocationfromNAICStoBEA}
\alias{mapFlowTotalsbySectorandLocationfromNAICStoBEA}
\title{Map a satellite table from NAICS-coded format to BEA-coded format.}
\usage{
mapFlowTotalsbySectorandLocationfromNAICStoBEA(
  totals_by_sector,
  totals_by_sector_year,
  model
)
}
\arguments{
\item{totals_by_sector}{A standardized satellite table with resource and emission names from original sources.}

\item{totals_by_sector_year}{Year of the satellite table.}

\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A satellite table aggregated by the USEEIO model sector codes.
}
\description{
Map a satellite table from NAICS-coded format to BEA-coded format.
}
