% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{mapFlowTotalsbySectorfromBEASchema2007to2012}
\alias{mapFlowTotalsbySectorfromBEASchema2007to2012}
\title{Map a satellite table from BEA Detail industry 2007 schema to 2012 schema.}
\usage{
mapFlowTotalsbySectorfromBEASchema2007to2012(totals_by_sector)
}
\arguments{
\item{totals_by_sector}{A standardized satellite table with resource and emission names from original sources.}
}
\value{
A satellite table aggregated by the USEEIO model sector codes.
}
\description{
Map a satellite table from BEA Detail industry 2007 schema to 2012 schema.
}
