% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCIAfmtFunctions.R
\name{mapLCIAmethodtoIndicators}
\alias{mapLCIAmethodtoIndicators}
\title{Temp function to map the trimmed LCIAmethod format from the LCIAformatter to 
the format for use in useeior. Very closely follows the mapFlowBySectorCollapsedtoRegionalTotals
in flowsa functions}
\usage{
mapLCIAmethodtoIndicators(lciamethodt)
}
\arguments{
\item{df}{with the trimmed LCIAmethod output of prepareLCIAmethodforIndicators}
}
\description{
Temp function to map the trimmed LCIAmethod format from the LCIAformatter to 
the format for use in useeior. Very closely follows the mapFlowBySectorCollapsedtoRegionalTotals
in flowsa functions
}
