% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowMappingFunctions.R
\name{mapListbyName}
\alias{mapListbyName}
\title{Map resource and emission names in a standardized satellite table to new names by flow name}
\usage{
mapListbyName(sattable, sattablemeta)
}
\arguments{
\item{sattable}{A standardized satellite table with resource and emission names from original sources.}

\item{sattablemeta}{Meta data of the satellite table.}
}
\value{
A standardized satellite table with new resource and emission names.
}
\description{
Map resource and emission names in a standardized satellite table to new names by flow name
}
