% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{mapLocationCodestoNames}
\alias{mapLocationCodestoNames}
\title{Maps location codes to names}
\usage{
mapLocationCodestoNames(codes, codesystem)
}
\arguments{
\item{codes}{A vector of location codes}

\item{codesystem}{A text value specifying code system, e.g. FIPS.}
}
\value{
A vector of location names where matches are found.
}
\description{
Maps location codes to names
}
