% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{mapSatTablefromNAICStoBEA}
\alias{mapSatTablefromNAICStoBEA}
\title{Map a satellite table from NAICS-coded format to BEA-coded format.}
\usage{
mapSatTablefromNAICStoBEA(sattable, satellitetableyear)
}
\arguments{
\item{sattable}{A standardized satellite table with resource and emission names from original sources.}

\item{satellitetableyear}{Year of the satellite table.}
}
\value{
A satellite table aggregated by the USEEIO model sector codes.
}
\description{
Map a satellite table from NAICS-coded format to BEA-coded format.
}
