% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{mapSatTablefromNAICStoUSEEIO}
\alias{mapSatTablefromNAICStoUSEEIO}
\title{Map a satellite table from NAICS-coded format to USEEIO-coded format.}
\usage{
mapSatTablefromNAICStoUSEEIO(sattable, majormodelversionnumber,
  satellitetableyear)
}
\arguments{
\item{sattable}{A standardized satellite table with resource and emission names from original sources.}

\item{majormodelversionnumber}{Model's major version number.}

\item{satellitetableyear}{Year of the satellite table.}
}
\value{
A satellite table aggregated by the USEEIO model sector codes.
}
\description{
Map a satellite table from NAICS-coded format to USEEIO-coded format.
}
