% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateOutput.R
\name{mapStateTabletoBEASummary}
\alias{mapStateTabletoBEASummary}
\title{Map state table (GDP, Tax, Employment Compensation, and GOS) to BEA Summary, apply allocation based on BLS QCEW Employment}
\usage{
mapStateTabletoBEASummary(statetable)
}
\arguments{
\item{statetable}{Pre-saved state table (GDP, Tax, Employment Compensation, and GOS).}
}
\value{
A dataframe contains state industry output for specified state with row names being BEA sector code.
}
\description{
Map state table (GDP, Tax, Employment Compensation, and GOS) to BEA Summary, apply allocation based on BLS QCEW Employment
}
