% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{MasterCrosswalk}
\alias{MasterCrosswalk}
\title{Master Crosswalk table (2017 schema)}
\format{
A dataframe with 4095 obs. and 12 variables:
\describe{
 \item{BEA_2017_Sector_Code}{text code}
 \item{BEA_2017_Sector_Name}{text code}
 \item{BEA_2017_Summary_Code}{text code}
 \item{BEA_2017_Summary_Name}{text code}
 \item{BEA_2017_Detail_Code}{text code}
 \item{BEA_2017_Detail_Name}{text code}
 \item{USEEIO_Code}{text code}
 \item{USEEIO_Name}{text code}
 \item{NAICS_2017_Code}{text code}
 \item{NAICS_2017_Name}{text code}
 \item{NAICS_2012_Code}{text code}
 \item{NAICS_2007_Code}{text code}
}
}
\usage{
MasterCrosswalk
}
\description{
Master Crosswalk table (2017 schema)
}
\keyword{datasets}
