% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{MasterCrosswalk2007}
\alias{MasterCrosswalk2007}
\title{Master Crosswalk table (2007 schema)}
\format{
A dataframe with 4141 obs. and 5 variables:
\describe{
 \item{BEA_2012_Sector_Code}{text code}
 \item{BEA_2012_Summary_Code}{text code}
 \item{BEA_2012_Detail_Code}{text code}
 \item{NAICS_2012_Code}{text code}
 \item{NAICS_2007_Code}{text code}
}
}
\usage{
MasterCrosswalk2007
}
\description{
Master Crosswalk table (2007 schema)
}
\keyword{datasets}
