% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{MasterCrosswalk2012}
\alias{MasterCrosswalk2012}
\title{Master Crosswalk table (2012 schema)}
\format{
A dataframe with 16776 obs. and 6 variables:
\describe{
 \item{BEA_2012_Sector_Code}{text code}
 \item{BEA_2012_Summary_Code}{text code}
 \item{BEA_2012_Detail_Code}{text code}
 \item{NAICS_2012_Code}{text code}
 \item{NAICS_2007_Code}{text code}
 \item{NAICS_2017_Code}{text code}
}
}
\usage{
MasterCrosswalk2012
}
\description{
Master Crosswalk table (2012 schema)
}
\keyword{datasets}
