% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{MasterCrosswalk2017}
\alias{MasterCrosswalk2017}
\title{Master Crosswalk table (2017 schema)}
\format{
A dataframe with 4095 obs. and 5 variables:
\describe{
 \item{BEA_2017_Sector_Code}{text code}
 \item{BEA_2017_Summary_Code}{text code}
 \item{BEA_2017_Detail_Code}{text code}
 \item{NAICS_2017_Code}{text code}
 \item{NAICS_2012_Code}{text code}
}
}
\usage{
MasterCrosswalk2017
}
\description{
Master Crosswalk table (2017 schema)
}
\keyword{datasets}
