% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioeconomyFunctions.R
\name{modifyMakeandUseTables}
\alias{modifyMakeandUseTables}
\title{Modify Make and Use Tables.}
\usage{
modifyMakeandUseTables(modModel, inputPurchases, valueAdded)
}
\arguments{
\item{modModel}{the model under modification after initial adjustments for Use Table.}

\item{inputPurchases}{(nComm+1)x 3 matrix with the amount spent in each of the nComm existing commodities to produce 1 GGE of biofuel in each of the new industries.}

\item{valueAdded}{2x3 matrix with the first 2 components of value added (Compensation of employees and Taxes on production and imports, less subsidies) required 
to produce 1 GGE of biofuel in each of the new industries.}
}
\value{
the model with Make and Use tables updated.

Example: modifyMakeandUseTable(modModel,matrix(c(rep(0.1,300),rep(0,106),rep(0.1,100),rep(0,306),rep(0,300),rep(0.102,106)), nrow=406, ncol=3),matrix(c(rep(1,2),rep(2,2),rep(1.5,2)), nrow=2, ncol=3))
}
\description{
This function modifies the pre-saved Make and Use tables for adding the new biofuels sector.It first modify the industry (intermediate) uses and
the final users uses, including those for the new sectors. Then it obtains the Total Industry Output for the new industries. Then it calls
modifyMakeTable() to modify the Make table and then ends modifying the Use table, adjusting the Value Added components.
}
