% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioeconomyFunctions.R
\name{modifyMakeTable}
\alias{modifyMakeTable}
\title{Modify Make BEA Table.}
\usage{
modifyMakeTable(modModel, UseTransactions, TotalIndustryOutputNewInd)
}
\arguments{
\item{modModel}{the model under modification after initial adjustments for Use Table.}

\item{UseTransactions}{dataframe with the input purchases/use transactions for each industry of each commodity after modifying Use Table.}

\item{TotalIndustryOutputNewInd}{a dataframe with the Total Industry Output for the 3 new industries based on the modified Use Table.}
}
\value{
a dataframe with the original Make table modified.

Example: modifyMakeTable(modModel,modUse[1:406,1:408],modUse[412,406:408])
}
\description{
This function modifies the pre-saved Make table for adding the new biofuels sector.
}
