% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioeconomyFunctions.R
\name{modifyUseTable}
\alias{modifyUseTable}
\title{Modify Use Table.}
\usage{
modifyUseTable(
  newSectorCode,
  similarSectorCode,
  percentage,
  inputPurchases,
  originalUse,
  newTotalIndustryOutput
)
}
\arguments{
\item{newSectorCode}{string/character that refers to the code/identifier that will be used for the new sector in the matrices}

\item{similarSectorCode}{string/character that refers to the code/identifier for the similar existing sector.}

\item{percentage}{numeric that refers to the \% of output of the original sector that new sector will produce.}

\item{inputPurchases}{(n+1)x 1 column vector with the amount spent in each of the n existing commodities to produce "Total Industry Output"
of the new sector.}

\item{originalUse}{dataframe with the Make table obtained in buildEEIOModel().}

\item{newTotalIndustryOutput}{the Total Industry Output of the new sector obtained from the updated make table.}
}
\value{
a dataframe with the originalUse table modified.

Example: modifyUseTable("prueba","324110",0.2,rep(1,times=406),model$Use)
}
\description{
This function modifies the pre-saved Use table for adding one new bioeconomy sector.
}
