% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{normalizeIOTransactions}
\alias{normalizeIOTransactions}
\title{Derive IO coefficients}
\usage{
normalizeIOTransactions(IO_transactions_df, IO_output_df)
}
\arguments{
\item{IO_transactions_df}{IO transactions of the model in dataframe format.}

\item{IO_output_df}{Output of the model in dataframe format.}
}
\value{
A matrix.
}
\description{
Derive IO coefficients
}
