% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{normalizeResultMatrixByTotalImpacts}
\alias{normalizeResultMatrixByTotalImpacts}
\title{Divide/Normalize a sector x flows matrix by the total of respective flow (column sum)}
\usage{
normalizeResultMatrixByTotalImpacts(m)
}
\arguments{
\item{m}{A sector x flows matrix.}
}
\value{
A normalized sector x flows matrix.
}
\description{
Divide/Normalize a sector x flows matrix by the total of respective flow (column sum)
}
