% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{PCEBridge2012}
\alias{PCEBridge2012}
\title{PCE Bridge Data for 2012}
\format{A data frame with 712 obs. and 9 variables:
\describe{
 \item{NIPACode}{text code}
 \item{PCECategory}{text category name, like 'Therapeutic medical equipment'}
 \item{CommodityCode}{BEA_2012_Detail_Code}
 \item{CommodityDescription}{BEA_2012_Detail_Name}
 \item{ProducersValue}{USD2012}
 \item{Transportation}{USD2012}
 \item{Wholesale}{USD2012}
 \item{Retail}{USD2012}
 \item{PurchasersValue}{USD2012}
}}
\source{
\url{https://apps.bea.gov/industry/xls/underlying-estimates/PCEBridge_2007_2012_DET.xlsx}
}
\usage{
PCEBridge2012
}
\description{
PCE Bridge Data for 2012
}
\keyword{datasets}
