% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationFunctions.R
\name{plotMatrixCoefficient}
\alias{plotMatrixCoefficient}
\title{Plot specified matrix coefficients as points to compare coefficient across models}
\usage{
plotMatrixCoefficient(
  model_list,
  matrix_name,
  coefficient_name,
  sector_to_remove,
  y_title,
  y_label,
  log_scale = FALSE
)
}
\arguments{
\item{model_list}{List of EEIO models with IOdata, satellite tables, and indicators loaded
Models must have the same coefficient in the rows}

\item{matrix_name}{Name of model matrix to extract data from, e.g. "B"}

\item{coefficient_name}{Row name in the specified matrix for the line plot}

\item{sector_to_remove}{Code of one or more BEA sectors that will be removed from the plot. Can be "".}

\item{y_title}{The title of y axis, excluding unit.}

\item{y_label}{The labels of y axis, can be "Name" or "Code".}

\item{log_scale}{A logical value indicating whether plotting in log scale.}
}
\description{
Plot specified matrix coefficients as points to compare coefficient across models
}
