% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{prepareDemandVectorForImportResults}
\alias{prepareDemandVectorForImportResults}
\title{Prepare demand vector for EEIO model results calculations}
\usage{
prepareDemandVectorForImportResults(
  model,
  demand = "Production",
  location = NULL
)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{demand}{A demand vector, can be name of a built-in model demand vector, e.g. "Production" or "Consumption",}

\item{location, }{str optional location code for demand vector, required for two-region models}
}
\description{
Prepare demand vector for EEIO model results calculations
}
