% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{prepareDemandVectorForStandardResults}
\alias{prepareDemandVectorForStandardResults}
\title{Prepare demand vector for EEIO model results calculations}
\usage{
prepareDemandVectorForStandardResults(
  model,
  demand = "Production",
  location = NULL,
  use_domestic_requirements = FALSE
)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{demand}{A demand vector, can be name of a built-in model demand vector, e.g. "Production" or "Consumption",
or an actual demand vector with names as one or more model sectors and
numeric values in USD with the same dollar year as model.}

\item{location, }{str optional location code for demand vector, required for two-region models}

\item{use_domestic_requirements}{A logical value: if TRUE, use domestic demand and L_d matrix;
if FALSE, use complete demand and L matrix.}
}
\description{
Prepare demand vector for EEIO model results calculations
}
