% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandFunctions.R
\name{prepareDomesticProductionDemand}
\alias{prepareDomesticProductionDemand}
\title{Prepares a demand vector representing domestic production
Formula for production vector: y_p <- y_dc + y_e + y_d_delta}
\usage{
prepareDomesticProductionDemand(model)
}
\arguments{
\item{model, }{a model}
}
\value{
A named vector with demand
}
\description{
Prepares a demand vector representing domestic production
Formula for production vector: y_p <- y_dc + y_e + y_d_delta
}
