% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{prepareEEIOModel}
\alias{prepareEEIOModel}
\title{Prepare economic components of an EEIO form USEEIO model.}
\usage{
prepareEEIOModel(modelname)
}
\arguments{
\item{modelname}{Name of the model from a config file.}
}
\value{
A list with USEEIO model economic components.
}
\description{
Prepare economic components of an EEIO form USEEIO model.
}
