% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowsaFunctions.R
\name{prepareFlowBySectorCollapsed}
\alias{prepareFlowBySectorCollapsed}
\title{Adjusts flowbysector data from flowsa}
\usage{
prepareFlowBySectorCollapsed(fbsc, satellite = TRUE)
}
\arguments{
\item{fbsc}{A FlowBySector collapsed df from flowsa}

\item{satellite}{bool, set to TRUE when used for env satellite tables}
}
\value{
A data frame of sector by region totals
}
\description{
Adjusts flowbysector data from flowsa
}
