% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandFunctions.R
\name{prepareImportDemand}
\alias{prepareImportDemand}
\title{Prepares a demand vector representing imports}
\usage{
prepareImportDemand(model)
}
\arguments{
\item{model, }{a model}
}
\value{
A named vector with demand
}
\description{
Prepares a demand vector representing imports
}
