% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandFunctions.R
\name{prepareImportProductionDemand}
\alias{prepareImportProductionDemand}
\title{Prepares a demand vector representing Import production}
\usage{
prepareImportProductionDemand(model, location)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{location, }{str of location code for demand vector}
}
\value{
A named vector with demand
}
\description{
Prepares a demand vector representing Import production
}
