% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCIAfmtFunctions.R
\name{prepareLCIAmethodforIndicators}
\alias{prepareLCIAmethodforIndicators}
\title{Prepares and reformats LCIAmethod data from LCIAformatter for use
modeled after prepareFlowBySectorCollapsedforSatellite}
\usage{
prepareLCIAmethodforIndicators(lciamethod)
}
\arguments{
\item{df}{A full LCIAmethod from LCIAformatter via getInventoryMethod}
}
\description{
Prepares and reformats LCIAmethod data from LCIAformatter for use
modeled after prepareFlowBySectorCollapsedforSatellite
}
