% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{prepareModelSectorCrosswalk}
\alias{prepareModelSectorCrosswalk}
\title{Prepare sectorcrosswalk table for a model}
\usage{
prepareModelSectorCrosswalk(model)
}
\arguments{
\item{model, }{any model object}
}
\value{
a data.frame, sectorcrosswalk table
}
\description{
Prepare sectorcrosswalk table for a model
}
