% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandFunctions.R
\name{prepareProductionDemand}
\alias{prepareProductionDemand}
\title{Prepares a demand vector representing production
Formula for production vector: y_p <- y_dc + y_e + y_d_delta}
\usage{
prepareProductionDemand(model)
}
\arguments{
\item{model, }{a model}
}
\value{
A named vector with demand
}
\description{
Prepares a demand vector representing production
Formula for production vector: y_p <- y_dc + y_e + y_d_delta
}
