% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{prepareTwoRegionDisaggregation}
\alias{prepareTwoRegionDisaggregation}
\title{Generate two-region disaggregation specs from a national spec}
\usage{
prepareTwoRegionDisaggregation(disagg, region, regions)
}
\description{
Generate two-region disaggregation specs from a national spec
}
