% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIOFunctions.R
\name{prepareWIODFfromFBS}
\alias{prepareWIODFfromFBS}
\title{Prepare make and use input files from FlowBySector file.}
\usage{
prepareWIODFfromFBS(fbs, spec, model)
}
\arguments{
\item{fbs}{FlowBySector dataframe.}

\item{spec}{WIO model spec}

\item{model}{An EEIO model object with model specs and IO tables loaded}
}
\value{
model with list of two dataframes: UseTableDF and MakeTableDF
}
\description{
Prepare make and use input files from FlowBySector file.
}
