% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateiorFunctions.R
\name{print2RValidationResults}
\alias{print2RValidationResults}
\title{Run validation checks for 2R models and print to console}
\usage{
print2RValidationResults(model)
}
\arguments{
\item{model}{A complete 2R EEIO model: a list with USEEIO model components and attributes}
}
\description{
Run validation checks for 2R models and print to console
}
