% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{printValidationResults}
\alias{printValidationResults}
\title{Run validation checks and print to console}
\usage{
printValidationResults(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes}
}
\description{
Run validation checks and print to console
}
