% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalImportFactors.R
\name{processImportFactors}
\alias{processImportFactors}
\title{Load and prepare import coefficients}
\usage{
processImportFactors(model, IFTable)
}
\arguments{
\item{model}{An EEIO form USEEIO model object with model specs loaded}

\item{IFTable, }{dataframe of unprocessed import factors}
}
\value{
IFTable, dataframe of processed of import coefficients (flow x sector).
}
\description{
Load and prepare import coefficients
}
