% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{RAS}
\alias{RAS}
\title{Generalized RAS procedure. Takes an initial matrix, a target row sum vector
and target colsum vector. Iterates until all row sums of matrix equal to row sum vector
and colsums of matrix equal col sum vector, within a tolerance}
\usage{
RAS(m0, t_r, t_c, t = 0.01, max_itr = 1000)
}
\description{
Generalized RAS procedure. Takes an initial matrix, a target row sum vector
and target colsum vector. Iterates until all row sums of matrix equal to row sum vector
and colsums of matrix equal col sum vector, within a tolerance
}
