% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalImportFactors.R
\name{readImportFactorTable}
\alias{readImportFactorTable}
\title{Load and prepare import coefficients}
\usage{
readImportFactorTable(IFSpec, configpaths = NULL)
}
\arguments{
\item{IFSpec}{list of specs for import factor file}

\item{configpaths}{str vector, paths (including file name) of model configuration file.
If NULL, built-in config files are used.}
}
\value{
IFtable, dataframe of unprocessed import factors
}
\description{
Load and prepare import coefficients
}
