% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{removeExtraSpaces}
\alias{removeExtraSpaces}
\title{Remove spaces around strings, like "321A "}
\usage{
removeExtraSpaces(s)
}
\value{
string
}
\description{
Remove spaces around strings, like "321A "
}
