% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{removeHybridProcesses}
\alias{removeHybridProcesses}
\title{Removes hybrid processes form a model object for successful validation}
\usage{
removeHybridProcesses(model, object)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes}

\item{object}{A model object in the form of a matrix or vector}
}
\value{
object with processes removed
}
\description{
Removes hybrid processes form a model object for successful validation
}
