% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{removeMissingSectors}
\alias{removeMissingSectors}
\title{Removes flow data where sectors are NA after mapping. Should only be used after checkSatelliteFlowLoss}
\usage{
removeMissingSectors(tbs)
}
\arguments{
\item{tbs, }{totals-by-sector df in model schema}
}
\value{
df, the modified tbs
}
\description{
Removes flow data where sectors are NA after mapping. Should only be used after checkSatelliteFlowLoss
}
