% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{removeModelSectors}
\alias{removeModelSectors}
\title{Remove sectors from the model objects according to the provided indexes}
\usage{
removeModelSectors(model, comIndexes, indIndexes)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{comIndexes}{A list containing the indexes of the commodities in the model}

\item{indIndexes}{A list containing the indexes of the industries in the model}
}
\value{
A model with the specified MUIO sectors in physical units and rearranged sector orders (optional)
}
\description{
Remove sectors from the model objects according to the provided indexes
}
