% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{removeNumberinSlashes}
\alias{removeNumberinSlashes}
\title{Remove numbers in slashes from a string, like /1/}
\usage{
removeNumberinSlashes(s)
}
\arguments{
\item{s, }{string}
}
\value{
A string with numbers in slashes removed
}
\description{
Remove numbers in slashes from a string, like /1/
}
