% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{removeRowsFromList}
\alias{removeRowsFromList}
\title{Remove specific rows from the specified list object in the model}
\usage{
removeRowsFromList(sectorList, indecesToAggregate)
}
\arguments{
\item{sectorList}{Model object to be aggregated}

\item{indecesToAggregate}{List of indeces of sectors to remove from list (i.e. aggregated sectors)}
}
\value{
An aggregated sectorList
}
\description{
Remove specific rows from the specified list object in the model
}
