% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIOFunctions.R
\name{removeSectorsFromWIOCW}
\alias{removeSectorsFromWIOCW}
\title{Remove sectors labeled by "Waste Treatment Commodities" and "Waste Treatment Industries", 
which are NOT specified by BEASectorsAsTreatmentSectors, from the spec object}
\usage{
removeSectorsFromWIOCW(model, spec)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{spec}{A dataframe with WIO specifications}
}
\value{
spec A dataframe with the specified sectors added to the spec$NAICSSetorCW object
}
\description{
Remove sectors labeled by "Waste Treatment Commodities" and "Waste Treatment Industries", 
which are NOT specified by BEASectorsAsTreatmentSectors, from the spec object
}
