% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{reorderModelSectors}
\alias{reorderModelSectors}
\title{Reorder sectors in the model objects according to the provided order.}
\usage{
reorderModelSectors(model, comOrder, indOrder)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{comOrder}{A list containing the order of the commodities in the model}

\item{indOrder}{A list containing the order of the industries in the model}
}
\value{
A model with the specified MUIO sectors in physical units and rearranged sector orders (optional)
}
\description{
Reorder sectors in the model objects according to the provided order.
}
