% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUIOFunctions.R
\name{reorderPhysicalSectors}
\alias{reorderPhysicalSectors}
\title{Reorder physical sectors such that they are in the top left of the Make and Use MUIO tables, as well as reorder all the relevant model objects.}
\usage{
reorderPhysicalSectors(
  model,
  physComIndeces,
  physIndIndeces,
  econComIndeces,
  econIndIndeces
)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{physComIndeces}{Indeces that specify the location of the physical MUIO commodities in the commodity lists}

\item{physIndIndeces}{Indeces that specify the location of the physical MUIO industries in the industry lists}

\item{econComIndeces}{Indeces that specify the location of the economic MUIO commodities (i.e., the "Other" sectors, X in code) in the commodity lists}

\item{econIndIndeces}{Indeces that specify the location of the economic MUIO industries (i.e., the "Other" sectors, X in code) in the industry lists}
}
\value{
A model with the specified MUIO sectors in physical units and rearranged sector orders (optional)
}
\description{
Reorder physical sectors such that they are in the top left of the Make and Use MUIO tables, as well as reorder all the relevant model objects.
}
