% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIOFunctions.R
\name{reorderWIOSectors}
\alias{reorderWIOSectors}
\title{Adjust WIO sectors such that they are in the correct order. Should only be needed in
the case that there is more than 1 WIOSpec.}
\usage{
reorderWIOSectors(model)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}
}
\value{
A model object with the WIO sectors in the correct order
}
\description{
Adjust WIO sectors such that they are in the correct order. Should only be needed in
the case that there is more than 1 WIOSpec.
}
