% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{replaceNonewithNA}
\alias{replaceNonewithNA}
\title{Replaces all `None` in a dataframe with the R NULL type NA}
\usage{
replaceNonewithNA(df)
}
\description{
Replaces all `None` in a dataframe with the R NULL type NA
}
