% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigDocumentation.R
\docType{data}
\name{SatelliteTableConfiguration}
\alias{SatelliteTableConfiguration}
\title{Satellite Table Configuration}
\format{
A list of satellite table attributes
\describe{
 \item{FullName}{string, Name of the satellite table}
 \item{Abbreviation}{string, Abbreviation for table}
 \item{StaticSource}{boolean, TRUE if loading table from static source}
 \item{StaticFile}{string, path to static file}
 \item{FileLocation}{string, description of the location of the source file}
 \item{DataYears}{list of data years}
 \item{Locations}{list of location acyronyms}
 \item{SectorListSource}{string, name of source "NAICS" or "BEA" accepted}
 \item{SectorListYear}{numeric, year of sector list schema, e.g. 2017}
 \item{SectorListLevel}{string, BEA IO levels. Options are "Detail","Summary", or "Sector"}
 \item{OriginalFlowSource}{string, name of original flow source for mapping}
 \item{DataSources}{list of data sources, @seealso [DataSourceConfiguration]}
}
}
\usage{
SatelliteTableConfiguration
}
\description{
Satellite Table Configuration
}
\keyword{datasets}
