% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQualityFunctions.R
\name{scoreContextualDQ}
\alias{scoreContextualDQ}
\title{Adds contextual data quality scores to a data frame with Year present}
\usage{
scoreContextualDQ(df)
}
\arguments{
\item{df}{A data frame containing a 'Year' column representing data year}
}
\value{
A data frame with contextual data quality scores added in columns
with names of the indicators. Only 'TemporalCorrelation' currently added.
}
\description{
Adds contextual data quality scores to a data frame with Year present
}
