% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQualityFunctions.R
\name{scoreTemporalDQ}
\alias{scoreTemporalDQ}
\title{Scores temporal data quality using a lookup based on difference between data year and target year}
\usage{
scoreTemporalDQ(data_year, target_year = NA, scoring_bounds)
}
\arguments{
\item{data_year}{integer year of data}

\item{target_year}{integer year of data, defaults to current year}

\item{scoring_bounds}{global scoring bounds}
}
\value{
An integer data quality score 1-5 or NA
}
\description{
Scores temporal data quality using a lookup based on difference between data year and target year
}
