% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_CommodityCodeName_2012}
\alias{Sector_CommodityCodeName_2012}
\title{Sector Commodity Code and Name (2012 schema)}
\format{A dataframe with 17 obs. and 2 variables}
\usage{
Sector_CommodityCodeName_2012
}
\description{
TBD
}
\keyword{datasets}
