% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_CPI_IO}
\alias{Sector_CPI_IO}
\title{Sector 2007-2017 CPI (2012 schema)}
\format{A dataframe with 15 obs. and 11 variables}
\usage{
Sector_CPI_IO
}
\description{
TBD
}
\keyword{datasets}
