% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_CPI_IO_17sch}
\alias{Sector_CPI_IO_17sch}
\title{Sector 1997-2023 CPI (2017 schema)}
\format{
A dataframe with 15 obs. and 27 variables
}
\source{
\url{https://apps.bea.gov/industry/Release/ZIP/UGdpByInd.zip}
}
\usage{
Sector_CPI_IO_17sch
}
\description{
Sector 1997-2023 CPI (2017 schema)
}
\keyword{datasets}
