% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_GrossOutput_IO}
\alias{Sector_GrossOutput_IO}
\title{Sector 2007-2018 Gross Output (2012 schema)}
\format{
A dataframe with 15 obs. and 17 variables
}
\source{
\url{https://apps.bea.gov//industry/iTables\%20Static\%20Files/AllTablesUnderlying.zip}
}
\usage{
Sector_GrossOutput_IO
}
\description{
TBD
}
\keyword{datasets}
