% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_GrossOutput_IO_17sch}
\alias{Sector_GrossOutput_IO_17sch}
\title{Sector 2017-2022 Gross Output (2017 schema)}
\format{
A dataframe with 15 obs. and 6 variables
}
\source{
\url{https://apps.bea.gov/industry/Release/ZIP/UGdpByInd.zip}
}
\usage{
Sector_GrossOutput_IO_17sch
}
\description{
Sector 2017-2022 Gross Output (2017 schema)
}
\keyword{datasets}
