% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_Make_2013_BeforeRedef}
\alias{Sector_Make_2013_BeforeRedef}
\title{Sector 2013 Make Before Redefinition (2012 schema)}
\format{A data frame with 16 obs. and 18 variables}
\usage{
Sector_Make_2013_BeforeRedef
}
\description{
TBD
}
\keyword{datasets}
