% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_Supply_2014}
\alias{Sector_Supply_2014}
\title{Sector 2014 Supply (2012 schema)}
\format{
A dataframe with 18 obs. and 27 variables
}
\source{
\url{https://apps.bea.gov/industry/iTables\%20Static\%20Files/AllTablesSUP.zip}
}
\usage{
Sector_Supply_2014
}
\description{
Sector 2014 Supply (2012 schema)
}
\keyword{datasets}
