% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_Use_2007_PUR_AfterRedef}
\alias{Sector_Use_2007_PUR_AfterRedef}
\title{Sector 2007 Use Purchaser's Value After Redefinition (2012 schema)}
\format{A data frame with 27 obs. and 28 variables}
\usage{
Sector_Use_2007_PUR_AfterRedef
}
\description{
TBD
}
\keyword{datasets}
