% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_Use_2012_PRO_BeforeRedef}
\alias{Sector_Use_2012_PRO_BeforeRedef}
\title{Sector 2012 Use Producer's Value Before Redefinition (2012 schema)}
\format{
A dataframe with 27 obs. and 28 variables
}
\source{
\url{https://edap-ord-data-commons.s3.amazonaws.com/useeior/AllTablesIO.zip}
}
\usage{
Sector_Use_2012_PRO_BeforeRedef
}
\description{
Sector 2012 Use Producer's Value Before Redefinition (2012 schema)
}
\keyword{datasets}
