% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_Use_2012_PUR_BeforeRedef}
\alias{Sector_Use_2012_PUR_BeforeRedef}
\title{Sector 2012 Use Purchaser's Value Before Redefinition (2012 schema)}
\format{
A dataframe with 27 obs. and 28 variables
}
\source{
\url{https://apps.bea.gov//industry/iTables\%20Static\%20Files/AllTablesIO.zip}
}
\usage{
Sector_Use_2012_PUR_BeforeRedef
}
\description{
TBD
}
\keyword{datasets}
