% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_Use_2020_PRO_AfterRedef}
\alias{Sector_Use_2020_PRO_AfterRedef}
\title{Sector 2020 Use Producer's Value After Redefinition (2012 schema)}
\format{
A dataframe with 27 obs. and 28 variables
}
\source{
\url{https://apps.bea.gov/industry/iTables\%20Static\%20Files/AllTablesIO.zip}
}
\usage{
Sector_Use_2020_PRO_AfterRedef
}
\description{
Sector 2020 Use Producer's Value After Redefinition (2012 schema)
}
\note{
Remove 'backslash' before 'percent sign' in url to access the source of data.
}
\keyword{datasets}
