% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_Use_2021_PRO_BeforeRedef}
\alias{Sector_Use_2021_PRO_BeforeRedef}
\title{Sector 2021 Use Producer's Value Before Redefinition (2012 schema)}
\format{
A dataframe with 23 obs. and 24 variables
}
\source{
\url{https://apps.bea.gov/industry/iTables\%20Static\%20Files/AllTablesIO.zip}
}
\usage{
Sector_Use_2021_PRO_BeforeRedef
}
\description{
Sector 2021 Use Producer's Value Before Redefinition (2012 schema)
}
\note{
Remove 'backslash' before 'percent sign' in url to access the source of data.
}
\keyword{datasets}
