% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_Use_SUT_2017}
\alias{Sector_Use_SUT_2017}
\title{Sector 2017 Use (2012 schema)}
\format{
A dataframe with 26 obs. and 22 variables
}
\source{
\url{https://apps.bea.gov/industry/iTables\%20Static\%20Files/AllTablesSUP.zip}
}
\usage{
Sector_Use_SUT_2017
}
\description{
Sector 2017 Use (2012 schema)
}
\keyword{datasets}
