% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_ValueAddedCodeName_2012}
\alias{Sector_ValueAddedCodeName_2012}
\title{Sector Value Added Code and Name (2012 schema)}
\format{
A dataframe with 3 obs. and 2 variables
}
\source{
\url{https://edap-ord-data-commons.s3.amazonaws.com/useeior/AllTablesIO.zip}
}
\usage{
Sector_ValueAddedCodeName_2012
}
\description{
Sector Value Added Code and Name (2012 schema)
}
\keyword{datasets}
