% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_ValueAddedCodeName_2017}
\alias{Sector_ValueAddedCodeName_2017}
\title{Sector Value Added Code and Name (2017 schema)}
\format{
A dataframe with 3 obs. and 2 variables
}
\source{
\url{https://apps.bea.gov/industry/iTables\%20Static\%20Files/AllTablesIO.zip}
}
\usage{
Sector_ValueAddedCodeName_2017
}
\description{
Sector Value Added Code and Name (2017 schema)
}
\keyword{datasets}
