% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{setCommonYearforFlow}
\alias{setCommonYearforFlow}
\title{Sets the Year in a tbs to be the year of the highest frequency for a given flow when that flow is reported
in more than a single year}
\usage{
setCommonYearforFlow(tbs)
}
\arguments{
\item{tbs, }{a model total by sector file}
}
\value{
df, the tbs
}
\description{
Sets the Year in a tbs to be the year of the highest frequency for a given flow when that flow is reported
in more than a single year
}
