% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQualityFunctions.R
\name{setDQScoringBounds}
\alias{setDQScoringBounds}
\title{Constants to determine which bound type to use for DQ indicator
and for each indicator, provide those bounds}
\usage{
setDQScoringBounds()
}
\value{
A list with DQ bounds
}
\description{
Constants to determine which bound type to use for DQ indicator
and for each indicator, provide those bounds
}
