% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{setWriteDirs}
\alias{setWriteDirs}
\title{Sets directories to write model output data to}
\usage{
setWriteDirs(model = NA)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes. Optional}
}
\value{
A named list of directories for model output writing
}
\description{
Sets directories to write model output data to. If model is not passed, just sets data directory.
}
