% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{specifiedMakeDisagg}
\alias{specifiedMakeDisagg}
\title{Disaggregate make table based on the allocations specified in the files referenced in the diaggregation specs.}
\usage{
specifiedMakeDisagg(model, disagg)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}
}
\value{
A standardized make table with old sectors removed and new disaggregated sectors added based on the allocations in the disaggregation specs.
}
\description{
Disaggregate make table based on the allocations specified in the files referenced in the diaggregation specs.
}
