% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{stackSatelliteTables}
\alias{stackSatelliteTables}
\title{Stacks two tables up}
\usage{
stackSatelliteTables(sattable1, sattable2)
}
\arguments{
\item{sattable1}{A standardized statellite table.}

\item{sattable2}{Another standardized statellite table.}
}
\value{
A complete standardized statellite table.
}
\description{
Stacks two tables up
}
