% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{standardizeandcastSatelliteTable}
\alias{standardizeandcastSatelliteTable}
\title{Converts flows table into flows x sector matrix-like format}
\usage{
standardizeandcastSatelliteTable(df, model)
}
\arguments{
\item{df}{a dataframe of flowables, contexts, units, sectors and locations}

\item{model}{an EEIO model with IO tables loaded}
}
\value{
a flows x sector matrix-like dataframe
}
\description{
Converts flows table into flows x sector matrix-like format
}
