% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIOFunctions.R
\name{subsetWIOSectors}
\alias{subsetWIOSectors}
\title{Get subset of sectors corresponding to WasteTreatmentIndustries, WasteTreatmentCommodities,
WasteGentTreatment, WasteGenMass, RecyclingTreatment, or RecyclingMass}
\usage{
subsetWIOSectors(model, WIO, WIOSection)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{WIO}{A list with WIO specifications and data}

\item{WIOSection}{A string that indicates which section of the WIO model the sectors belong to}
}
\value{
A model with the model WIO model objects with the correct subset of sectors and format
}
\description{
Get subset of sectors corresponding to WasteTreatmentIndustries, WasteTreatmentCommodities,
WasteGentTreatment, WasteGenMass, RecyclingTreatment, or RecyclingMass
}
