% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandFunctions.R
\name{sumDemandCols}
\alias{sumDemandCols}
\title{Sums across sectors for a given set of codes/cols in a given final demand df}
\usage{
sumDemandCols(Y, codes)
}
\arguments{
\item{Y, }{a model Demand df}

\item{codes, }{sector code(s) for a subset of Final Demand cols}
}
\value{
A named vector with model sectors and demand amounts
}
\description{
Sums across sectors for a given set of codes/cols in a given final demand df
}
