% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandFunctions.R
\name{sumforConsumption}
\alias{sumforConsumption}
\title{Sums the demand cols representing final consumption, i.e. household, investment, and government}
\usage{
sumforConsumption(model, Y)
}
\arguments{
\item{model, }{a model}

\item{Y, }{a model Demand df.}
}
\value{
a named vector with model sectors and demand amounts
}
\description{
Sums the demand cols representing final consumption, i.e. household, investment, and government
}
