% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{summarizeImpactResult}
\alias{summarizeImpactResult}
\title{Calculates the result matrix by sector for a specified indciator}
\usage{
summarizeImpactResult(y, model, indicator)
}
\arguments{
\item{y, }{a model demand vector}

\item{model, }{A complete EEIO Model object}

\item{indicator, }{str, index of a model indicator for use in the C matrix, e.g. "Acidification Potential"}
}
\value{
impacts, matrix of impacts sector purchased x sector sourced
}
\description{
Calculates the result matrix by sector for a specified indciator
}
