% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_CPI_IO_17sch}
\alias{Summary_CPI_IO_17sch}
\title{Summary 2017-2022 CPI (2017 schema)}
\format{
A dataframe with 71 obs. and 6 variables
}
\source{
\url{https://apps.bea.gov/industry/Release/ZIP/UGdpByInd.zip}
}
\usage{
Summary_CPI_IO_17sch
}
\description{
Summary 2017-2022 CPI (2017 schema)
}
\keyword{datasets}
