% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_FinalDemandCodeName_2017}
\alias{Summary_FinalDemandCodeName_2017}
\title{Summary Final Demand Code and Name (2017 schema)}
\format{
A dataframe with 20 obs. and 2 variables
}
\source{
\url{https://apps.bea.gov/industry/iTables\%20Static\%20Files/AllTablesIO.zip}
}
\usage{
Summary_FinalDemandCodeName_2017
}
\description{
Summary Final Demand Code and Name (2017 schema)
}
\keyword{datasets}
