% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_GrossOutput_IO}
\alias{Summary_GrossOutput_IO}
\title{Summary 2007-2017 Gross Output (2012 schema)}
\format{A data frame with 71 obs. and 11 variables}
\usage{
Summary_GrossOutput_IO
}
\description{
TBD
}
\keyword{datasets}
