% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Import_2012}
\alias{Summary_Import_2012}
\title{Summary 2012 Import (2012 schema)}
\format{A dataframe with 73 obs. and 93 variables}
\source{
\url{https://apps.bea.gov/industry/xls/io-annual/ImportMatrices_Before_Redefinitions_SUM_1997-2018.xlsx}
}
\usage{
Summary_Import_2012
}
\description{
TBD
}
\keyword{datasets}
