% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Import_2018_BeforeRedef}
\alias{Summary_Import_2018_BeforeRedef}
\title{Summary 2018 Import Before Redefinition (2012 schema)}
\format{
A dataframe with 73 obs. and 93 variables
}
\source{
\url{https://apps.bea.gov/industry/xls/io-annual/ImportMatrices_Before_Redefinitions_SUM_1997-2018.xlsx}
}
\usage{
Summary_Import_2018_BeforeRedef
}
\description{
Summary 2018 Import Before Redefinition (2012 schema)
}
\keyword{datasets}
