% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Import_2022_BeforeRedef_17sch}
\alias{Summary_Import_2022_BeforeRedef_17sch}
\title{Summary 2022 Import Before Redefinition (2017 schema)}
\format{
A dataframe with 73 obs. and 93 variables
}
\source{
\url{https://apps.bea.gov/industry/xls/io-annual/ImportMatrices_Before_Redefinitions_SUM_2017-2022.xlsx}
}
\usage{
Summary_Import_2022_BeforeRedef_17sch
}
\description{
Summary 2022 Import Before Redefinition (2017 schema)
}
\keyword{datasets}
