% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_IndustryCodeName_2012}
\alias{Summary_IndustryCodeName_2012}
\title{Summary Industry Code and Name (2012 schema)}
\format{A dataframe with 73 obs. and 2 variables}
\usage{
Summary_IndustryCodeName_2012
}
\description{
TBD
}
\keyword{datasets}
