% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Make_2010_BeforeRedef}
\alias{Summary_Make_2010_BeforeRedef}
\title{Summary 2010 Make Before Redefinition (2012 schema)}
\format{A dataframe with 72 obs. and 74 variables}
\usage{
Summary_Make_2010_BeforeRedef
}
\description{
TBD
}
\keyword{datasets}
