% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Make_2013_AfterRedef}
\alias{Summary_Make_2013_AfterRedef}
\title{Summary 2013 Make After Redefinition (2012 schema)}
\format{
A dataframe with 72 obs. and 74 variables
}
\source{
\url{https://apps.bea.gov//industry/iTables\%20Static\%20Files/AllTablesIO.zip}
}
\usage{
Summary_Make_2013_AfterRedef
}
\description{
TBD
}
\keyword{datasets}
