% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Make_2015_BeforeRedef}
\alias{Summary_Make_2015_BeforeRedef}
\title{Summary 2015 Make Before Redefinition (2012 schema)}
\format{A dataframe with 72 obs. and 74 variables}
\usage{
Summary_Make_2015_BeforeRedef
}
\description{
TBD
}
\keyword{datasets}
