% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Make_2016_AfterRedef}
\alias{Summary_Make_2016_AfterRedef}
\title{Summary 2016 Make After Redefinition (2012 schema)}
\format{A dataframe with 72 obs. and 74 variables}
\usage{
Summary_Make_2016_AfterRedef
}
\description{
TBD
}
\keyword{datasets}
