% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Make_2018_AfterRedef}
\alias{Summary_Make_2018_AfterRedef}
\title{Summary 2018 Make After Redefinition (2012 schema)}
\format{
A dataframe with 72 obs. and 74 variables
}
\source{
\url{https://edap-ord-data-commons.s3.amazonaws.com/useeior/AllTablesIO.zip}
}
\usage{
Summary_Make_2018_AfterRedef
}
\description{
Summary 2018 Make After Redefinition (2012 schema)
}
\keyword{datasets}
