% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Make_2020_AfterRedef}
\alias{Summary_Make_2020_AfterRedef}
\title{Summary 2020 Make After Redefinition (2012 schema)}
\format{
A dataframe with 72 obs. and 74 variables
}
\source{
\url{https://apps.bea.gov/industry/iTables\%20Static\%20Files/AllTablesIO.zip}
}
\usage{
Summary_Make_2020_AfterRedef
}
\description{
Summary 2020 Make After Redefinition (2012 schema)
}
\note{
Remove 'backslash' before 'percent sign' in url to access the source of data.
}
\keyword{datasets}
