% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Make_2022_AfterRedef_17sch}
\alias{Summary_Make_2022_AfterRedef_17sch}
\title{Summary 2022 Make After Redefinition (2017 schema)}
\format{
A dataframe with 72 obs. and 74 variables
}
\source{
\url{https://apps.bea.gov/industry/iTables\%20Static\%20Files/AllTablesIO.zip}
}
\usage{
Summary_Make_2022_AfterRedef_17sch
}
\description{
Summary 2022 Make After Redefinition (2017 schema)
}
\keyword{datasets}
