% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Make_2023_AfterRedef_17sch}
\alias{Summary_Make_2023_AfterRedef_17sch}
\title{Summary 2023 Make After Redefinition (2017 schema)}
\format{
A dataframe with 72 obs. and 74 variables
}
\source{
\url{https://apps.bea.gov/industry/release/zip/MAKE-USE-IMPORTS\%20(AFTER\%20REDEFINITIONS).zip}
}
\usage{
Summary_Make_2023_AfterRedef_17sch
}
\description{
Summary 2023 Make After Redefinition (2017 schema)
}
\keyword{datasets}
