% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Supply_2022}
\alias{Summary_Supply_2022}
\title{Summary 2022 Supply (2017 schema)}
\format{
A dataframe with 74 obs. and 83 variables
}
\source{
\url{https://apps.bea.gov/industry/iTables\%20Static\%20Files/AllTablesSUP.zip}
}
\usage{
Summary_Supply_2022
}
\description{
Summary 2022 Supply (2017 schema)
}
\keyword{datasets}
