% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Use_2007_PUR_BeforeRedef}
\alias{Summary_Use_2007_PUR_BeforeRedef}
\title{Summary 2007 Use Purchaser's Value Before Redefinition (2012 schema)}
\format{A data frame with 80 obs. and 98 variables}
\usage{
Summary_Use_2007_PUR_BeforeRedef
}
\description{
TBD
}
\keyword{datasets}
