% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Use_2010_PRO_BeforeRedef}
\alias{Summary_Use_2010_PRO_BeforeRedef}
\title{Summary 2010 Use Producer's Value Before Redefinition (2012 schema)}
\format{
A dataframe with 83 obs. and 98 variables
}
\source{
\url{https://edap-ord-data-commons.s3.amazonaws.com/useeior/AllTablesIO.zip}
}
\usage{
Summary_Use_2010_PRO_BeforeRedef
}
\description{
Summary 2010 Use Producer's Value Before Redefinition (2012 schema)
}
\keyword{datasets}
