% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Use_2012_PUR_BeforeRedef}
\alias{Summary_Use_2012_PUR_BeforeRedef}
\title{Summary 2012 Use Purchaser's Value Before Redefinition (2012 schema)}
\format{
A dataframe with 80 obs. and 98 variables
}
\source{
\url{https://apps.bea.gov//industry/iTables\%20Static\%20Files/AllTablesIO.zip}
}
\usage{
Summary_Use_2012_PUR_BeforeRedef
}
\description{
TBD
}
\keyword{datasets}
