% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Use_2014_PRO_BeforeRedef}
\alias{Summary_Use_2014_PRO_BeforeRedef}
\title{Summary 2014 Use Producer's Value Before Redefinition (2012 schema)}
\format{A data frame with 83 obs. and 98 variables}
\usage{
Summary_Use_2014_PRO_BeforeRedef
}
\description{
TBD
}
\keyword{datasets}
