% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Use_2015_PRO_AfterRedef}
\alias{Summary_Use_2015_PRO_AfterRedef}
\title{Summary 2015 Use Producer's Value After Redefinition (2012 schema)}
\format{A data frame with 83 obs. and 98 variables}
\usage{
Summary_Use_2015_PRO_AfterRedef
}
\description{
TBD
}
\keyword{datasets}
