% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Use_2018_PUR_AfterRedef}
\alias{Summary_Use_2018_PUR_AfterRedef}
\title{Summary 2018 Use Purchaser's Value After Redefinition (2012 schema)}
\format{A data frame with 80 obs. and 98 variables}
\usage{
Summary_Use_2018_PUR_AfterRedef
}
\description{
TBD
}
\keyword{datasets}
