% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Use_2019_PRO_BeforeRedef}
\alias{Summary_Use_2019_PRO_BeforeRedef}
\title{Summary 2019 Use Producer's Value Before Redefinition (2012 schema)}
\format{
A dataframe with 83 obs. and 98 variables
}
\source{
\url{https://apps.bea.gov/industry/iTables\%20Static\%20Files/AllTablesIO.zip}
}
\usage{
Summary_Use_2019_PRO_BeforeRedef
}
\description{
Summary 2019 Use Producer's Value Before Redefinition (2012 schema)
}
\note{
Remove 'backslash' before 'percent sign' in url to access the source of data.
}
\keyword{datasets}
