% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Use_2020_PRO_BeforeRedef_17sch}
\alias{Summary_Use_2020_PRO_BeforeRedef_17sch}
\title{Summary 2020 Use Producer's Value Before Redefinition (2017 schema)}
\format{
A dataframe with 79 obs. and 94 variables
}
\source{
\url{https://apps.bea.gov/industry/release/zip/MAKE-USE-IMPORTS\%20(BEFORE\%20REDEFINITIONS).zip}
}
\usage{
Summary_Use_2020_PRO_BeforeRedef_17sch
}
\description{
Summary 2020 Use Producer's Value Before Redefinition (2017 schema)
}
\keyword{datasets}
