% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Use_SUT_2010}
\alias{Summary_Use_SUT_2010}
\title{Summary 2010 Use (2012 schema)}
\format{
A dataframe with 82 obs. and 92 variables
}
\source{
\url{https://apps.bea.gov/industry/iTables\%20Static\%20Files/AllTablesSUP.zip}
}
\usage{
Summary_Use_SUT_2010
}
\description{
Summary 2010 Use (2012 schema)
}
\keyword{datasets}
