% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Use_SUT_2020_17sch}
\alias{Summary_Use_SUT_2020_17sch}
\title{Summary 2020 Use (2017 schema)}
\format{
A dataframe with 82 obs. and 92 variables
}
\source{
\url{https://apps.bea.gov/industry/iTables\%20Static\%20Files/AllTablesSUP.zip}
}
\usage{
Summary_Use_SUT_2020_17sch
}
\description{
Summary 2020 Use (2017 schema)
}
\keyword{datasets}
