% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_ValueAdded_IO_17sch}
\alias{Summary_ValueAdded_IO_17sch}
\title{Summary 1997-2023 Value Added (2017 schema)}
\format{
A dataframe with 71 obs. and 27 variables
}
\source{
\url{https://apps.bea.gov/industry/Release/ZIP/UGdpByInd.zip}
}
\usage{
Summary_ValueAdded_IO_17sch
}
\description{
Summary 1997-2023 Value Added (2017 schema)
}
\keyword{datasets}
