% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{testCalculationFunctions}
\alias{testCalculationFunctions}
\title{Test that model calculation functions are successful
Includes tests for the following functions:
adjustResultMatrixPrice, calculateFlowContributiontoImpact,
calculateSectorContributiontoImpact, disaggregateTotalToDirectAndTier1,
calculateSectorPurchasedbySectorSourcedImpact, aggregateResultMatrix,
calculateMarginSectorImpacts}
\usage{
testCalculationFunctions(model)
}
\arguments{
\item{model, }{A fully built EEIO model object}
}
\description{
Test that model calculation functions are successful
Includes tests for the following functions:
adjustResultMatrixPrice, calculateFlowContributiontoImpact,
calculateSectorContributiontoImpact, disaggregateTotalToDirectAndTier1,
calculateSectorPurchasedbySectorSourcedImpact, aggregateResultMatrix,
calculateMarginSectorImpacts
}
