% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{testVisualizationFunctions}
\alias{testVisualizationFunctions}
\title{Test that visualization functions are successful
Includes tests for the following functions:
barplotFloworImpactFractionbyRegion, barplotIndicatorScoresbySector, 
heatmapSatelliteTableCoverage, heatmapSectorRanking, plotMatrixCoefficient}
\usage{
testVisualizationFunctions(model)
}
\arguments{
\item{model, }{A fully built EEIO model object}
}
\description{
Test that visualization functions are successful
Includes tests for the following functions:
barplotFloworImpactFractionbyRegion, barplotIndicatorScoresbySector, 
heatmapSatelliteTableCoverage, heatmapSectorRanking, plotMatrixCoefficient
}
