% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{ToleranceforRAS}
\alias{ToleranceforRAS}
\title{Calculate tolerance for RAS. Takes a target row sum vector and target colsum vector.
Specify either relative difference or absolute difference .}
\usage{
ToleranceforRAS(t_r, t_c, relative_diff = NULL, absolute_diff = NULL)
}
\description{
Calculate tolerance for RAS. Takes a target row sum vector and target colsum vector.
Specify either relative difference or absolute difference .
}
