% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIOFunctions.R
\name{transformBEASectorToDFInput}
\alias{transformBEASectorToDFInput}
\title{Take an existing USEEIO sector and transform it into a dataframe formatted as a WIO-style input file}
\usage{
transformBEASectorToDFInput(model, spec, vectorToTransform)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{spec}{A model object contain the WIO specifications}

\item{vectorToTransform}{A string indicating which table and vector to use, e.g., UseRows, MakeCols, etc.}
}
\value{
A model with the UseTransactions matrix modified with WIO specs.
}
\description{
Take an existing USEEIO sector and transform it into a dataframe formatted as a WIO-style input file
}
