% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{transformDirectRequirementswithMarketShares}
\alias{transformDirectRequirementswithMarketShares}
\title{Transform Direct Requirements matrix with Market Shares matrix, works for both commodity-by-commodity and industry-by-industry model types.}
\usage{
transformDirectRequirementswithMarketShares(B, D, model)
}
\arguments{
\item{B}{Marginal impact per unit of the environmental flows.}

\item{D}{Market Shares matrix.}

\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
Direct Requirements matrix.
}
\description{
Transform Direct Requirements matrix with Market Shares matrix, works for both commodity-by-commodity and industry-by-industry model types.
}
