% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{transformFinalDemandwithMarketShares}
\alias{transformFinalDemandwithMarketShares}
\title{Transform Final Demand df with Market Shares matrix.}
\usage{
transformFinalDemandwithMarketShares(Fdf, model)
}
\arguments{
\item{Fdf}{Final Demand dataframe.}

\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
Final Demand matrix.
}
\description{
Transform Final Demand df with Market Shares matrix.
}
