% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{transformIndustryCPItoCommodityCPIforYear}
\alias{transformIndustryCPItoCommodityCPIforYear}
\title{Generate Commodity CPI by transforming Industry CPI using Commodity Mix matrix.}
\usage{
transformIndustryCPItoCommodityCPIforYear(year, model)
}
\arguments{
\item{year}{Year of Industry CPI.}

\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A dataframe contains adjusted Commodity CPI.
}
\description{
Generate Commodity CPI by transforming Industry CPI using Commodity Mix matrix.
}
