% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{uniformDisagg}
\alias{uniformDisagg}
\title{Disaggregate make or use table uniformly based on the number of new sectors}
\usage{
uniformDisagg(model, disagg, table)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}

\item{table}{DataFrame of make or use table}
}
\value{
A standardized make table with old sectors removed and new, uniformly disaggregated sectors added.
}
\description{
Disaggregate make or use table uniformly based on the number of new sectors
}
