% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{UniformMakeDisagg}
\alias{UniformMakeDisagg}
\title{Disaggregate make table uniformly based on the number of new sectors}
\usage{
UniformMakeDisagg(model, disagg)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}
}
\value{
A standardized make table with old sectors removed and new, uniformly disaggregated sectors added.
}
\description{
Disaggregate make table uniformly based on the number of new sectors
}
