% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{uniformUseDisagg}
\alias{uniformUseDisagg}
\title{Disaggregate use table uniformly based on the number of new sectors}
\usage{
uniformUseDisagg(model, disagg, domestic = FALSE)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}
}
\value{
A standardized use table with old sectors removed and new, uniformly disaggregated sectors added.
}
\description{
Disaggregate use table uniformly based on the number of new sectors
}
