% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateiorFunctions.R
\name{validate2RCommodityTotals}
\alias{validate2RCommodityTotals}
\title{Validate commodity totals between 2R Use table, Make table, and total commodity output objects}
\usage{
validate2RCommodityTotals(model)
}
\arguments{
\item{model}{A complete 2R EEIO model: a list with USEEIO model components and attributes}
}
\value{
A list containing failures of commodity total comparisons between various model objects.
}
\description{
Validate commodity totals between 2R Use table, Make table, and total commodity output objects
}
