% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{validate2RegionLagainstOutput}
\alias{validate2RegionLagainstOutput}
\title{Validate Leontief matrix (L) of two-region model and final demand against
SoI and RoUS output.}
\usage{
validate2RegionLagainstOutput(model, state = NULL)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes}

\item{state}{A text value specifying state of interest.}
}
\value{
A list of validation components and result.
}
\description{
Validate Leontief matrix (L) of two-region model and final demand against
SoI and RoUS output.
}
