% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{validateHouseholdEmissions}
\alias{validateHouseholdEmissions}
\title{Validate the calculation of household_emissions}
\usage{
validateHouseholdEmissions(model)
}
\arguments{
\item{model, }{A fully built EEIO model object}
}
\description{
Validate the calculation of household_emissions
}
