% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{validateImportFactorsApproach}
\alias{validateImportFactorsApproach}
\title{Validate the results of the model build using the Import Factor approach (i.e., coupled model approach)}
\usage{
validateImportFactorsApproach(model, demand = "Consumption")
}
\arguments{
\item{model, }{An EEIO model object with model specs and crosswalk table loaded}

\item{demand, }{A demand vector, has to be name of a built-in model demand vector, e.g. "Production" or "Consumption". Consumption used as default.}
}
\value{
A calculated direct requirements table
}
\description{
Validate the results of the model build using the Import Factor approach (i.e., coupled model approach)
}
