% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{validateResult}
\alias{validateResult}
\title{Validate result based on specified tolerance}
\usage{
validateResult(result, abs_diff = TRUE, tolerance)
}
\arguments{
\item{result}{A data object to be validated}

\item{abs_diff}{A logical value indicating whether to validate absolute values}

\item{tolerance}{A numeric value setting tolerance of the comparison}
}
\value{
A list contains confrontation details and validation results
}
\description{
Validate result based on specified tolerance
}
