% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{writeDatatoRDA}
\alias{writeDatatoRDA}
\title{Write external data to .rda.}
\usage{
writeDatatoRDA(data, data_name)
}
\arguments{
\item{data}{An R data object.}

\item{data_name}{A string specifying data name.}
}
\description{
Write external data to .rda.
}
