% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{writeMatrixasBinFile}
\alias{writeMatrixasBinFile}
\title{Write matrix as bin file}
\usage{
writeMatrixasBinFile(matrix, path)
}
\arguments{
\item{matrix}{A matrix to be written}

\item{path}{Path to write the bin file to}
}
\description{
Write matrix as bin file
}
