% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{writeMetadatatoJSON}
\alias{writeMetadatatoJSON}
\title{Write metadata of downloaded data to JSON.}
\usage{
writeMetadatatoJSON(name, year, source, url)
}
\arguments{
\item{name}{A string specifying data name.}

\item{year}{A numeric value specifying data year.}

\item{source}{A string specifying data source.}

\item{url}{A string specifying data url.}
}
\description{
Write metadata of downloaded data to JSON.
}
