% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeModelComponents}
\alias{writeModelComponents}
\title{Write model components to output folder using IO-Model-Builder format.}
\usage{
writeModelComponents(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\description{
Only writes model economic components (DRC, Marketshares, Demand) for now.
}
