% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeModelDemandstoJSON}
\alias{writeModelDemandstoJSON}
\title{Write model demand vectors as JSON files for API to output folder.}
\usage{
writeModelDemandstoJSON(model, outputfolder)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{outputfolder}{A directory to write model demand vectors.}
}
\description{
Writes model demand vectors, including y and y_d for consumption and production.
}
