% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeModelforAPI}
\alias{writeModelforAPI}
\title{Writes all model data and metadata components to the API}
\usage{
writeModelforAPI(model, basedir)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{basedir}{Base directory to write the model components to}
}
\description{
Writes all model data and metadata components to the API
}
