% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeModelforPY}
\alias{writeModelforPY}
\title{Write model components to output folder for useeiopy building using IO-Model-Builder format.}
\usage{
writeModelforPY(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{modelfolder}{Directory to write the model components to}
}
\description{
Only writes model economic components (DRC, Marketshares, Demand) for now.
}
