% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeModelMatricesforAPI}
\alias{writeModelMatricesforAPI}
\title{Write model matrices as BIN files for API to output folder.}
\usage{
writeModelMatricesforAPI(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\description{
Writes model matrices, including A, B, C, D, L, and U.
}
