% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeModelMatricestoXLSX}
\alias{writeModelMatricestoXLSX}
\title{Write model matrices and results to XLSX file}
\usage{
writeModelMatricestoXLSX(model)
}
\arguments{
\item{model, }{any model object}
}
\description{
Write model matrices and results to XLSX file
}
