% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeSessionInfotoFile}
\alias{writeSessionInfotoFile}
\title{Write out session information to a "Rsessioninfo.txt file in the given path}
\usage{
writeSessionInfotoFile(path)
}
\arguments{
\item{path, }{str, a path without the file}
}
\value{
None
}
\description{
Write out session information to a "Rsessioninfo.txt file in the given path
}
