% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{adjustMultiplierPrice}
\alias{adjustMultiplierPrice}
\title{Adjust multipliers based on currency year, price type, and margin type.}
\usage{
adjustMultiplierPrice(currency_year, purchaser_price = TRUE,
  margin_type = "industry", model)
}
\arguments{
\item{currency_year}{An integer representing the currency year.}

\item{purchaser_price}{A boolean value indicating whether to adjust producer's price to purchaser's price.}

\item{margin_type}{A character value: can be "industry" or "final consumer".}

\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A list of price-adjusted multipliers.
}
\description{
Adjust multipliers based on currency year, price type, and margin type.
}
