% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{aggregateMakeTable}
\alias{aggregateMakeTable}
\title{Aggregate the MakeTable based on specified source file}
\usage{
aggregateMakeTable(model, aggregationSpecs)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded.}

\item{aggregationSpecs}{Specifications for aggregation}
}
\value{
An aggregated MakeTable.
}
\description{
Aggregate the MakeTable based on specified source file
}
