% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{aggregateMasterCrosswalk}
\alias{aggregateMasterCrosswalk}
\title{Aggregate the MasterCrosswalk on the selected sectors}
\usage{
aggregateMasterCrosswalk(model, aggregationSpecs)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{aggregationSpecs}{Specifications for aggregation}
}
\value{
crosswalk with aggregated sectors removed
}
\description{
Aggregate the MasterCrosswalk on the selected sectors
}
