% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{aggregateMultiYearCPI}
\alias{aggregateMultiYearCPI}
\title{Aggregate MultiYear CPI model objects}
\usage{
aggregateMultiYearCPI(model, mainIndex, indecesToAggregate, type)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded.}

\item{mainIndex}{Index to aggregate the others to.}

\item{indecesToAggregate}{List of indeces to aggregate.}

\item{type}{String to designate either commodity or industry}
}
\value{
newCPI A dataframe with the aggregatded CPI values by year.
}
\description{
Aggregate MultiYear CPI model objects
}
