% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{aggregateMultiYearOutput}
\alias{aggregateMultiYearOutput}
\title{Aggregate MultiYear Output model objects}
\usage{
aggregateMultiYearOutput(originalOutput, mainIndex, indecesToAggregate)
}
\arguments{
\item{originalOutput}{MultiYear Output dataframe}

\item{mainIndex}{Index to aggregate the others to.}

\item{indecesToAggregate}{List of indeces to aggregate.}
}
\value{
model A dataframe with the disaggregated GDPGrossOutputIO by year.
}
\description{
Aggregate MultiYear Output model objects
}
