% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{aggregateResultMatrix}
\alias{aggregateResultMatrix}
\title{Aggregate result matrix by rows and columns}
\usage{
aggregateResultMatrix(matrix, to_level, crosswalk)
}
\arguments{
\item{matrix}{A matrix of inventory or impact results with sectors as both rows and columns,
such as the output from calculateSectorPurchasedbySectorSourcedImpact()}

\item{to_level}{The level of BEA code this matrix will be aggregated to}

\item{crosswalk}{Sector crosswalk between levels of detail}
}
\value{
An aggregated matrix with sectors as rows and columns
}
\description{
Aggregate result matrix by rows and columns
}
