% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{aggregateSatelliteTable}
\alias{aggregateSatelliteTable}
\title{Aggregate (FlowAmount in) satellite tables from BEA level to model configuration}
\usage{
aggregateSatelliteTable(sattable, from_level, model)
}
\arguments{
\item{sattable}{A satellite table to be aggregated based on the level (Detail, Summary, or Sector) of BEA code.}

\item{from_level}{The level of BEA code in the satellite table.}

\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A more aggregated satellite table.
}
\description{
Aggregate (FlowAmount in) satellite tables from BEA level to model configuration
}
