% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{balanceDisagg}
\alias{balanceDisagg}
\title{Balance the Make and Use tables after disaggregation.}
\usage{
balanceDisagg(model, disagg)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}
}
\value{
model object with RAS-balanced disaggregation sectors
}
\description{
Balance the Make and Use tables after disaggregation.
}
