% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationFunctions.R
\name{barplotFlowbyGroup}
\alias{barplotFlowbyGroup}
\title{Bar plot of flow by group}
\usage{
barplotFlowbyGroup(methodname, NAICS_digit, group)
}
\arguments{
\item{methodname}{Name of the parquet file loaded from flowsa}

\item{NAICS_digit}{Digit of NAICS code}

\item{group}{Group of flows}
}
\description{
Bar plot of flow by group
}
