% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationFunctions.R
\name{barplotFloworImpactFractionbyRegion}
\alias{barplotFloworImpactFractionbyRegion}
\title{Proportional bar chart splitting out flows or impacts by a region and the Rest of the region}
\usage{
barplotFloworImpactFractionbyRegion(R1_calc_result, Total_calc_result, x_title)
}
\arguments{
\item{R1_calc_result}{A matrix from model result.}

\item{Total_calc_result}{A matrix from model result.}

\item{x_title}{The title of x axis, excluding unit.}
}
\value{
a ggplot bar chart with horizontal orientation
}
\description{
Proportional bar chart splitting out flows or impacts by a region and the Rest of the region
}
