% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationFunctions.R
\name{barplotIndicatorScoresbySector}
\alias{barplotIndicatorScoresbySector}
\title{Bar plot of indicator scores calculated from totals by sector and displayed by BEA Sector Level to compare scores across models}
\usage{
barplotIndicatorScoresbySector(
  model_list,
  totals_by_sector_name,
  indicator_name,
  sector,
  y_title
)
}
\arguments{
\item{model_list}{List of EEIO models with IOdata, satellite tables, and indicators loaded}

\item{totals_by_sector_name}{The name of one of the totals by sector tables available in model$SatelliteTables$totals_by_sector}

\item{indicator_name}{The name of the indicator of interest from the model$Indicators$factors}

\item{sector}{Can be logical value or text. If non-boolean, it must be code of one or more BEA sectors.}

\item{y_title}{The title of y axis, excluding unit.}
}
\description{
Bar plot of indicator scores calculated from totals by sector and displayed by BEA Sector Level to compare scores across models
}
