% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{buildEEIOModel}
\alias{buildEEIOModel}
\title{Build an EEIO from USEEIO model.}
\usage{
buildEEIOModel(model)
}
\arguments{
\item{model}{the model prepared with prepareEEIOModel().}
}
\value{
A list with USEEIO model components and attributes.
}
\description{
Build an EEIO from USEEIO model.
}
