% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{buildModel}
\alias{buildModel}
\title{Build an EEIO model.}
\usage{
buildModel(modelname, configpaths = NULL)
}
\arguments{
\item{modelname}{Name of the model from a config file.}

\item{configpaths}{str vector, paths (including file name) of model configuration file
and optional agg/disagg configuration file(s). If NULL, built-in config files are used.}
}
\value{
A list of EEIO model complete components and attributes
}
\description{
Build an EEIO model.
}
