% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalImportFactors.R
\name{buildModelwithImportFactors}
\alias{buildModelwithImportFactors}
\title{Create import Use table and validate domestic+import against full use model .}
\usage{
buildModelwithImportFactors(model, configpaths = NULL)
}
\arguments{
\item{model, }{An EEIO model object with model specs and crosswalk table loaded}

\item{configpaths}{str vector, paths (including file name) of model configuration file
and optional agg/disagg configuration file(s). If NULL, built-in config files are used.}
}
\value{
A model object with explicit import components.
}
\description{
Create import Use table and validate domestic+import against full use model .
}
