% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{calculateAndValidateImportA}
\alias{calculateAndValidateImportA}
\title{Create import direct requirements table based on import use and import y tables}
\usage{
calculateAndValidateImportA(model, y = NULL, y_d = NULL)
}
\arguments{
\item{model, }{An EEIO model object with model specs and crosswalk table loaded}

\item{y}{A final demand VECTOR used for calculating the conventional model results, L*Y}

\item{y_d}{A final demand VECTOR used for validating the model results using the A_m calculation.}
}
\value{
A calculated direct requirements table
}
\description{
Create import direct requirements table based on import use and import y tables
}
