% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{calculateDirectPerspectiveLCIA}
\alias{calculateDirectPerspectiveLCIA}
\title{The direct perspective LCIA aligns impacts with sectors consumed by direct use.
Multiply the D matrix (the product of C matrix and B matrix) and scaling vector s.}
\usage{
calculateDirectPerspectiveLCIA(D, s)
}
\arguments{
\item{D}{Direct impact per unit of the environmental flows.}

\item{s}{Scaling vector.}
}
\value{
A matrix with direct impacts in form of sector x impact categories.
}
\description{
The direct perspective LCIA aligns impacts with sectors consumed by direct use.
Multiply the D matrix (the product of C matrix and B matrix) and scaling vector s.
}
\references{
Yang, Yi, Wesley W. Ingwersen, Troy R. Hawkins, Michael Srocka, and David E. Meyer.
2017. “USEEIO: A New and Transparent United States Environmentally-Extended Input-Output Model.”
Journal of Cleaner Production 158 (August): 308–18. https://doi.org/10.1016/j.jclepro.2017.04.150.
SI1, Equation 9.
}
