% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{calculateEEIOModel}
\alias{calculateEEIOModel}
\title{Calculate total emissions/resources (LCI) and total impacts (LCIA) for an EEIO model
for a given perspective and demand vector.}
\usage{
calculateEEIOModel(
  model,
  perspective,
  demand = "Production",
  location = NULL,
  use_domestic_requirements = FALSE,
  household_emissions = FALSE,
  show_RoW = FALSE
)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{perspective}{Perspective of the model, can be "DIRECT" or "FINAL". "DIRECT" perspective
aligns results with the sectors in which they are produced, while "FINAL" perspective aligns
results with the sectors consumed by the final user.}

\item{demand}{A demand vector, can be name of a built-in model demand vector, e.g. "Production" or "Consumption",
or an actual demand vector with names as one or more model sectors and
numeric values in USD with the same dollar year as model.}

\item{location, }{str optional location code for demand vector, required for two-region models}

\item{use_domestic_requirements}{A logical value: if TRUE, use domestic demand and L_d matrix;
if FALSE, use complete demand and L matrix.}

\item{household_emissions, }{bool, if TRUE, include calculation of emissions from households}

\item{show_RoW, }{bool, if TRUE, include rows for commodities in RoW, e.g. `111CA/RoW` in result objects.
Only valid currently for models with ExternalImportFactors.}
}
\value{
A list with LCI and LCIA results (in data.frame format) of the EEIO model.
}
\description{
Calculate total emissions/resources (LCI) and total impacts (LCIA) for an EEIO model
for a given perspective and demand vector.
}
