% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{calculateEEIOModel}
\alias{calculateEEIOModel}
\title{Calculate total emissions/resources (LCI) or total impact for USEEIO model for a given demand vector and perspective.}
\usage{
calculateEEIOModel(
  model,
  perspective,
  demand = "Production",
  use_domestic_requirements = FALSE
)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{perspective}{Perspective of the model, can be "DIRECT", "INTERMEDIATE", or "FINAL".}

\item{demand}{A name of a built in model demand vector or a named vector with names as one or more model sectors and numeric values in USD with the same dollar year as model}

\item{use_domestic_requirements}{A boolean value: if TRUE, use domestic A_d; if FALSE, use A matrices.}
}
\value{
A list with LCI and LCIA results of the EEIO model.
}
\description{
Calculate total emissions/resources (LCI) or total impact for USEEIO model for a given demand vector and perspective.
}
