% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{calculateEEIOModel}
\alias{calculateEEIOModel}
\title{Calculate total emissions/resources (LCI) and total impacts (LCIA) for an EEIO model
for a given perspective and demand vector.}
\usage{
calculateEEIOModel(
  model,
  perspective,
  demand = "Production",
  use_domestic_requirements = FALSE
)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{perspective}{Perspective of the model, can be "DIRECT", "INTERMEDIATE", or "FINAL".}

\item{demand}{A demand vector, can be name of a built-in model demand vector, e.g. "Production",
or an actual demand vector with names as one or more model sectors and
numeric values in USD with the same dollar year as model.}

\item{use_domestic_requirements}{A logical value: if TRUE, use domestic demand and L_d matrix;
if FALSE, use complete demand and L matrix.}
}
\value{
A list with LCI and LCIA results (in data.frame format) of the EEIO model.
}
\description{
Calculate total emissions/resources (LCI) and total impacts (LCIA) for an EEIO model
for a given perspective and demand vector.
}
