% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{calculateFlowContributiontoImpact}
\alias{calculateFlowContributiontoImpact}
\title{Calculate the percent contribution of M flows to an N indicator result
Uses model M matrix for flows and C matrix for indicator}
\usage{
calculateFlowContributiontoImpact(model, sector, indicator, domestic = FALSE)
}
\arguments{
\item{model, }{A complete EEIO Model object}

\item{sector, }{str, index of a model sector for use in the M matrix, e.g. "221100/us"}

\item{indicator, }{str, index of a model indicator for use in the C matrix, e.g. "Acidification Potential"}

\item{domestic, }{boolean, sets model to use domestic flow matrix.  Default is FALSE.}
}
\value{
df, dataframe sorted from highest flow contribution "contribution", also showing "x","y","impact"
}
\description{
Calculate the percent contribution of M flows to an N indicator result
Uses model M matrix for flows and C matrix for indicator
}
