% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{calculateGhoshG}
\alias{calculateGhoshG}
\title{Function to calculate inverse of Ghosh matrix B, returns the inverse G}
\usage{
calculateGhoshG(GhoshB)
}
\arguments{
\item{GhoshB}{Ghosh matrix B}
}
\value{
G, Ghosh matrix G
}
\description{
Function to calculate inverse of Ghosh matrix B, returns the inverse G
}
