% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{calculateOutputRatio}
\alias{calculateOutputRatio}
\title{Generate Output Ratio table, flexible to Commodity/Industry output and model Commodity/Industry type}
\usage{
calculateOutputRatio(model, output_type = "Commodity")
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{output_type}{Either Commodity or Industry, default is Commodity}
}
\value{
A data frame of output ratio
}
\description{
Generate Output Ratio table, flexible to Commodity/Industry output and model Commodity/Industry type
}
