% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{calculateResultsWithExternalFactors}
\alias{calculateResultsWithExternalFactors}
\title{Calculate total emissions/resources (LCI) and total impacts (LCIA) for an EEIO model that has external import factors
for a given demand vector.
Note that for this calculation, perspective is always FINAL}
\usage{
calculateResultsWithExternalFactors(model, demand = "Production")
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{demand}{A demand vector, can be name of a built-in model demand vector, e.g. "Production" or "Consumption",
or an actual demand vector with names as one or more model sectors and
numeric values in USD with the same dollar year as model.}
}
\value{
A list with LCI and LCIA results (in data.frame format) of the EEIO model.
}
\description{
Calculate total emissions/resources (LCI) and total impacts (LCIA) for an EEIO model that has external import factors
for a given demand vector.
Note that for this calculation, perspective is always FINAL
}
