% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{calculateResultsWithExternalFactors}
\alias{calculateResultsWithExternalFactors}
\title{Calculate total emissions/resources (LCI) and total impacts (LCIA) for an EEIO model that has external import factors
for a given demand vector.
Note that for this calculation, perspective is always FINAL}
\usage{
calculateResultsWithExternalFactors(
  model,
  demand = "Consumption",
  use_domestic_requirements = FALSE,
  household_emissions = FALSE
)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{demand}{A demand vector, can be name of a built-in model demand vector, e.g. "Production" or "Consumption"}

\item{use_domestic_requirements}{bool, if TRUE, return only domestic portion of results}

\item{household_emissions, }{bool, if TRUE, include calculation of emissions from households}
}
\value{
A list with LCI and LCIA results (in data.frame format) of the EEIO model.
}
\description{
Calculate total emissions/resources (LCI) and total impacts (LCIA) for an EEIO model that has external import factors
for a given demand vector.
Note that for this calculation, perspective is always FINAL
}
