% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{calculateTotalImpactbyTier1Purchases}
\alias{calculateTotalImpactbyTier1Purchases}
\title{Calculate sector x sector total impacts (single indicator) for Tier 1 purchases
Multiply each row of sector x sector A matrix by scalar elements of an
indicator (single) x sector array from N}
\usage{
calculateTotalImpactbyTier1Purchases(model, indicator)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes}

\item{indicator}{str, index of a model indicator, e.g. "Greenhouse Gases".}
}
\value{
A sector x sector, impact-per-tier-1-purchase matrix.
}
\description{
Calculate sector x sector total impacts (single indicator) for Tier 1 purchases
Multiply each row of sector x sector A matrix by scalar elements of an
indicator (single) x sector array from N
}
