% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{compareCommodityOutputandDomesticUseplusProductionDemand}
\alias{compareCommodityOutputandDomesticUseplusProductionDemand}
\title{Compares the total commodity output against the summation of model domestic Use and production demand}
\usage{
compareCommodityOutputandDomesticUseplusProductionDemand(
  model,
  tolerance = 0.05
)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes}

\item{tolerance, }{a numeric value, tolerance level of the comparison}
}
\value{
A list with pass/fail validation result and the cell-by-cell relative diff matrix
}
\description{
Compares the total commodity output against the summation of model domestic Use and production demand
}
