% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{compareEandLCIResult}
\alias{compareEandLCIResult}
\title{Compares the total flows against the model flow totals result calculation with the total demand}
\usage{
compareEandLCIResult(
  model,
  output_type,
  use_domestic = FALSE,
  tolerance = 0.05
)
}
\arguments{
\item{model, }{EEIOmodel object completely built}
}
\description{
Compares the total flows against the model flow totals result calculation with the total demand
}
