% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareModels.R
\name{compareFlowTotals}
\alias{compareFlowTotals}
\title{Compare flow totals for two models}
\usage{
compareFlowTotals(modelA, modelB)
}
\arguments{
\item{modelA, }{a useeior model}

\item{modelB}{a useeior model}
}
\value{
comparison, a list with pass/fail comparison results
}
\description{
Compare flow totals for two models
}
