% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareModels.R
\name{compareFlowTotals}
\alias{compareFlowTotals}
\title{Compare flow totals for two models}
\usage{
compareFlowTotals(TbSA, TbSB)
}
\arguments{
\item{TbSA}{a totals-by-sector list from a model}

\item{TbSB}{a totals-by-sector list from a model}
}
\value{
df, the validate::confrontation result
}
\description{
Compare flow totals for two models
}
