% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareModels.R
\name{compareFlowTotals}
\alias{compareFlowTotals}
\title{Compare flow totals for two models}
\usage{
compareFlowTotals(modelA, modelB)
}
\arguments{
\item{modelA}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{modelB}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
a list with pass/fail comparison results
}
\description{
Compare flow totals for two models
}
