% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{compareIndustryOutputinMakeandUse}
\alias{compareIndustryOutputinMakeandUse}
\title{Gets industry output from model Use and Make and checks if they are the same}
\usage{
compareIndustryOutputinMakeandUse(model)
}
\arguments{
\item{model, }{a built model object}
}
\description{
Gets industry output from model Use and Make and checks if they are the same
}
