% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{compareOutputandLeontiefXDemand}
\alias{compareOutputandLeontiefXDemand}
\title{Compares the total sector output against the model result calculation with the demand vector. and direct perspective.
Uses the model$FinalDemand and model$L
Works for the domestic model with the equivalent tables}
\usage{
compareOutputandLeontiefXDemand(model, use_domestic = FALSE, tolerance = 0.05)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes}

\item{use_domestic, }{a logical value indicating whether to use domestic demand vector}

\item{tolerance, }{a numeric value, tolerance level of the comparison}
}
\value{
A list with pass/fail validation result and the cell-by-cell relative diff matrix
}
\description{
Compares the total sector output against the model result calculation with the demand vector. and direct perspective.
Uses the model$FinalDemand and model$L
Works for the domestic model with the equivalent tables
}
