% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{constructEEIOMatrices}
\alias{constructEEIOMatrices}
\title{Construct EEIO matrices based on loaded IO tables, built satellite tables,
and indicator tables.}
\usage{
constructEEIOMatrices(model)
}
\arguments{
\item{model}{Model file loaded with IO tables, satellite tables, and indicator tables.}
}
\value{
A list with EEIO matrices..
}
\description{
Construct EEIO matrices based on loaded IO tables, built satellite tables,
and indicator tables.
}
