% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{convertStrEncodingLatintoASCII}
\alias{convertStrEncodingLatintoASCII}
\title{Forces a string encoding to ASCII from Latin-1}
\usage{
convertStrEncodingLatintoASCII(s)
}
\arguments{
\item{s, }{string with Latin-1 encoding}
}
\value{
A string with ASCII encoding
}
\description{
Forces a string encoding to ASCII from Latin-1
}
