% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{createBfromFlowDataandOutput}
\alias{createBfromFlowDataandOutput}
\title{Creates the B matrix from the flow data}
\usage{
createBfromFlowDataandOutput(model)
}
\arguments{
\item{model, }{a model with econ and flow data loaded}
}
\value{
B, a matrix in flow x sector format with values of flow per $ output sector
}
\description{
Creates the B matrix from the flow data
}
