% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQualityFunctions.R
\name{createDdqi}
\alias{createDdqi}
\title{Creates D_dqi matrix from B_dqi}
\usage{
createDdqi(model)
}
\arguments{
\item{model}{A complete EEIO model object.}
}
\value{
A 3d matrix of dqi scores for D
}
\description{
Using the formula: (C \%*\% (B * B_dqi)) / D
}
