% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMarginImpacts.R
\name{deriveMarginSectorImpacts}
\alias{deriveMarginSectorImpacts}
\title{Prepare M and U matrices with sector margin impacts}
\usage{
deriveMarginSectorImpacts(model)
}
\arguments{
\item{model}{A complete EEIO model}
}
\value{
A list with M_margin and U_margin
}
\description{
Prepare M and U matrices with sector margin impacts
}
