% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMarginImpacts.R
\name{deriveMarginSectorImpacts}
\alias{deriveMarginSectorImpacts}
\title{Prepare M and U matrices with sector margin impacts}
\usage{
deriveMarginSectorImpacts(model, margin_type = "intermediate")
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{margin_type}{A character value: can be "intermediate" or "final consumer".}
}
\value{
A list with M_m and U_m
}
\description{
Prepare M and U matrices with sector margin impacts
}
