% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_CPI_IO}
\alias{Detail_CPI_IO}
\title{Detail 2007-2018 CPI (2012 schema)}
\format{
A dataframe with 405 obs. and 17 variables
}
\source{
\url{https://apps.bea.gov//industry/iTables\%20Static\%20Files/AllTablesUnderlying.zip}
}
\usage{
Detail_CPI_IO
}
\description{
TBD
}
\keyword{datasets}
