% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_CPI_IO}
\alias{Detail_CPI_IO}
\title{Detail 2007-2018 CPI (2012 schema)}
\format{
A dataframe with 405 obs. and 17 variables
}
\source{
\url{https://edap-ord-data-commons.s3.amazonaws.com/useeior/AllTablesUnderlying.zip}
}
\usage{
Detail_CPI_IO
}
\description{
Detail 2007-2018 CPI (2012 schema)
}
\keyword{datasets}
