% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_Make_2012_BeforeRedef}
\alias{Detail_Make_2012_BeforeRedef}
\title{Detail 2012 Make Before Redefinition (2012 schema)}
\format{
A dataframe with 406 obs. and 406 variables
}
\source{
\url{https://edap-ord-data-commons.s3.amazonaws.com/useeior/AllTablesIO.zip}
}
\usage{
Detail_Make_2012_BeforeRedef
}
\description{
Detail 2012 Make Before Redefinition (2012 schema)
}
\keyword{datasets}
