% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Detail_Use_2012_PUR_BeforeRedef}
\alias{Detail_Use_2012_PUR_BeforeRedef}
\title{Detail 2012 Use Purchaser's Value Before Redefinition (2012 schema)}
\format{
A dataframe with 403 obs. and 428 variables
}
\source{
\url{https://edap-ord-data-commons.s3.amazonaws.com/useeior/AllTablesIO.zip}
}
\usage{
Detail_Use_2012_PUR_BeforeRedef
}
\description{
Detail 2012 Use Purchaser's Value Before Redefinition (2012 schema)
}
\keyword{datasets}
