% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateFinalDemand}
\alias{disaggregateFinalDemand}
\title{Disaggregate Final Demand based on specs}
\usage{
disaggregateFinalDemand(model, disagg, domestic = FALSE)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}

\item{domestic}{A logical value indicating whether to disaggregate domestic final demand.}
}
\value{
A standardized final demand table with old sectors removed and new sectors with manual and default allocations added.
}
\description{
Disaggregate Final Demand based on specs
}
