% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateMasterCrosswalk}
\alias{disaggregateMasterCrosswalk}
\title{Disaggregate the MasterCrosswalk to include the new sectors for disaggregation}
\usage{
disaggregateMasterCrosswalk(model, disagg)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}
}
\value{
crosswalk with new sectors added.
}
\description{
Disaggregate the MasterCrosswalk to include the new sectors for disaggregation
}
