% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateModel}
\alias{disaggregateModel}
\title{Disaggregate a model based on specified source file}
\usage{
disaggregateModel(model)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}
}
\value{
A disaggregated model.
}
\description{
Disaggregate a model based on specified source file
}
