% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateMultiYearOutput}
\alias{disaggregateMultiYearOutput}
\title{Disaggregate MultiYear Output model objects}
\usage{
disaggregateMultiYearOutput(model, disagg, output_type = "Commodity")
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}

\item{output_type}{A string that indicates whether the Commodity or Industry output should be disaggregated}
}
\value{
model A dataframe with the disaggregated GDPGrossOutputIO by year
}
\description{
Disaggregate MultiYear Output model objects
}
