% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateSatelliteSubsetByRatio}
\alias{disaggregateSatelliteSubsetByRatio}
\title{Disaggregate a portion of a satellite table based on an allocation_vector}
\usage{
disaggregateSatelliteSubsetByRatio(
  sattable,
  disagg,
  allocating_sectors,
  allocation_vector = NULL
)
}
\arguments{
\item{sattable}{A standardized satellite table to be disaggregated.}

\item{disagg}{Specifications for disaggregating the current Table}

\item{allocating_sectors}{vector of sectors to allocate to}

\item{allocation_vector}{named vector of allocation ratios}
}
\value{
A satellite table with new sectors added.
}
\description{
Disaggregate a portion of a satellite table based on an allocation_vector
}
