% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{disaggregateSatelliteTable}
\alias{disaggregateSatelliteTable}
\title{Disaggregate satellite tables from static file based on specs}
\usage{
disaggregateSatelliteTable(disagg, tbs, sat_spec)
}
\arguments{
\item{disagg}{Specifications for disaggregating the current Table}

\item{tbs}{A standardized satellite table with resource and emission names from original sources.}

\item{sat_spec, }{a standard specification for a single satellite table.}
}
\value{
A standardized satellite table with old sectors removed and new sectors added.
}
\description{
Disaggregate satellite tables from static file based on specs
}
