% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{disaggregateTotalToDirectAndTier1}
\alias{disaggregateTotalToDirectAndTier1}
\title{For a given indicator, disaggregate total impacts per purchase (N) into 
direct impacts (D) and upstream, Tier 1 purchase impacts. Return a long format
dataframe of exchanges, with sector names mapped to sector codes.}
\usage{
disaggregateTotalToDirectAndTier1(model, indicator)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes}

\item{indicator}{str, index of a model indicator, e.g. "Greenhouse Gases".}
}
\value{
A data frame of direct and per-tier-1-purchase sector impacts
}
\description{
For a given indicator, disaggregate total impacts per purchase (N) into 
direct impacts (D) and upstream, Tier 1 purchase impacts. Return a long format
dataframe of exchanges, with sector names mapped to sector codes.
}
