% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowMappingFunctions.R
\docType{data}
\name{flowmapping}
\alias{flowmapping}
\title{Functions to map elementary or waste flows in satellite tables drawing on an internal mapping file
Import flow mapping file.}
\format{
An object of class \code{data.frame} with 615 rows and 28 columns.
}
\usage{
flowmapping
}
\description{
Functions to map elementary or waste flows in satellite tables drawing on an internal mapping file
Import flow mapping file.
}
\keyword{datasets}
