% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{formatLocationforStateModels}
\alias{formatLocationforStateModels}
\title{Format location in state models from formal state name to US-ST}
\usage{
formatLocationforStateModels(location)
}
\arguments{
\item{location}{A text value of input location name}
}
\value{
A text value of formatted location for state models
}
\description{
Format location in state models from formal state name to US-ST
}
