% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{formatValidationResult}
\alias{formatValidationResult}
\title{Format validation result}
\usage{
formatValidationResult(result, abs_diff = TRUE, tolerance)
}
\arguments{
\item{abs_diff}{A boolean value indicating whether to validate absolute values}

\item{tolerance}{A numeric value setting tolerance of the comparison}

\item{df}{A data.frame to be validated}
}
\value{
A list contains formatted validation results
}
\description{
Format validation result
}
