% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{formatValidationResult}
\alias{formatValidationResult}
\title{Format validation result}
\usage{
formatValidationResult(result, abs_diff = TRUE, tolerance)
}
\arguments{
\item{result}{Validation result to be formatted}

\item{abs_diff}{A logical value indicating whether to validate absolute values}

\item{tolerance}{A numeric value setting tolerance of the comparison}
}
\value{
A list contains formatted validation results
}
\description{
Format validation result
}
