% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{generateCbSfromTbSandModel}
\alias{generateCbSfromTbSandModel}
\title{Prepare coefficients (x unit/$) from the totals by flow and sector (x unit)}
\usage{
generateCbSfromTbSandModel(model)
}
\arguments{
\item{model, }{a model with econ and flow data loaded}

\item{TbS, }{a totals by sector dataframe}
}
\value{
df, a Coefficients-by-sector table
}
\description{
Prepare coefficients (x unit/$) from the totals by flow and sector (x unit)
}
