% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{generateCbSfromTbSandModel}
\alias{generateCbSfromTbSandModel}
\title{Prepare coefficients (x unit/$) from the totals by flow and sector (x unit)}
\usage{
generateCbSfromTbSandModel(model)
}
\arguments{
\item{model}{An EEIO model object with model specs, IO tables, satellite tables, and indicators loaded}
}
\value{
A dataframe of Coefficients-by-Sector (CbS) table
}
\description{
Prepare coefficients (x unit/$) from the totals by flow and sector (x unit)
}
