% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{generateDirectRequirementsfromUse}
\alias{generateDirectRequirementsfromUse}
\title{Generate Direct Requirements matrix from Use table.}
\usage{
generateDirectRequirementsfromUse(model, domestic)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{domestic}{A logical parameter indicating whether to calculate DR or Domestic DR.}
}
\value{
Direct Requirements matrix of the model.
}
\description{
Generate Direct Requirements matrix from Use table.
}
