% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_LCIAFunctions.R
\name{generateLCIA}
\alias{generateLCIA}
\title{Deprecated function for generating LCIA output using LCIA_indicators static file}
\usage{
generateLCIA(version)
}
\description{
Deprecated function for generating LCIA output using LCIA_indicators static file
}
