% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIndicators.R
\name{generateLCIA}
\alias{generateLCIA}
\title{Deprecated function for generating LCIA output using LCIA_indicators static file}
\usage{
generateLCIA(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A LCIA dataframe not yet formatted for IOMB.
}
\description{
Deprecated function for generating LCIA output using LCIA_indicators static file
}
