% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{generateModelSectorSchema}
\alias{generateModelSectorSchema}
\title{Create sector schema for a model}
\usage{
generateModelSectorSchema(model)
}
\arguments{
\item{model}{An EEIO model object with model specs loaded}
}
\value{
A string of sector schema for a model
}
\description{
Create sector schema for a model
}
