% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{getAdjustedOutput}
\alias{getAdjustedOutput}
\title{Adjust Industry output based on CPI.}
\usage{
getAdjustedOutput(outputyear, referenceyear, location_acronym, IsRoUS, model)
}
\arguments{
\item{outputyear}{Year of Industry output.}

\item{referenceyear}{Year of the currency reference.}

\item{location_acronym}{Abbreviated location name of the model, e.g. "US" or "GA".}

\item{IsRoUS}{A logical parameter indicating whether to adjust Industry output for Rest of US (RoUS).}

\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A dataframe contains adjusted Industry output with row names being BEA sector code.
}
\description{
Adjust Industry output based on CPI.
}
