% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{getAggregationSpecs}
\alias{getAggregationSpecs}
\title{Obtain aggregation specs from input files}
\usage{
getAggregationSpecs(model, configpaths = NULL)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{configpaths}{str vector, paths (including file name) of agg configuration file(s).
If NULL, built-in config files are used.}
}
\value{
A model with the specified aggregation and disaggregation specs.
}
\description{
Obtain aggregation specs from input files
}
