% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateOutput.R
\name{getAlternativeStateIndustryOutputEstimates}
\alias{getAlternativeStateIndustryOutputEstimates}
\title{Estimate state output based on alternative sources.}
\usage{
getAlternativeStateIndustryOutputEstimates(year)
}
\arguments{
\item{year}{A numeric value between 2007 and 2017 specifying the year of interest.}
}
\value{
A dataframe contains state output for all states and a specific industry.
}
\description{
Estimate state output based on alternative sources.
}
