% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCIAfmtFunctions.R
\name{getCombinedImpactMethods}
\alias{getCombinedImpactMethods}
\title{Get and combine impact methods using the Python LCIAformatter package's get_mapped_method function.}
\usage{
getCombinedImpactMethods(ind_spec)
}
\arguments{
\item{ind_spec}{Specification of an indicator}
}
\value{
An LCIAmethod with the specified indicators
}
\description{
Get and combine impact methods using the Python LCIAformatter package's get_mapped_method function.
}
