% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigurationFunctions.R
\name{getConfiguration}
\alias{getConfiguration}
\title{Gets a stored configuration file}
\usage{
getConfiguration(modelname, spectype)
}
\arguments{
\item{modelname, }{str, the name of the model}

\item{spectype, }{str, specification type, either "model" or "disagg"}
}
\value{
A list of model specifications.
}
\description{
Gets a stored configuration file
}
