% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{getDisaggregationSpecs}
\alias{getDisaggregationSpecs}
\title{Obtain aggregation and disaggregation specs from input files}
\usage{
getDisaggregationSpecs(model, configpaths = NULL)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{configpaths}{str vector, paths (including file name) of disagg configuration file(s).
If NULL, built-in config files are used.}
}
\value{
A model with the specified aggregation and disaggregation specs.
}
\description{
Obtain aggregation and disaggregation specs from input files
}
