% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQuality.R
\name{getDQfields}
\alias{getDQfields}
\title{Draws from https://github.com/USEPA/ElectricityLCI/blob/master/electricitylci/dqi.py 
Get DQ fields in existing data frame}
\usage{
getDQfields(df)
}
\arguments{
\item{df}{totals data frame}
}
\value{
string vector with names of data quality fields
}
\description{
Draws from https://github.com/USEPA/ElectricityLCI/blob/master/electricitylci/dqi.py 
Get DQ fields in existing data frame
}
