% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQualityFunctions.R
\name{getDQfields}
\alias{getDQfields}
\title{Functions to support data quality assessment
Draws from https://github.com/USEPA/ElectricityLCI/blob/master/electricitylci/dqi.py 
Get DQ fields in existing data frame}
\usage{
getDQfields(df)
}
\arguments{
\item{df}{A totals_by_sector data frame}
}
\value{
A string vector with names of data quality fields
}
\description{
Functions to support data quality assessment
Draws from https://github.com/USEPA/ElectricityLCI/blob/master/electricitylci/dqi.py 
Get DQ fields in existing data frame
}
