% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{getFlowbySector}
\alias{getFlowbySector}
\title{Calls the Python flowsa package's getFlowBySector method}
\usage{
getFlowbySector(method_name)
}
\arguments{
\item{method_name}{The name}
}
\value{
A dataframe in flowbysector format
}
\description{
Calls the Python flowsa package's getFlowBySector method
}
