% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowsaFunctions.R
\name{getFlowbySectorCollapsed}
\alias{getFlowbySectorCollapsed}
\title{Load flowsa's FlowBySector df and collapse sector columns}
\usage{
getFlowbySectorCollapsed(sat_spec)
}
\arguments{
\item{sat_spec, }{a standard specification for a single satellite table}
}
\value{
A data frame for flowsa data in sector by region totals format
}
\description{
Load flowsa's FlowBySector df and collapse sector columns
}
