% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCIAfmtFunctions.R
\name{getImpactMethod}
\alias{getImpactMethod}
\title{Get impact method in the format of the Python LCIAformatter package's get_mapped_method function.}
\usage{
getImpactMethod(parameters)
}
\arguments{
\item{parameters}{List of parameters, must include 'filename', 'methods' and 'indicators' are optional
which are a list of one or more indicators to include from the specified method.}
}
\value{
An LCIAmethod with the specified indicators
}
\description{
Get impact method in the format of the Python LCIAformatter package's get_mapped_method function.
}
