% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCIAfmtFunctions.R
\name{getImpactMethod}
\alias{getImpactMethod}
\title{Get impact method in the format of the Python LCIAformatter package's get_mapped_method function.}
\usage{
getImpactMethod(ind_spec)
}
\arguments{
\item{ind_spec}{Specification of an indicator}
}
\value{
An LCIAmethod with the specified indicators
}
\description{
Get impact method in the format of the Python LCIAformatter package's get_mapped_method function.
}
