% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationFunctions.R
\name{getIndicatorColorMapping}
\alias{getIndicatorColorMapping}
\title{Uses VizualizationEssentials.yml to get a mapping of the to indicator Color scheme}
\usage{
getIndicatorColorMapping()
}
\value{
df with mapping with model indicator to colors
}
\description{
Uses VizualizationEssentials.yml to get a mapping of the to indicator Color scheme
}
