% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCIAfmtFunctions.R
\name{getInventoryMethod}
\alias{getInventoryMethod}
\title{Get inventory method using the Python LCIAformatter package's get_mapped_method function.}
\usage{
getInventoryMethod(parameters)
}
\arguments{
\item{parameters}{List of parameters, must include 'indicators' which is a list of one or 
more indicators to include from inventory method.}
}
\value{
An LCIAmethod data frame with the specified indicators.
}
\description{
Get inventory method using the Python LCIAformatter package's get_mapped_method function.
}
