% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadMargins.R
\name{getMarginsTable}
\alias{getMarginsTable}
\title{Generate Margins table using BEA Margin Details table which include all industries and final demand.}
\usage{
getMarginsTable(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A data.frame containing CommodityCode, and margins for ProducersValue, Transportation, Wholesale, Retail and PurchasersValue.
}
\description{
Generate Margins table using BEA Margin Details table which include all industries and final demand.
}
