% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrosswalkFunctions.R
\name{getNAICS2to6Digits}
\alias{getNAICS2to6Digits}
\title{Get 2-6 digit NAICS codes in a crosswalk format for year specified.}
\usage{
getNAICS2to6Digits(year)
}
\arguments{
\item{year}{int, 2012 or 2007 accepted.}
}
\value{
data frame with columns NAICS_2, NAICS_3, NAICS_4, NAICS_5, NAICS_6.
}
\description{
Get 2-6 digit NAICS codes in a crosswalk format for year specified.
}
