% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrosswalkFunctions.R
\name{getNAICStoBEAAllocation}
\alias{getNAICStoBEAAllocation}
\title{Determine allocation factors between NAICS and BEA sectors based on Industry output.}
\usage{
getNAICStoBEAAllocation(year, model)
}
\arguments{
\item{year}{Year of model Industry output.}

\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A table of allocation factors between NAICS and BEA sectors.
}
\description{
Determine allocation factors between NAICS and BEA sectors based on Industry output.
}
