% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculationFunctions.R
\name{getSectorLinkages}
\alias{getSectorLinkages}
\title{Function to get backward or forward economic linkages from a USEEIO model,
Provides results based on input demand, sorted by total (direct+indirect)
Backward linkages use A and L matrices
Forward linkages use Ghosh counterparts to A and L}
\usage{
getSectorLinkages(
  model,
  demand,
  type = "backward",
  location = NULL,
  use_domestic_requirements = FALSE,
  cutoff = 0.01
)
}
\arguments{
\item{model}{An EEIO model object with model specs, IO tables, satellite tables,
and indicators loaded}

\item{demand}{A demand vector, can be name of a built-in model demand vector,
e.g. "Production" or "Consumption",
or an actual demand vector with names as one or more model sectors and
numeric values in USD with the same dollar year as model.}

\item{type}{"backward" linkages use A and L matrices "forward" linkages use Ghosh
counterparts to A and L}

\item{location}{str optional location code for demand vector, required for
two-region models}

\item{use_domestic_requirements}{A logical value: if TRUE, use domestic demand
and L_d matrix; if FALSE, use complete demand and L matrix.}

\item{cutoff}{numeric value, shows the cutoff value for sorted results.
Values smaller than cutoff are not shown}
}
\value{
dataframe of direct, indirect, and total economic linkages
}
\description{
Function to get backward or forward economic linkages from a USEEIO model,
Provides results based on input demand, sorted by total (direct+indirect)
Backward linkages use A and L matrices
Forward linkages use Ghosh counterparts to A and L
}
