% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateiorFunctions.R
\name{getTwoRegionIOData}
\alias{getTwoRegionIOData}
\title{Load two-region IO data of model iolevel and year from user's local directory
or the EPA Data Commons.}
\usage{
getTwoRegionIOData(model, dataname)
}
\arguments{
\item{model}{An EEIO form USEEIO model object with model specs and IO meta data loaded.}

\item{dataname}{Name of desired IO data, can be "Make", "Use", "DomesticUse",
"UseTransactions", "FinalDemand", "InternationalTradeAdjustment,
"DomesticUseTransactions", "DomesticFinalDemand",
"CommodityOutput, "IndustryOutput", and "DomesticUsewithTrade".}
}
\value{
A list of two-region IO data of model iolevel and year.
}
\description{
Load two-region IO data of model iolevel and year from user's
local directory or the EPA Data Commons.
}
