% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationFunctions.R
\name{heatmapSatelliteTableCoverage}
\alias{heatmapSatelliteTableCoverage}
\title{Heatmap showing coverage of satellite tables}
\usage{
heatmapSatelliteTableCoverage(model)
}
\arguments{
\item{model}{A complete EEIO model}
}
\description{
Heatmap showing coverage of satellite tables
}
