% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationFunctions.R
\name{heatmapSatelliteTableCoverage}
\alias{heatmapSatelliteTableCoverage}
\title{Heatmap showing coverage of satellite tables}
\usage{
heatmapSatelliteTableCoverage(model, form = "Commodity")
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes}

\item{form}{Form of sectors in satellite table, can be"Commodity" and "Industry".}
}
\description{
Heatmap showing coverage of satellite tables
}
