% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationFunctions.R
\name{heatmapSectorRanking}
\alias{heatmapSectorRanking}
\title{SMM tool like heatmap showing ranking of sectors}
\usage{
heatmapSectorRanking(
  model,
  matrix,
  indicators,
  sector_to_remove,
  N_sector,
  x_title = NULL
)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes}

\item{matrix}{A matrix from model result}

\item{indicators}{A vector of indicators to plot}

\item{sector_to_remove}{Code of one or more BEA sectors that will be removed from the plot. Can be "".}

\item{N_sector}{A numeric value indicating number of sectors to show in the ranking}

\item{x_title}{A string specifying desired title on the x-axis, default is NULL, the title will be "modelname indicators"}
}
\description{
SMM tool like heatmap showing ranking of sectors
}
