% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIOFunctions.R
\name{includeWIOSectorDFs}
\alias{includeWIOSectorDFs}
\title{Include the WIO elements in the model$Commodity and model$Industry objects}
\usage{
includeWIOSectorDFs(model, WIO)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{WIO}{A list with WIO specifications and data}
}
\value{
A model object which contain the model$Commodity and model$Industry objects with WIO sectors
}
\description{
Include the WIO elements in the model$Commodity and model$Industry objects
}
