% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigDocumentation.R
\docType{data}
\name{IndicatorConfiguration}
\alias{IndicatorConfiguration}
\title{Indicator Configuration}
\format{
A list of indicator attributes
\describe{
 \item{FullName}{string, Name of the indicator}
 \item{Abbreviation}{string, Name of the indicator}
 \item{Category}{string, Impact category of the indicator}
 \item{Unit}{string, unit of the indicator}
 \item{SatticSource}{boolean, TRUE if this is a static source}
 \item{DataSources}{list of data sources, @seealso [DataSourceConfiguration]}
 }
}
\usage{
IndicatorConfiguration
}
\description{
Indicator Configuration
}
\keyword{datasets}
