% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigDocumentation.R
\docType{data}
\name{IndicatorConfiguration}
\alias{IndicatorConfiguration}
\title{Indicator Configuration}
\format{
A list of indicator attributes
\describe{
 \item{Name}{string, Name of the indicator}
 \item{Code}{string, Abbreviation for the indicator}
 \item{Group}{string, Impact category of the indicator}
 \item{Unit}{string, unit of the indicator}
 \item{SimpleUnit}{string, simple unit of the indicator}
 \item{SimpleName}{string, simple name of the indicator}
 \item{StaticSource}{boolean, TRUE if this is a static source}
 \item{StaticFile}{string, name of the source file}
 \item{FileLocation}{string, description of the location of the source file}
 \item{ScriptFunctionCall}{string, the function used to process the indicator data}
 \item{ScriptFunctionParameters}{list, parameters passed to ScriptFunctionCall}
 \item{DataSources}{list of data sources, @seealso [DataSourceConfiguration]}
 }
}
\usage{
IndicatorConfiguration
}
\description{
Indicator Configuration
}
\keyword{datasets}
