% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationFunctions.R
\name{lineplotFlowCoefficients}
\alias{lineplotFlowCoefficients}
\title{Line plot of flow coefficients}
\usage{
lineplotFlowCoefficients(flow, matrix_list, modelspecs_list)
}
\arguments{
\item{flow}{An index name for a flow in a model matrix (full flow name)}

\item{matrix_list}{List of model matrix, supports a model B, M, or U matrix}

\item{modelspecs_list}{List of specs for given models}
}
\description{
Line plot of flow coefficients
}
