% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIndicators.R
\name{loadandbuildIndicators}
\alias{loadandbuildIndicators}
\title{Loads data for all model indicators as listed in model specs}
\usage{
loadandbuildIndicators(model)
}
\arguments{
\item{model}{A model object with model specs, IO tables, and satelilte tables loaded}
}
\value{
A model object with Indicators added
}
\description{
Loads data for all model indicators as listed in model specs
}
