% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIOTables.R
\name{loadBEAtables}
\alias{loadBEAtables}
\title{Load BEA IO tables in a list based on model config and io_codes.}
\usage{
loadBEAtables(specs, io_codes)
}
\arguments{
\item{specs}{Model specifications.}

\item{io_codes}{A list of BEA IO codes.}
}
\value{
A list with BEA IO tables
}
\description{
Load BEA IO tables in a list based on model config and io_codes.
}
