% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDemandVectors.R
\name{loadDefaultDemandVectorMeta}
\alias{loadDefaultDemandVectorMeta}
\title{Loads a package stored demand vector metadata (.yml) for vectors to be created for every model with type and system specified
This function adds additional year, location and IDs along with the type and system based on the model specs.}
\usage{
loadDefaultDemandVectorMeta(model)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}
}
\value{
a data frame of metadata with columns Type, System, Name, Year, Location and ID with rows for each default
}
\description{
Loads a package stored demand vector metadata (.yml) for vectors to be created for every model with type and system specified
This function adds additional year, location and IDs along with the type and system based on the model specs.
}
