% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDemandVectors.R
\name{loadDemandVectors}
\alias{loadDemandVectors}
\title{Adds demand vectors and metadata based on model specs to model object}
\usage{
loadDemandVectors(model)
}
\arguments{
\item{model}{A model list object with the specs object listed}
}
\value{
model with a list of demand vectors and a meta file
}
\description{
Adds demand vectors and metadata based on model specs to model object
}
