% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDemandVectors.R
\name{loadDemandVectors}
\alias{loadDemandVectors}
\title{Adds demand vectors and metadata based on useeior defaults and model specs to model object}
\usage{
loadDemandVectors(model)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}
}
\value{
model with a list of demand vectors and a meta file stored appended as model$DemandVectors
}
\description{
Adds demand vectors and metadata based on useeior defaults and model specs to model object
}
