% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadExternalImportFactors.R
\name{loadExternalImportFactors}
\alias{loadExternalImportFactors}
\title{Prepare economic components of an EEIO form USEEIO model.}
\usage{
loadExternalImportFactors(model, configpaths = NULL)
}
\arguments{
\item{model}{An EEIO form USEEIO model object with model specs loaded}

\item{configpaths}{str vector, paths (including file name) of model configuration file
If NULL, built-in config files are used.}
}
\value{
A list with EEIO form USEEIO model economic components.
}
\description{
Prepare economic components of an EEIO form USEEIO model.
}
