% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalImportFactors.R
\name{loadExternalImportFactors}
\alias{loadExternalImportFactors}
\title{Load and prepare import coefficients}
\usage{
loadExternalImportFactors(model, configpaths = NULL)
}
\arguments{
\item{model}{An EEIO form USEEIO model object with model specs loaded}

\item{configpaths}{str vector, paths (including file name) of model configuration file.
If NULL, built-in config files are used.}
}
\value{
M_m, matrix of import coefficients (flow x sector).
}
\description{
Load and prepare import coefficients
}
