% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIndicators.R
\name{loadIndicators}
\alias{loadIndicators}
\title{Load indicators and associated factors in a list based on model config.}
\usage{
loadIndicators(specs)
}
\arguments{
\item{specs}{Specifications of the model.}
}
\value{
A list with df for indicators and df for factors.
}
\description{
Load indicators and associated factors in a list based on model config.
}
