% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIndicators.R
\name{loadindicators}
\alias{loadindicators}
\title{Load indicator factors in a list based on model config.}
\usage{
loadindicators(specs)
}
\arguments{
\item{specs}{Specifications of the model.}
}
\value{
A list of indicator factors not yet formatted for IOMB.
}
\description{
Load indicator factors in a list based on model config.
}
