% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIndicators.R
\name{loadIndicators}
\alias{loadIndicators}
\title{Load indicators and associated factors in a list based on model config.}
\usage{
loadIndicators(model)
}
\arguments{
\item{model}{A model object with IO tables and satellite tables loaded.}
}
\value{
A list with data.frame for indicators and data.frame for factors.
}
\description{
Load indicators and associated factors in a list based on model config.
}
