% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIOTables.R
\name{loadIOData}
\alias{loadIOData}
\title{Prepare economic components of an EEIO form USEEIO model.}
\usage{
loadIOData(modelname)
}
\arguments{
\item{modelname}{Name of the model from a config file.}
}
\value{
A list with USEEIO model economic components.
}
\description{
Prepare economic components of an EEIO form USEEIO model.
}
