% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIOTables.R
\name{loadIOData}
\alias{loadIOData}
\title{Prepare economic components of an EEIO model.}
\usage{
loadIOData(model)
}
\arguments{
\item{model}{A list of model specs}
}
\value{
A list with EEIO model economic components.
}
\description{
Prepare economic components of an EEIO model.
}
