% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIOTables.R
\name{loadIOData}
\alias{loadIOData}
\title{Prepare economic components of an EEIO form USEEIO model.}
\usage{
loadIOData(model, configpaths = NULL)
}
\arguments{
\item{model}{An EEIO form USEEIO model object with model specs loaded}

\item{configpaths}{str vector, paths (including file name) of model configuration file
and optional agg/disagg configuration file(s). If NULL, built-in config files are used.}
}
\value{
A list with EEIO form USEEIO model economic components.
}
\description{
Prepare economic components of an EEIO form USEEIO model.
}
