% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_LCIAFunctions.R
\name{loadLCIAfactors}
\alias{loadLCIAfactors}
\title{Loads all LCIA factors from static source file after melting it to long file
Returns df with "Name""Category""Subcategory""Unit""UUID""Abbreviation""Amount"}
\usage{
loadLCIAfactors()
}
\description{
Loads all LCIA factors from static source file after melting it to long file
Returns df with "Name""Category""Subcategory""Unit""UUID""Abbreviation""Amount"
}
