% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIndicators.R
\name{loadLCIAfactors}
\alias{loadLCIAfactors}
\title{Loads all LCIA factors from static source file after melting it to long file}
\usage{
loadLCIAfactors()
}
\value{
A dataframe of LCIA factors.
}
\description{
Loads all LCIA factors from static source file after melting it to long file
}
