% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSatellites.R
\name{loadsattables}
\alias{loadsattables}
\title{Load satellite tables in a list based on model.}
\usage{
loadsattables(model)
}
\arguments{
\item{model}{Configuration of the model.}
}
\value{
A list with satellite tables.
}
\description{
Load satellite tables in a list based on model.
}
