% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSatellites.R
\name{loadsattables}
\alias{loadsattables}
\title{Load totals by sector/region and builds satellite tables based on model specs.
Supports BEA and NAICS based totals. These totals can be provided as static files
or dynamic function calls are supported. NAICS-based totals are aggregated/allocated to BEA sectors
as part of the preparation.}
\format{
A list with lists of totals by sector and formatted satellite tables
\describe{
 \itemize{
   \item totals_by_sector
     \itemize{
       \item Flowable {Name of the flow}
       \item Sector {Code of the sector in the model IO schema}
       \item SectorSourceName {Source of the sector categorization, default is NAICS_2012_Code}
       \item Context {Full context of the flow, compartment and subcompartment combined}
       \item Location {Activity location, at a national, state, or county level}
       \item Unit {SI unit acronym. 'kg' for mass flows; 'MJ' for energy flows.}
       \item Year {Year of the data}
       \item DistributionType {Form of the frequency distribution, if given. Acceptable values are 'NORMAL', 'LOGNORMAL', 'TRIANGULAR', 'UNIFORM'.}
       \item FlowAmount {Amount of the flow}
       \item Min {The minimum FlowAmount, if provided for the data range.}
       \item Max {The maximum FlowAmount, if provided for the data range.}
       \item DataReliability {A 1-5 score of data reliability based on reporting values associated with the amount.}
       \item TemporalCorrelation {A 1-5 score of data collection based on reporting values associated with the amount.}
       \item GeographicalCorrelation {A 1-5 score of data collection based on reporting values associated with the amount.}
       \item TechnologicalCorrelation {A 1-5 score of data collection based on reporting values associated with the amount.}
       \item DataCollection {A 1-5 score of data collection based on reporting values associated with the amount.}
     }
   \item coeffs_by_sector
     \itemize{
       \item FlowName
       \item CAS
       \item FlowCategory
       \item FlowSubCategory
       \item FlowUUID
       \item ProcessName
       \item ProcessCode
       \item ProcessLocation
       \item FlowAmount       
       \item UncertaintyDistribution
       \item UncertaintyExpectedValue
       \item UncertaintyDispersion
       \item UncertaintyMin
       \item UncertaintyMax
       \item DQReliability
       \item DQTemporal
       \item DQGeographical
       \item DQTechnological
       \item DQDataCollection
       \item MetaYearofData
       \item MetaTags
       \item MetaSources
       \item MetaYearofData        
       \item MetaOther
     }
 }
}
}
\usage{
loadsattables(model)
}
\arguments{
\item{model}{A model list object with the specs object listed}
}
\value{
Lists of national totals by sector and formatted satellite tables
}
\description{
Only works for static national totals by BEA sector in a set format
}
