% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSatellites.R
\name{loadSatTables}
\alias{loadSatTables}
\title{Load totals by sector/region and prepares them based on model specs.}
\format{
A list with lists of totals by sector
\describe{
 \itemize{
   \item totals_by_sector
     \itemize{
       \item Flowable {Name of the flow}
       \item Context {Full context of the flow, compartment and subcompartment combined}
       \item Unit {SI unit acronym. 'kg' for mass flows; 'MJ' for energy flows.}
       \item FlowUUID {unique hex code for flow}
       \item SectorName {Name of the sector}
       \item Sector {Code of the sector in the model IO schema}
       \item FlowAmount {Amount of the flow}
       \item Location {Activity location, at a national, state, or county level}
       \item Year {Year of the data}
       \item DistributionType {Form of the frequency distribution, if given. Acceptable values are 'NORMAL', 'LOGNORMAL', 'TRIANGULAR', 'UNIFORM'.}
       \item Min {The minimum FlowAmount, if provided for the data range.}
       \item Max {The maximum FlowAmount, if provided for the data range.}
       \item DataReliability {A 1-5 score of data reliability based on reporting values associated with the amount.}
       \item TemporalCorrelation {A 1-5 score of data collection based on reporting values associated with the amount.}
       \item GeographicalCorrelation {A 1-5 score of data collection based on reporting values associated with the amount.}
       \item TechnologicalCorrelation {A 1-5 score of data collection based on reporting values associated with the amount.}
       \item DataCollection {A 1-5 score of data collection based on reporting values associated with the amount.}
       \item MetaSources {Tag for the data source.}
   }
 }
}
}
\usage{
loadSatTables(model)
}
\arguments{
\item{model}{A model list object with model specs and IO tables listed}
}
\value{
Lists of totals by sector by region and unique flows
}
\description{
Supports BEA and NAICS based totals. These totals can be provided as static files
or dynamic function calls are supported. NAICS-based totals are aggregated/allocated to BEA sectors
as part of the preparation.
}
