% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadIOTables.R
\name{loadTwoRegionStateIOtables}
\alias{loadTwoRegionStateIOtables}
\title{Load two-region state IO tables in a list based on model config.}
\usage{
loadTwoRegionStateIOtables(model)
}
\arguments{
\item{model}{An EEIO form USEEIO model object with model specs and IO meta data loaded.}
}
\value{
A list with state IO tables.
}
\description{
Load two-region state IO tables in a list based on model config.
}
