% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{mapFIPS5toLocationCodes}
\alias{mapFIPS5toLocationCodes}
\title{Maps a vector of FIPS codes to location codes
! Placeholder only works for '00000' now}
\usage{
mapFIPS5toLocationCodes(fipscodes)
}
\arguments{
\item{fipscodes}{A vector of 5 digit FIPS codes}
}
\value{
A vector of location codes where matches are found
}
\description{
Maps a vector of FIPS codes to location codes
! Placeholder only works for '00000' now
}
