% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\docType{data}
\name{matrices}
\alias{matrices}
\title{The vector of matrices to write out}
\format{
An object of class \code{character} of length 16.
}
\usage{
matrices
}
\description{
The vector of matrices to write out
}
\keyword{datasets}
