% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioeconomyFunctions.R
\name{modifyBmatrix}
\alias{modifyBmatrix}
\title{This function modifies the B table created in buildEEIOModel() for adding one new bioeconomy sector}
\usage{
modifyBmatrix(newEnvData, model)
}
\arguments{
\item{newEnvData}{(# environmental flows x 1) column vector with the data for all the environmental flows per dollar of output for the new sector.}

\item{model}{the list of USEEIO model components and attributes under modification.}

\item{newSectorCode}{string/character that refers to the code/identifier that will be used for the new sector in the matrices.}
}
\value{
B matrix modified.
}
\description{
This function modifies the B table created in buildEEIOModel() for adding one new bioeconomy sector
}
