% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioeconomyFunctions.R
\name{modifyMakeTable}
\alias{modifyMakeTable}
\title{Modify Make BEA Table.}
\usage{
modifyMakeTable(
  newSectorCode,
  similarSectorCode,
  percentage,
  originalMake,
  inputPurchases
)
}
\arguments{
\item{newSectorCode}{string/character that refers to the code/identifier that will be used for the new sector in the matrices}

\item{similarSectorCode}{string/character that refers to the code/identifier for the similar existing sector.}

\item{percentage}{numeric that refers to the \% of output of the original sector that new sector will produce.}

\item{originalMake}{dataframe with the Make table obtained in buildEEIOModel().}

\item{inputPurchases}{(n+1)x 1 column vector with the amount spent in each of the n existing commodities to produce "Total Industry Output"
of the new sector.}
}
\value{
a dataframe with the originalMake table modified.

Example: modifyMakeTable("prueba","324110",0.2,model$Make)
}
\description{
This function modifies the pre-saved Make table for adding one new bioeconomy sector.
}
