% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandFunctions.R
\name{prepareConsumptionDemand}
\alias{prepareConsumptionDemand}
\title{Prepares a demand vector representing consumption}
\usage{
prepareConsumptionDemand(model)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}
}
\value{
a named vector with demand
}
\description{
Prepares a demand vector representing consumption
}
