% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandFunctions.R
\name{prepareDomesticConsumptionDemand}
\alias{prepareDomesticConsumptionDemand}
\title{Prepares a demand vector representing domestic consumption}
\usage{
prepareDomesticConsumptionDemand(model)
}
\arguments{
\item{model, }{a model}
}
\value{
a named vector with demand
}
\description{
Prepares a demand vector representing domestic consumption
}
