% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandFunctions.R
\name{prepareDomesticProductionDemand}
\alias{prepareDomesticProductionDemand}
\title{Prepares a demand vector representing domestic production
Formula for production vector: y_p <- y_dc + y_e + y_d_delta + mu}
\usage{
prepareDomesticProductionDemand(model, location)
}
\arguments{
\item{model}{An EEIO model object with model specs and IO tables loaded}

\item{location, }{str of location code for demand vector}
}
\value{
A named vector with demand
}
\description{
Prepares a demand vector representing domestic production
Formula for production vector: y_p <- y_dc + y_e + y_d_delta + mu
}
