% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateModel.R
\name{prepareEfromtbs}
\alias{prepareEfromtbs}
\title{Concatenate all satellite flows in model}
\usage{
prepareEfromtbs(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes}
}
\description{
Concatenate all satellite flows in model
}
