% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandFunctions.R
\name{prepareHouseholdDemand}
\alias{prepareHouseholdDemand}
\title{Prepares a demand vector representing household consumption}
\usage{
prepareHouseholdDemand(model)
}
\arguments{
\item{model, }{a model}
}
\value{
a named vector with demand
}
\description{
Prepares a demand vector representing household consumption
}
