% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCIAfmtFunctions.R
\name{prepareLCIAmethodforIndicators}
\alias{prepareLCIAmethodforIndicators}
\title{Prepares and reformats LCIAmethod data from LCIAformatter for use
modeled after prepareFlowBySectorCollapsedforSatellite}
\usage{
prepareLCIAmethodforIndicators(lciamethod)
}
\arguments{
\item{lciamethod}{A full LCIAmethod data frame from LCIAformatter via getInventoryMethod or getImpactMethod.}
}
\value{
A LCIAmethod data frame formatted for indicators
}
\description{
Prepares and reformats LCIAmethod data from LCIAformatter for use
modeled after prepareFlowBySectorCollapsedforSatellite
}
