% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandFunctions.R
\name{prepareProductionDemand}
\alias{prepareProductionDemand}
\title{Prepares a demand vector representing production}
\usage{
prepareProductionDemand(model)
}
\arguments{
\item{model, }{a model}
}
\value{
a named vector with demand
}
\description{
Prepares a demand vector representing production
}
