% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{prepareWriteDirs}
\alias{prepareWriteDirs}
\title{Sets directories to write model output data to}
\usage{
prepareWriteDirs(model, dirs)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{dirs}{A named list of directories}
}
\description{
Sets directories to write model output data to.
If model is not passed, just uses data directory.
}
