% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{RAS}
\alias{RAS}
\title{Generalized RAS procedure. Takes an initial matrix, a target row sum vector
and target colsum vector. Iterates until all row sums of matrix equal to target row sum vector
and colsums of matrix equal target col sum vector, within a tolerance.}
\usage{
RAS(m0, t_r, t_c, t, max_itr = 1e+06)
}
\arguments{
\item{m0}{A matrix object.}

\item{t_r}{A vector setting the target row sums of the matrix.}

\item{t_c}{A vector setting the target column sums of the matrix.}

\item{t}{A numeric value setting the tolerance of RAS.}

\item{max_itr}{A numeric value setting the maximum number of iterations to try for convergence.
Default is 1,000,000.}
}
\value{
A RAS balanced matrix.
}
\description{
Generalized RAS procedure. Takes an initial matrix, a target row sum vector
and target colsum vector. Iterates until all row sums of matrix equal to target row sum vector
and colsums of matrix equal target col sum vector, within a tolerance.
}
