% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{replaceNonewithNA}
\alias{replaceNonewithNA}
\title{Replaces all `None` in a dataframe with the R NULL type NA}
\usage{
replaceNonewithNA(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame without `None`
}
\description{
Replaces all `None` in a dataframe with the R NULL type NA
}
