% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_CPI_IO}
\alias{Sector_CPI_IO}
\title{Sector 2007-2018 CPI (2012 schema)}
\format{
A dataframe with 15 obs. and 17 variables
}
\source{
\url{https://edap-ord-data-commons.s3.amazonaws.com/useeior/AllTablesUnderlying.zip}
}
\usage{
Sector_CPI_IO
}
\description{
Sector 2007-2018 CPI (2012 schema)
}
\keyword{datasets}
