% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_Make_2011_BeforeRedef}
\alias{Sector_Make_2011_BeforeRedef}
\title{Sector 2011 Make Before Redefinition (2012 schema)}
\format{
A dataframe with 16 obs. and 18 variables
}
\source{
\url{https://apps.bea.gov//industry/iTables\%20Static\%20Files/AllTablesIO.zip}
}
\usage{
Sector_Make_2011_BeforeRedef
}
\description{
TBD
}
\keyword{datasets}
