% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Sector_Make_2011_BeforeRedef}
\alias{Sector_Make_2011_BeforeRedef}
\title{Sector 2011 Make Before Redefinition (2012 schema)}
\format{
A dataframe with 16 obs. and 18 variables
}
\source{
\url{https://edap-ord-data-commons.s3.amazonaws.com/useeior/AllTablesIO.zip}
}
\usage{
Sector_Make_2011_BeforeRedef
}
\description{
Sector 2011 Make Before Redefinition (2012 schema)
}
\keyword{datasets}
