% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisaggregateFunctions.R
\name{specifiedUseDisagg}
\alias{specifiedUseDisagg}
\title{Disaggregate use table based on the allocations specified in the files referenced in the disaggregation specs.}
\usage{
specifiedUseDisagg(model, disagg, domestic = FALSE)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{disagg}{Specifications for disaggregating the current Table}

\item{domestic}{Flag that indicates where to use the Domestic Use or UseTransactions table}
}
\value{
A standardized make table with old sectors removed and new disaggregated sectors added based on the allocations in the disaggregation specs.
}
\description{
Disaggregate use table based on the allocations specified in the files referenced in the disaggregation specs.
}
