% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatelliteFunctions.R
\name{stackSatelliteTables}
\alias{stackSatelliteTables}
\title{Stacks two tables up}
\usage{
stackSatelliteTables(sattable1, sattable2)
}
\arguments{
\item{sattable1}{A standardized satellite table.}

\item{sattable2}{Another standardized satellite table.}
}
\value{
A complete standardized satellite table.
}
\description{
Stacks two tables up
}
