% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{standardizeandcastSatelliteTable}
\alias{standardizeandcastSatelliteTable}
\title{Converts flows table into flows x sector matrix-like format}
\usage{
standardizeandcastSatelliteTable(df, model)
}
\arguments{
\item{df}{a dataframe of flowables, contexts, units, sectors and locations}

\item{model}{An EEIO model object with model specs, IO tables, satellite tables, and indicators loaded}
}
\value{
A matrix-like dataframe of flows x sector
}
\description{
Converts flows table into flows x sector matrix-like format
}
