% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_CommodityCodeName_2012}
\alias{Summary_CommodityCodeName_2012}
\title{Summary Commodity Code and Name (2012 schema)}
\format{
A dataframe with 73 obs. and 2 variables
}
\source{
\url{https://edap-ord-data-commons.s3.amazonaws.com/useeior/AllTablesIO.zip}
}
\usage{
Summary_CommodityCodeName_2012
}
\description{
Summary Commodity Code and Name (2012 schema)
}
\keyword{datasets}
