% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{Summary_Make_2015_BeforeRedef}
\alias{Summary_Make_2015_BeforeRedef}
\title{Summary 2015 Make Before Redefinition (2012 schema)}
\format{
A dataframe with 72 obs. and 74 variables
}
\source{
\url{https://edap-ord-data-commons.s3.amazonaws.com/useeior/AllTablesIO.zip}
}
\usage{
Summary_Make_2015_BeforeRedef
}
\description{
Summary 2015 Make Before Redefinition (2012 schema)
}
\keyword{datasets}
