% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_IOfunctions.R
\name{transformFinalDemandwithMarketShares}
\alias{transformFinalDemandwithMarketShares}
\title{Transform Final Demand df with Market Shares matrix.}
\usage{
transformFinalDemandwithMarketShares(Fdf)
}
\arguments{
\item{B}{Marginal impact per unit of the environmental flows.}

\item{D}{Market Shares matrix.}
}
\value{
Final Demand matrix.
}
\description{
Transform Final Demand df with Market Shares matrix.
}
