% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{transformFinalDemandwithMarketShares}
\alias{transformFinalDemandwithMarketShares}
\title{Transform Final Demand (commodity x sector) with Market Shares matrix
to Final Demand (industry x sector)}
\usage{
transformFinalDemandwithMarketShares(Fdf, model)
}
\arguments{
\item{Fdf}{Final Demand data.frame.}

\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
Final Demand (industry x sector) data.frame
}
\description{
Transform Final Demand (commodity x sector) with Market Shares matrix
to Final Demand (industry x sector)
}
