% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{transformIndustryOutputtoCommodityOutputforYear}
\alias{transformIndustryOutputtoCommodityOutputforYear}
\title{Generate Commodity output by transforming Industry output using Commodity Mix matrix.}
\usage{
transformIndustryOutputtoCommodityOutputforYear(year, model)
}
\arguments{
\item{year}{Year of Industry output}

\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\value{
A dataframe contains adjusted Commodity output.
}
\description{
Generate Commodity output by transforming Industry output using Commodity Mix matrix.
}
