% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeMatrixtoJSON}
\alias{writeMatrixtoJSON}
\title{Write model matrix as JSON files to output folder.}
\usage{
writeMatrixtoJSON(matrix, outputpath)
}
\arguments{
\item{outputpath}{A directory and file to write to.}

\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\description{
Writes model objects.
}
