% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{writeMetadatatoJSON}
\alias{writeMetadatatoJSON}
\title{Write metadata of downloaded data to JSON.}
\usage{
writeMetadatatoJSON(
  package,
  name,
  year,
  source,
  url,
  date_last_modified,
  date_accessed
)
}
\arguments{
\item{package}{A string specifying package.}

\item{name}{A string specifying data name.}

\item{year}{A numeric value specifying data year.}

\item{source}{A string specifying data source.}

\item{url}{A string specifying data url.}

\item{date_last_modified}{A string specifying when the original data was
last modified by provider, e.g. BEA.}

\item{date_accessed}{A string specifying when the original data was accessed
by package.}
}
\description{
Write metadata of downloaded data to JSON.
}
