% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{writeModelComponents}
\alias{writeModelComponents}
\title{Write model components to output folder.}
\usage{
writeModelComponents(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\description{
Write model components to output folder.
}
