% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeModelMatrices}
\alias{writeModelMatrices}
\title{Write model matrices as CSV and BIN files to output folder.}
\usage{
writeModelMatrices(model)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\description{
Writes model matrices, including A, B, C, D, L, and U.
}
