% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeModelMatrices}
\alias{writeModelMatrices}
\title{Write model matrices as CSV files to output folder.}
\usage{
writeModelMatrices(model, outputfolder)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{outputfolder}{A directory to write matrices out to}
}
\description{
Writes model matrices as CSV files, including A, A_d, B, C, D, L, L_d, M, M_d, N, N_d, and Rho (CPI ratio matrix)
}
