% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeModelMatrices}
\alias{writeModelMatrices}
\title{Write model matrices as .csv or .bin files to output folder.}
\usage{
writeModelMatrices(model, to_format, outputfolder)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{to_format}{A string specifying the format of write-to file, can be "csv" or "bin".}

\item{outputfolder}{A directory to write matrices out to}
}
\description{
Writes model matrices as .csv or .bin files to output folder.
}
