% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeModelMatricesforAPI}
\alias{writeModelMatricesforAPI}
\title{Write model matrices as BIN files for API to output folder.}
\usage{
writeModelMatricesforAPI(model, modelfolder)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{modelfolder}{Directory to write the model components to}
}
\description{
Writes model matrices, including Make, Use, A, A_d, B, C, D, L, L_d, M, N, Rho (CPI ratio matrix), x (Industry Output), and q (Commodity Output).
}
