% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeModelMetadata}
\alias{writeModelMetadata}
\title{Write model metadata (indicators and demands, sectors, and flows) as CSV files to output folder.}
\usage{
writeModelMetadata(model, dirs)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}
}
\description{
Writes model metadata, including indicators and demands.
}
