% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeModeltoXLSX}
\alias{writeModeltoXLSX}
\title{Write selected model objects and user meta to XLSX file}
\usage{
writeModeltoXLSX(model)
}
\arguments{
\item{model, }{any model object}
}
\description{
Write selected model objects and user meta to XLSX file
}
