% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeModeltoXLSX}
\alias{writeModeltoXLSX}
\title{Write selected model matrices, demand vectors, and metadata as XLSX file to output folder}
\usage{
writeModeltoXLSX(model, outputfolder)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{outputfolder}{A directory to write model matrices and metadata as XLSX file.}
}
\description{
Writes model matrices demand vectors, and metadata as XLSX file to output folder.
}
