% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeSectorCrosswalk}
\alias{writeSectorCrosswalk}
\title{Write the master sector crosswalk out for the API}
\usage{
writeSectorCrosswalk()
}
\description{
Writes master sector crosswalk out for the API in csv
}
