% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeSectorCrosswalk}
\alias{writeSectorCrosswalk}
\title{Write the model sector crosswalk as .csv file}
\usage{
writeSectorCrosswalk(model, outputfolder)
}
\arguments{
\item{model}{A complete EEIO model: a list with USEEIO model components and attributes.}

\item{outputfolder}{A directory to write model sector crosswalk}
}
\description{
Writes the model sector crosswalk as .csv file
}
