% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteModel.R
\name{writeSectorCrosswalkforAPI}
\alias{writeSectorCrosswalkforAPI}
\title{Write the master sector crosswalk out for the API}
\usage{
writeSectorCrosswalkforAPI()
}
\description{
Writes master sector crosswalk out for the API in csv
}
